% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_draw_the_line_data.R
\name{bind_draw_the_line_data}
\alias{bind_draw_the_line_data}
\title{Bind draw the line chart data}
\usage{
bind_draw_the_line_data(
  .,
  data = NULL,
  x = NULL,
  lines_reveal = NULL,
  lines_fixed = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{data}{Data.}

\item{x}{Time (x axis). Labels for the X axis. Flourish type hint: column}

\item{lines_reveal}{Lines reveal}

\item{lines_fixed}{Lines fixed}
}
\value{
A Flourish chart
}
\description{
Bind draw the line chart data
}
\examples{
try(
  flourish(chart_type = "draw_the_line", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_draw_the_line_data(gapminder)
)
}
