% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_election_results_chart_data.R
\name{bind_election_results_chart_data}
\alias{bind_election_results_chart_data}
\title{Bind election results chart data}
\usage{
bind_election_results_chart_data(
  .,
  values_now_data = NULL,
  values_historical_data = NULL,
  values_now_name = NULL,
  values_now_values = NULL,
  values_historical_name = NULL,
  values_historical_values = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{values_now_data}{Values now data.}

\item{values_historical_data}{Values historical data.}

\item{values_now_name}{Region name. The name of the data series (eg. voting district). Flourish type hint: column}

\item{values_now_values}{Results by party. One column per party, with party names as header rows. Flourish type hint: columns}

\item{values_historical_name}{Region name. The name of the data series (eg. voting district). Flourish type hint: column}

\item{values_historical_values}{Results by party. One column per party, with party names as header rows. Flourish type hint: columns}
}
\value{
A Flourish chart
}
\description{
Bind election results chart data
}
\examples{
try(
  flourish(chart_type = "election_results", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_election_results_chart_data(gapminder)
)
}
