% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_parliament_chart_data.R
\name{bind_parliament_chart_data}
\alias{bind_parliament_chart_data}
\title{Bind parliament chart data}
\usage{
bind_parliament_chart_data(., data = NULL, party = NULL, seats = NULL)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{data}{Data.}

\item{party}{Party}

\item{seats}{Seats. This should point to the column with election data. Flourish type hint: columns}
}
\value{
A Flourish chart
}
\description{
Bind parliament chart data
}
\examples{
try(
  flourish(chart_type = "parliament", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_parliament_chart_data(gapminder)
)
}
