% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_text_annotator_data.R
\name{bind_text_annotator_data}
\alias{bind_text_annotator_data}
\title{Bind text annotator data}
\usage{
bind_text_annotator_data(., data = NULL, text = NULL, note = NULL)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{data}{Data.}

\item{text}{Text. The piece of text you want to highlight. Note that text is case and space sensitive. Flourish type hint: column}

\item{note}{Annotation. The annotation content. Flourish type hint: column}
}
\value{
A Flourish chart
}
\description{
Bind text annotator data
}
\examples{
try(
  flourish(chart_type = "text_annotator", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_text_annotator_data(gapminder)
)
}
