% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_word_cloud_data.R
\name{bind_word_cloud_data}
\alias{bind_word_cloud_data}
\title{Bind word cloud data}
\usage{
bind_word_cloud_data(
  .,
  words_data = NULL,
  word = NULL,
  value = NULL,
  category = NULL,
  metadata = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{words_data}{Words data.}

\item{word}{Word}

\item{value}{Value}

\item{category}{Category}

\item{metadata}{Info for popups. One or more columns of information (text, image URLs, embedded charts etc) to include in popups and panels Flourish type hint: columns}
}
\value{
A Flourish chart
}
\description{
Bind word cloud data
}
\examples{
try(
  flourish(chart_type = "word_cloud", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_word_cloud_data(gapminder)
)
}
