% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flourish_api_documentation}
\alias{flourish_api_documentation}
\title{Flourish template API documentation}
\format{
A data frame with 57 rows and 3 columns:
\describe{
\item{template_name}{Template name. Found in API documentation: \url{https://app.flourish.studio/@flourish}. }
\item{chart_type}{Chart type. One template may have multiple chart types.}
\item{url}{URL for template API documentation.}
}
}
\source{
\url{https://flourish.studio/index.html}
}
\usage{
flourish_api_documentation
}
\value{
A data frame
}
\description{
A data set containing template metadata about Flourish graphs
}
\examples{
flourish_api_documentation
}
\keyword{datasets}
