% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{Gapminder data}
\format{
A data frame with 1704 rows and 6 columns:
\describe{
\item{country}{Country name.}
\item{continent}{Continent name.}
\item{year}{Year data was collected.}
\item{lifeExp}{Life expectancy.}
\item{pop}{Population size.}
\item{gdpPercap}{GDP per capita.}
}
}
\source{
\url{https://www.gapminder.org/}
}
\usage{
gapminder
}
\value{
A data frame
}
\description{
A data set containing Gapminder data used for vignettes
}
\examples{
gapminder
}
\keyword{datasets}
