% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_calculator_details.R
\name{set_calculator_details}
\alias{set_calculator_details}
\title{Set calculator chart details}
\usage{
set_calculator_details(
  .,
  question_flow = NULL,
  design_border_type = NULL,
  design_border_color = NULL,
  design_background = NULL,
  design_advanced = NULL,
  design_align_horizontal = NULL,
  design_align_vertical = NULL,
  design_max_width = NULL,
  design_margin_vertical = NULL,
  design_padding_vertical = NULL,
  design_padding_horizontal = NULL,
  design_border_width = NULL,
  design_border_style = NULL,
  design_border_radius = NULL,
  design_shadow = NULL,
  design_shadow_color = NULL,
  result_variables = NULL,
  result_input = NULL,
  result_advanced = NULL,
  result_font_size = NULL,
  result_font_size_custom = NULL,
  result_font_color = NULL,
  result_font_weight = NULL,
  result_font_style = NULL,
  result_border_type = NULL,
  result_border_width = NULL,
  result_border_style = NULL,
  result_border_color = NULL,
  result_bg_type = NULL,
  result_bg_color = NULL,
  result_bg_image_src = NULL,
  result_bg_image_size = NULL,
  design_question_font_size = NULL,
  design_question_font_size_custom = NULL,
  design_question_font_color = NULL,
  design_question_font_weight = NULL,
  design_question_font_style = NULL,
  design_question_margin_vertical = NULL,
  design_question_margin_horizontal = NULL,
  design_question_padding = NULL,
  design_context_font_size = NULL,
  design_context_font_size_custom = NULL,
  design_context_font_color = NULL,
  design_context_font_weight = NULL,
  design_context_font_style = NULL,
  design_context_margin_vertical = NULL,
  design_context_margin_horizontal = NULL,
  design_question_bg_color = NULL,
  design_question_bg_opacity = NULL,
  design_question_border_radius = NULL,
  design_answer_margin_vertical = NULL,
  design_answer_margin_horizontal = NULL,
  input_color = NULL,
  input_background_color = NULL,
  input_width = NULL,
  number_input_placeholder = NULL,
  text_input_placeholder = NULL,
  input_input_advanced = NULL,
  input_border_color = NULL,
  input_border_radius = NULL,
  input_border_width = NULL,
  input_border_style = NULL,
  input_align = NULL,
  input_margin = NULL,
  input_padding = NULL,
  button_font_size = NULL,
  button_font_size_custom = NULL,
  chart_button_control_style_background = NULL,
  chart_button_control_style_background_selected = NULL,
  chart_button_control_style_background_hover = NULL,
  chart_button_control_style_font_color = NULL,
  chart_button_control_style_font_color_selected = NULL,
  chart_button_control_style_font_color_hover = NULL,
  chart_button_control_style_button_styles_advanced = NULL,
  chart_button_control_style_border_width = NULL,
  chart_button_control_style_border_color = NULL,
  chart_button_control_style_border_transparency = NULL,
  chart_button_control_style_border_radius = NULL,
  button_border_style = NULL,
  button_container_width = NULL,
  button_justify_content = NULL,
  button_margin = NULL,
  button_padding_factor = NULL,
  dropdown_font_size = NULL,
  dropdown_font_size_custom = NULL,
  dropdown_placeholder = NULL,
  chart_dropdown_control_style_background = NULL,
  chart_dropdown_control_style_font_color = NULL,
  chart_dropdown_control_style_border_styles_advanced = NULL,
  chart_dropdown_control_style_border_style = NULL,
  chart_dropdown_control_style_border_width = NULL,
  chart_dropdown_control_style_border_color = NULL,
  chart_dropdown_control_style_border_transparency = NULL,
  chart_dropdown_control_style_border_radius = NULL,
  dropdown_padding_factor = NULL,
  dropdown_border_style = NULL,
  dropdown_font_weight = NULL,
  dropdown_font_style = NULL,
  dropdown_shadow = NULL,
  dropdown_shadow_color = NULL,
  date_format = NULL,
  date_input_placeholder = NULL,
  date_advanced = NULL,
  date_width = NULL,
  date_margin = NULL,
  date_padding = NULL,
  date_border_width = NULL,
  date_border_style = NULL,
  date_border_color = NULL,
  date_border_radius = NULL,
  date_align = NULL,
  date_calendar_position_vertical = NULL,
  date_calendar_position_horizontal = NULL,
  rating_icon_type = NULL,
  rating_icon_fill_base = NULL,
  rating_icon_stroke_base = NULL,
  rating_icon_fill_active = NULL,
  rating_icon_stroke_active = NULL,
  rating_advanced = NULL,
  rating_icon_width = NULL,
  rating_icon_height = NULL,
  rating_icon_margin = NULL,
  rating_icon_stroke_width = NULL,
  rating_container_width = NULL,
  rating_wrap_margin_vertical = NULL,
  rating_wrap_margin_horizontal = NULL,
  rating_wrap_padding_vertical = NULL,
  rating_wrap_padding_horizontal = NULL,
  rating_label_width = NULL,
  rating_label_color = NULL,
  rating_label_position = NULL,
  chart_slider_filter_controls_handle_color = NULL,
  chart_slider_filter_controls_track_active_color = NULL,
  chart_slider_filter_controls_track_base_color = NULL,
  chart_slider_filter_controls_handle_size = NULL,
  chart_slider_filter_controls_ticks = NULL,
  chart_slider_filter_controls_advanced = NULL,
  chart_slider_filter_controls_container_width = NULL,
  chart_slider_filter_controls_wrap_margin_vertical = NULL,
  chart_slider_filter_controls_wrap_margin_horizontal = NULL,
  chart_slider_filter_controls_track_height = NULL,
  chart_slider_filter_controls_track_border_radius = NULL,
  chart_slider_filter_controls_track_border_color = NULL,
  chart_slider_filter_controls_track_border_width = NULL,
  chart_slider_filter_controls_handle_shape = NULL,
  chart_slider_filter_controls_handle_width = NULL,
  chart_slider_filter_controls_handle_border_color = NULL,
  chart_slider_filter_controls_handle_border_width = NULL,
  chart_slider_filter_controls_handle_border_radius = NULL,
  chart_slider_filter_controls_handle_box_shadow = NULL,
  chart_slider_filter_controls_tooltips = NULL,
  chart_slider_filter_controls_tooltip_background = NULL,
  chart_slider_filter_controls_tick_color = NULL,
  chart_slider_filter_controls_tick_padding = NULL,
  chart_slider_filter_controls_tick_number_padding = NULL,
  submit_comment_text = NULL,
  submit_button_text = NULL,
  submit_bg_type = NULL,
  submit_bg_color = NULL,
  submit_bg_image_src = NULL,
  submit_border = NULL,
  submit_bg_image_size = NULL,
  submit_advanced = NULL,
  submit_comment_font_size = NULL,
  submit_comment_font_size_custom = NULL,
  submit_comment_font_weight = NULL,
  submit_comment_font_style = NULL,
  submit_flex_direction = NULL,
  submit_comment_font_color = NULL,
  submit_comment_margin = NULL,
  submit_button_font_size = NULL,
  submit_button_font_size_custom = NULL,
  submit_button_font_weight = NULL,
  submit_button_font_style = NULL,
  submit_button_margin = NULL,
  submit_button_padding_factor = NULL,
  submit_button_box_shadow = NULL,
  submit_button_use_default_styles = NULL,
  submit_button_bg_color = NULL,
  submit_button_bg_color_hover = NULL,
  submit_button_font_color = NULL,
  submit_button_font_color_hover = NULL,
  submit_button_border_style = NULL,
  submit_button_border_color = NULL,
  submit_button_border_width = NULL,
  submit_button_border_transparency = NULL,
  submit_button_border_radius = NULL,
  design_controls_position = NULL,
  design_controls_button_symbol_back = NULL,
  design_controls_button_symbol_forward = NULL,
  design_controls_button_font_color = NULL,
  design_controls_button_font_size = NULL,
  design_controls_advanced = NULL,
  design_controls_button_margin = NULL,
  design_controls_button_padding_factor = NULL,
  design_controls_button_bg_color = NULL,
  design_controls_button_bg_opacity = NULL,
  design_controls_button_font_weight = NULL,
  design_controls_button_font_style = NULL,
  design_controls_button_border_width = NULL,
  design_controls_button_border_style = NULL,
  design_controls_button_border_color = NULL,
  design_controls_button_border_radius = NULL,
  design_progress_show = NULL,
  design_progress_position = NULL,
  design_progress_height = NULL,
  design_progress_margin = NULL,
  design_progress_bg_color = NULL,
  design_progress_bg_opacity = NULL,
  chart_number_date_format_prefix = NULL,
  chart_number_date_format_suffix = NULL,
  chart_number_date_format_n_dec = NULL,
  chart_number_date_format_advanced = NULL,
  chart_number_date_format_negative_sign = NULL,
  chart_number_date_format_strip_zeros = NULL,
  chart_number_date_format_strip_separator = NULL,
  chart_number_date_format_transform_labels = NULL,
  chart_number_date_format_transform = NULL,
  chart_number_date_format_multiply_divide_constant = NULL,
  chart_number_date_format_exponentiate_constant = NULL,
  chart_number_localization_input_decimal_separator = NULL,
  chart_number_localization_output_separators = NULL,
  external_write_enabled = NULL,
  external_write_endpoint = NULL,
  chart_layout_body_font = NULL,
  chart_layout_font_color = NULL,
  chart_layout_background_color_enabled = NULL,
  chart_layout_background_image_enabled = NULL,
  chart_layout_background_color = NULL,
  chart_layout_background_image_src = NULL,
  chart_layout_background_image_size = NULL,
  chart_layout_background_image_position = NULL,
  chart_layout_max_width_target = NULL,
  chart_layout_max_width = NULL,
  chart_layout_max_width_align = NULL,
  chart_layout_layout_order = NULL,
  chart_layout_space_between_sections = NULL,
  chart_layout_space_between_sections_custom = NULL,
  chart_layout_margin_top = NULL,
  chart_layout_margin_right = NULL,
  chart_layout_margin_bottom = NULL,
  chart_layout_margin_left = NULL,
  chart_layout_border_enabled = NULL,
  chart_layout_border_top_width = NULL,
  chart_layout_border_top_style = NULL,
  chart_layout_border_top_color = NULL,
  chart_layout_border_right_width = NULL,
  chart_layout_border_right_style = NULL,
  chart_layout_border_right_color = NULL,
  chart_layout_border_bottom_width = NULL,
  chart_layout_border_bottom_style = NULL,
  chart_layout_border_bottom_color = NULL,
  chart_layout_border_left_width = NULL,
  chart_layout_border_left_style = NULL,
  chart_layout_border_left_color = NULL,
  chart_layout_read_direction = NULL,
  chart_layout_font_size_mobile_small = NULL,
  chart_layout_font_size_mobile_big = NULL,
  chart_layout_font_size_tablet = NULL,
  chart_layout_font_size_desktop = NULL,
  chart_layout_font_size_big_screen = NULL,
  chart_layout_breakpoint_mobile_small = NULL,
  chart_layout_breakpoint_mobile_big = NULL,
  chart_layout_breakpoint_tablet = NULL,
  chart_layout_breakpoint_desktop = NULL,
  chart_layout_breakpoint_big_screen = NULL,
  chart_layout_header_align = NULL,
  chart_layout_title = NULL,
  chart_layout_title_styling = NULL,
  chart_layout_title_font = NULL,
  chart_layout_title_size = NULL,
  chart_layout_title_size_custom = NULL,
  chart_layout_title_weight = NULL,
  chart_layout_title_color = NULL,
  chart_layout_title_line_height = NULL,
  chart_layout_title_space_above = NULL,
  chart_layout_title_space_above_custom = NULL,
  chart_layout_subtitle = NULL,
  chart_layout_subtitle_styling = NULL,
  chart_layout_subtitle_font = NULL,
  chart_layout_subtitle_size = NULL,
  chart_layout_subtitle_size_custom = NULL,
  chart_layout_subtitle_weight = NULL,
  chart_layout_subtitle_color = NULL,
  chart_layout_subtitle_line_height = NULL,
  chart_layout_subtitle_space_above = NULL,
  chart_layout_subtitle_space_above_custom = NULL,
  chart_layout_header_text = NULL,
  chart_layout_header_text_styling = NULL,
  chart_layout_header_text_size = NULL,
  chart_layout_header_text_size_custom = NULL,
  chart_layout_header_text_weight = NULL,
  chart_layout_header_text_color = NULL,
  chart_layout_header_text_line_height = NULL,
  chart_layout_header_text_space_above = NULL,
  chart_layout_header_text_space_above_custom = NULL,
  chart_layout_header_border = NULL,
  chart_layout_header_border_width = NULL,
  chart_layout_header_border_color = NULL,
  chart_layout_header_border_style = NULL,
  chart_layout_header_border_space = NULL,
  chart_layout_header_logo_enabled = NULL,
  chart_layout_header_logo_src = NULL,
  chart_layout_header_logo_alt = NULL,
  chart_layout_header_logo_link_url = NULL,
  chart_layout_header_logo_height = NULL,
  chart_layout_header_logo_align = NULL,
  chart_layout_header_logo_position_inside = NULL,
  chart_layout_header_logo_position_outside = NULL,
  chart_layout_header_logo_margin_top = NULL,
  chart_layout_header_logo_margin_right = NULL,
  chart_layout_header_logo_margin_bottom = NULL,
  chart_layout_header_logo_margin_left = NULL,
  chart_layout_footer_align = NULL,
  chart_layout_footer_text_size = NULL,
  chart_layout_footer_text_color = NULL,
  chart_layout_footer_styling = NULL,
  chart_layout_footer_font = NULL,
  chart_layout_footer_text_weight = NULL,
  chart_layout_source_name = NULL,
  chart_layout_source_url = NULL,
  chart_layout_multiple_sources = NULL,
  chart_layout_source_name_2 = NULL,
  chart_layout_source_url_2 = NULL,
  chart_layout_source_name_3 = NULL,
  chart_layout_source_url_3 = NULL,
  chart_layout_source_label = NULL,
  chart_layout_footer_note = NULL,
  chart_layout_footer_note_secondary = NULL,
  chart_layout_advanced_note_styling = NULL,
  chart_layout_footer_timestamp_format = NULL,
  chart_layout_footer_logo_enabled = NULL,
  chart_layout_footer_logo_src = NULL,
  chart_layout_footer_logo_src_light = NULL,
  chart_layout_footer_logo_alt = NULL,
  chart_layout_footer_logo_link_url = NULL,
  chart_layout_footer_logo_height = NULL,
  chart_layout_footer_logo_margin = NULL,
  chart_layout_footer_logo_order = NULL,
  chart_layout_footer_align_vertical = NULL,
  chart_layout_footer_border = NULL,
  chart_layout_footer_border_width = NULL,
  chart_layout_footer_border_color = NULL,
  chart_layout_footer_border_style = NULL,
  chart_layout_footer_border_space = NULL,
  chart_layout_screenreader_text_primary = NULL,
  chart_layout_screenreader_label = NULL,
  chart_layout_screenreader_hide_primary = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{question_flow}{Question flow. Choices: vertical, horizontal. Flourish type hint: string, defaults to \code{NULL}}

\item{design_border_type}{Border. Choices: border-bottom, border. Flourish type hint: string, defaults to \code{NULL}}

\item{design_border_color}{Border color. Flourish type hint: color, defaults to \code{NULL}}

\item{design_background}{Background. The background color of the cards. If not set, the background color in Layout is used. Flourish type hint: color, defaults to \code{NULL}}

\item{design_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{design_align_horizontal}{Align. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{design_align_vertical}{Position. Choices: start, center, flex-end. Set the vertical text position if the window height is fixed and the text doesnt stretch across the full vertical card space. Flourish type hint: string, defaults to \code{NULL}}

\item{design_max_width}{Max width (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{design_margin_vertical}{Vertical margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_padding_vertical}{Vertical padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_padding_horizontal}{Horizontal padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{design_border_style}{Border style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{design_border_radius}{Roundness. Flourish type hint: number, defaults to \code{NULL}}

\item{design_shadow}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{design_shadow_color}{Shadow color. Shadow color as an RGBA value (red, green, blue, alpha). The alpha parameter is a number between 0.0 (fully transparent) and 1.0 (fully opaque). Flourish type hint: string, defaults to \code{NULL}}

\item{result_variables}{Variables.  Define variables to use in the Result field below. Set one variable per line in the format name = value Values can be:   Numbers   Text (single quotes for spaces)   Question references like q_2  with the number representing the row number of each question in the Questions datasheet  Previously defined variables (if wrapped into square   brackets ) which will resolve to the answered value   Excel functions, e.g., a = SUM(q_2, q_3)  Flourish type hint: text, defaults to \code{NULL}}

\item{result_input}{Result. Here you can compose your result view You have access to all variables you created above as well as all "q_" variables by referencing them in double curly braces like for example: \code{{{ my_variable }}} or \code{{{ q_6 }}} Flourish type hint: html, defaults to \code{NULL}}

\item{result_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{result_font_size}{Text size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{result_font_size_custom}{Custom. Specify a custom responsive text size. Best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{result_font_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{result_font_weight}{Text weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{result_font_style}{Text style. Choices: normal, italic. Flourish type hint: string, defaults to \code{NULL}}

\item{result_border_type}{Border. Choices: border-bottom, border. Flourish type hint: string, defaults to \code{NULL}}

\item{result_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{result_border_style}{Border style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{result_border_color}{Border color. Flourish type hint: color, defaults to \code{NULL}}

\item{result_bg_type}{Background. Choices: color, image. Flourish type hint: string, defaults to \code{NULL}}

\item{result_bg_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{result_bg_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{result_bg_image_size}{Image size. Choices: auto, cover, contain. Flourish type hint: string, defaults to \code{NULL}}

\item{design_question_font_size}{Text size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{design_question_font_size_custom}{Custom. Specify a custom responsive text size. Best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{design_question_font_color}{Color. The text color used for your questions. If not set, the text color in Layout is used. Flourish type hint: color, defaults to \code{NULL}}

\item{design_question_font_weight}{Text weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{design_question_font_style}{Text style. Choices: normal, italic. Flourish type hint: string, defaults to \code{NULL}}

\item{design_question_margin_vertical}{Vertical margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_question_margin_horizontal}{Horizontal margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_question_padding}{Padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_context_font_size}{Text size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{design_context_font_size_custom}{Custom. Specify a custom responsive text size. Best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{design_context_font_color}{Color. The text color used for your question context. If not set, the text color in Layout is used. Flourish type hint: color, defaults to \code{NULL}}

\item{design_context_font_weight}{Text weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{design_context_font_style}{Text style. Choices: normal, italic. Flourish type hint: string, defaults to \code{NULL}}

\item{design_context_margin_vertical}{Vertical margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_context_margin_horizontal}{Horizontal margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_question_bg_color}{Background color. The background color for your questions and context. To change the background color of the whole question card, go to the General design settings. Flourish type hint: color, defaults to \code{NULL}}

\item{design_question_bg_opacity}{Background opacity. The opacity of the question background color. Flourish type hint: number, defaults to \code{NULL}}

\item{design_question_border_radius}{Roundness. The roundness of the question background container. Flourish type hint: number, defaults to \code{NULL}}

\item{design_answer_margin_vertical}{Vertical margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_answer_margin_horizontal}{Horizontal margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{input_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{input_background_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{input_width}{Width. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{number_input_placeholder}{Number input placeholder. Flourish type hint: string, defaults to \code{NULL}}

\item{text_input_placeholder}{Text input placeholder. Flourish type hint: string, defaults to \code{NULL}}

\item{input_input_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{input_border_color}{Border color. Flourish type hint: color, defaults to \code{NULL}}

\item{input_border_radius}{Roundness. Flourish type hint: number, defaults to \code{NULL}}

\item{input_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{input_border_style}{Border style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{input_align}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{input_margin}{Margin. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{input_padding}{Padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{button_font_size}{Text size. Choices: 1, 1.2, 1.6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{button_font_size_custom}{Custom. Specify a custom responsive text size. Best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_background_selected}{Selected. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_background_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color_selected}{Selected. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_button_styles_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_button_control_style_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_border_transparency}{Transparency. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{button_border_style}{Border style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{button_container_width}{Button group width (\%). The width of the button group as a percentage of the card width. Flourish type hint: number, defaults to \code{NULL}}

\item{button_justify_content}{Align. Choices: flex-start, center, flex-end. Flourish type hint: string, defaults to \code{NULL}}

\item{button_margin}{Margin. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{button_padding_factor}{Padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{dropdown_font_size}{Text size. Choices: 1, 1.2, 1.6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{dropdown_font_size_custom}{Custom. Specify a custom responsive text size. Best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{dropdown_placeholder}{Placeholder. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_dropdown_control_style_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_font_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_styles_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_style}{Border style. Choices: all, bottom. Show border on all sides, or only at the bottom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_transparency}{Transparency. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{dropdown_padding_factor}{Padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{dropdown_border_style}{Border look. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{dropdown_font_weight}{Text weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{dropdown_font_style}{Text style. Choices: normal, italic. Flourish type hint: string, defaults to \code{NULL}}

\item{dropdown_shadow}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{dropdown_shadow_color}{Shadow color. Flourish type hint: string, defaults to \code{NULL}}

\item{date_format}{Format. Choices: y-m-d, d/m/_y, d/m/y, j/n/_y, j/n/y, d-m-_y, d-m-y, j-n-_y, j-n-y, d.m._y, d.m.y, j.n._y, j.n.y, m/d/_y, m/d/y, n/j/_y, n/j/y, m-d-_y, m-d-y, n-j-_y, n-j-y, m j, _y, f j, _y. Flourish type hint: string, defaults to \code{NULL}}

\item{date_input_placeholder}{Placeholder. Flourish type hint: string, defaults to \code{NULL}}

\item{date_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{date_width}{Width. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{date_margin}{Margin. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{date_padding}{Padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{date_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{date_border_style}{Border style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{date_border_color}{Border color. Flourish type hint: color, defaults to \code{NULL}}

\item{date_border_radius}{Roundness. Flourish type hint: number, defaults to \code{NULL}}

\item{date_align}{Text alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{date_calendar_position_vertical}{Calendar vertical position. Choices: auto, above, below. Flourish type hint: string, defaults to \code{NULL}}

\item{date_calendar_position_horizontal}{Calendar horizontal position. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{rating_icon_type}{Choices: star, circle, check. Flourish type hint: string, defaults to \code{NULL}}

\item{rating_icon_fill_base}{Base fill color. Flourish type hint: color, defaults to \code{NULL}}

\item{rating_icon_stroke_base}{Base stroke color. Flourish type hint: color, defaults to \code{NULL}}

\item{rating_icon_fill_active}{Active fill color. Flourish type hint: color, defaults to \code{NULL}}

\item{rating_icon_stroke_active}{Active stroke color. Flourish type hint: color, defaults to \code{NULL}}

\item{rating_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{rating_icon_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{rating_icon_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{rating_icon_margin}{Vertical margin. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{rating_icon_stroke_width}{Stroke width (px). Flourish type hint: number, defaults to \code{NULL}}

\item{rating_container_width}{Rating group width (\%). The width of the rating group as a percentage of the card width. Flourish type hint: number, defaults to \code{NULL}}

\item{rating_wrap_margin_vertical}{Vertical margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{rating_wrap_margin_horizontal}{Horizontal margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{rating_wrap_padding_vertical}{Vertical padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{rating_wrap_padding_horizontal}{Horizontal padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{rating_label_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{rating_label_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{rating_label_position}{Position. Choices: above, below. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_slider_filter_controls_handle_color}{Handle color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_filter_controls_track_active_color}{Track active color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_filter_controls_track_base_color}{Track background color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_filter_controls_handle_size}{Handle size. percentage of track height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_filter_controls_ticks}{Range ticks. Display the minimum and the maximum of the range at the left and right. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_slider_filter_controls_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_slider_filter_controls_container_width}{Track width (\%). The width of the track as a percentage of the card width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_filter_controls_wrap_margin_vertical}{Vertical margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_filter_controls_wrap_margin_horizontal}{Horizontal margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_filter_controls_track_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_filter_controls_track_border_radius}{Roundness. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_filter_controls_track_border_color}{Border color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_filter_controls_track_border_width}{Border width. The width of the border around the card, in rems Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_filter_controls_handle_shape}{Shape. Choices: rectangle, circle. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_slider_filter_controls_handle_width}{Width. \% of Size (the handle height). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_filter_controls_handle_border_color}{Stroke color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_filter_controls_handle_border_width}{Stroke width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_filter_controls_handle_border_radius}{Roundness. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_filter_controls_handle_box_shadow}{Box shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_slider_filter_controls_tooltips}{Tooltip. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_slider_filter_controls_tooltip_background}{Tooltip background. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_slider_filter_controls_tick_color}{Tick color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_filter_controls_tick_padding}{Tick padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_filter_controls_tick_number_padding}{Number padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_comment_text}{Submit label. Flourish type hint: text, defaults to \code{NULL}}

\item{submit_button_text}{Submit button text. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_bg_type}{Background. Choices: color, image. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_bg_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{submit_bg_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{submit_border}{Border. Flourish type hint: boolean, defaults to \code{NULL}}

\item{submit_bg_image_size}{Image size. Choices: auto, cover, contain. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{submit_comment_font_size}{Label size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_comment_font_size_custom}{Custom. Specify a custom responsive text size. Best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_comment_font_weight}{Text weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_comment_font_style}{Text style. Choices: normal, italic. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_flex_direction}{Label position. Choices: column, column-reverse. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_comment_font_color}{Label color. Flourish type hint: color, defaults to \code{NULL}}

\item{submit_comment_margin}{Vertical margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_button_font_size}{Button text size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_button_font_size_custom}{Custom. Specify a custom responsive text size. Best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_button_font_weight}{Button text weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_button_font_style}{Text style. Choices: normal, italic. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_button_margin}{Margin. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_button_padding_factor}{Padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_button_box_shadow}{Hover shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{submit_button_use_default_styles}{Use answer button styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{submit_button_bg_color}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{submit_button_bg_color_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{submit_button_font_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{submit_button_font_color_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{submit_button_border_style}{Border style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_button_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{submit_button_border_width}{Width. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_button_border_transparency}{Transparency. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_button_border_radius}{Roundness. Flourish type hint: number, defaults to \code{NULL}}

\item{design_controls_position}{Position. Choices: top_left, top_centre, top_right, bottom_left, bottom_centre, bottom_right. Flourish type hint: string, defaults to \code{NULL}}

\item{design_controls_button_symbol_back}{Back button. Flourish type hint: string, defaults to \code{NULL}}

\item{design_controls_button_symbol_forward}{Forward button. Flourish type hint: string, defaults to \code{NULL}}

\item{design_controls_button_font_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{design_controls_button_font_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_controls_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{design_controls_button_margin}{Margin. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_controls_button_padding_factor}{Padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_controls_button_bg_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{design_controls_button_bg_opacity}{Background opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{design_controls_button_font_weight}{Text weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{design_controls_button_font_style}{Text style. Choices: normal, italic. Flourish type hint: string, defaults to \code{NULL}}

\item{design_controls_button_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{design_controls_button_border_style}{Border style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{design_controls_button_border_color}{Border color. Flourish type hint: color, defaults to \code{NULL}}

\item{design_controls_button_border_radius}{Roundness. Flourish type hint: number, defaults to \code{NULL}}

\item{design_progress_show}{Show progress bar. Flourish type hint: boolean, defaults to \code{NULL}}

\item{design_progress_position}{Position. Choices: top, bottom. Flourish type hint: string, defaults to \code{NULL}}

\item{design_progress_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{design_progress_margin}{Horizontal margins. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{design_progress_bg_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{design_progress_bg_opacity}{Background opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_prefix}{Prefix. Text to place in front of number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_suffix}{Suffix. Text to place after number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_n_dec}{Decimal places. Enter a negative integer to round to a whole number with that many zeros. For example, -2 will round to the nearest hundred. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_negative_sign}{Styling of negative numbers. Choices: -$nk, $-nk, ($nk), $(n)k, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_strip_zeros}{Remove trailing zeros. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_strip_separator}{Hide thousands separator below 10,000. Turn off if you want four-digit numbers to include a separator, e.g. 1,234 rather than 1234. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_transform_labels}{Multiply/divide values. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_transform}{Choices: multiply, divide, exponentiate. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_multiply_divide_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_exponentiate_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_localization_input_decimal_separator}{Decimal separator in data sheet. Choices: ., ,. Used for interpreting your data. Only change if data is not displaying on the chart as expected. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_localization_output_separators}{Number format to display. Choices: ,., .,, ., ,, ., ,. How the numbers should appear on chart labels. Flourish type hint: string, defaults to \code{NULL}}

\item{external_write_enabled}{Store result data externally. Flourish type hint: boolean, defaults to \code{NULL}}

\item{external_write_endpoint}{Storage endpoint. Your external database, store, API or web app endpoint to post data to. If you were to use Google Sheets to store your data, please check the External write docs for integration help. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_body_font}{Main font. This font will apply to the whole graphic by default. You can optionally change the font for the title, subtitle, footer, etc in the Header and Footer settings panels. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_font_color}{Text color. This color will apply to the whole graphic by default, You can optionally change the color for individual text elements, in other settings panels. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_color_enabled}{Color. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_image_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_background_image_size}{Size. Choices: cover, contain, auto, 100\% 100\%. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_background_image_position}{Position. Choices: top left, top center, top right, center left, center center, center right, bottom left, bottom center, bottom right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width_target}{Maximum width. Choices: none, wrapper, primary. Apply a maximum width to just the main graphic or everything (main graphic plus header, footer, etc). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width}{Maximum width. Leave blank to stretch to container width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_max_width_align}{Align. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_layout_order}{Layout order. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections}{Space between sections. Choices: 0.5, 1, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_enabled}{Show borders around visualisation. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_border_top_width}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_top_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_top_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_right_width}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_right_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_right_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_bottom_width}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_bottom_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_bottom_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_left_width}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_left_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_left_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_read_direction}{Read direction. Choices: ltr, rtl. This will change the reading direction of the main text elements on the page. Its not possible to adjust this on all elements, such as axes.  Note that when direction is set to right to left any alignment icons will be reversed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_align}{Alignment. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_title_font}{Title Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_title_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_title_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_subtitle_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_subtitle_font}{Subtitle Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_subtitle_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_subtitle_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_header_text_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_text_size}{Size. Choices: 1.2, 1.4, 1.6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_text_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_space}{Space. Space between border and header text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_enabled}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_header_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_align}{Align. Choices: inside, outside. Align logo inside either the header or the main visualization container. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_inside}{Position. Choices: top, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_outside}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_align}{Alignment. Choices: left, center, right, justify. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_styling}{Advanced footer styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_font}{Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_footer_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_multiple_sources}{Multiple sources. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_source_name_2}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_2}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name_3}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_3}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_label}{Source label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_note}{Note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_footer_note_secondary}{Note (secondary). The secondary note is placed below the source and primary note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_advanced_note_styling}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_timestamp_format}{Time/date stamp formatting. Choices: \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M, \%H:\%M \%p, \%H:\%M:\%S, \%d \%b \%Y, \%H:\%M, \%d \%b \%Y, \%H:\%M, \%d/\%m/\%Y, \%H:\%M, \%m/\%d/\%Y, \%H:\%M, \%d-\%m-\%Y, \%H:\%M, \%m-\%d-\%Y, \%d \%b \%Y, \%m/\%d/\%Y, \%d/\%m/\%Y, \%m-\%d-\%Y, \%d-\%m-\%Y, \%Y-\%m-\%d, \%Y-\%m-\%d \%H:\%M:\%S. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src_light}{Image (light version). If provided this version will be used whenever the background color is dark. Flourish type hint: hidden, defaults to \code{NULL}}

\item{chart_layout_footer_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_margin}{Margin. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_order}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_align_vertical}{V. align. Choices: flex-start, center, flex-end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_space}{Space. Space between border and footer text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_screenreader_text_primary}{Screenreader description. A text alternative to the visual content that will only be visible to screenreaders, e.g. The line chart shows China consistently higher than the other countries since 1990.  Do not replicate your title, since that will also be read by screenreaders. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_layout_screenreader_label}{Screenreader label. A short text label given to the main Flourish embed wrapper to provide an accessible name that is only visible to screenreaders Added in the form of an "aria-label". Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_screenreader_hide_primary}{Screenreader mode for main visual container. Choices: \code{TRUE}, \code{FALSE}. Whether the main visual container is visible to screenreaders (Text in the header and footer are always available to screenreaders.). Flourish type hint: boolean, defaults to \code{NULL}}
}
\value{
A Flourish chart
}
\description{
Set calculator chart details
}
\examples{
try(
  flourish(chart_type = "calculator", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_calculator_data(gapminder) |>
    set_calculator_details()
)
}
