% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_chord_diagram_details.R
\name{set_chord_diagram_details}
\alias{set_chord_diagram_details}
\title{Set chord diagram chart details}
\usage{
set_chord_diagram_details(
  .,
  opacity_default = NULL,
  opacity_focused = NULL,
  opacity_defocused = NULL,
  selected_chords = NULL,
  use_mouseover_label = NULL,
  show_selected_label = NULL,
  chord_labels_reponsive = NULL,
  chord_label_font_size = NULL,
  chord_label_font_size_min = NULL,
  chord_label_font_size_max = NULL,
  show_chord_stroke = NULL,
  chord_stroke_color = NULL,
  chord_stroke_opacity = NULL,
  arc_thickness = NULL,
  arc_spacing = NULL,
  pad_angle = NULL,
  label_alignment = NULL,
  label_responsive = NULL,
  label_font_size = NULL,
  label_font_size_min = NULL,
  label_font_size_max = NULL,
  label_color = NULL,
  use_label_color_arc = NULL,
  chart_color_scale_categorical_palette = NULL,
  chart_color_scale_categorical_extend = NULL,
  chart_color_scale_categorical_custom_palette = NULL,
  chart_color_scale_numeric_type = NULL,
  chart_color_scale_binning = NULL,
  chart_color_scale_bin_mode = NULL,
  chart_color_scale_bin_count = NULL,
  chart_color_scale_bin_thresholds = NULL,
  chart_color_scale_sequential_palette = NULL,
  chart_color_scale_sequential_reverse = NULL,
  chart_color_scale_sequential_custom_min = NULL,
  chart_color_scale_sequential_custom_max = NULL,
  chart_color_scale_sequential_color_space = NULL,
  chart_color_scale_sequential_custom_domain = NULL,
  chart_color_scale_sequential_domain_min = NULL,
  chart_color_scale_sequential_domain_max = NULL,
  chart_color_scale_diverging_palette = NULL,
  chart_color_scale_diverging_reverse = NULL,
  chart_color_scale_diverging_custom_min = NULL,
  chart_color_scale_diverging_custom_mid = NULL,
  chart_color_scale_diverging_custom_max = NULL,
  chart_color_scale_diverging_color_space = NULL,
  chart_color_scale_diverging_custom_domain = NULL,
  chart_color_scale_diverging_domain_min = NULL,
  chart_color_scale_diverging_domain_mid = NULL,
  chart_color_scale_diverging_domain_max = NULL,
  gradient_type = NULL,
  chart_popup_show_popups = NULL,
  chart_popup_is_custom = NULL,
  chart_popup_custom_template = NULL,
  chart_popup_show_pointer = NULL,
  chart_popup_show_shadow = NULL,
  chart_popup_style_popups = NULL,
  chart_popup_text_color = NULL,
  chart_popup_align = NULL,
  chart_popup_text_size = NULL,
  chart_popup_fill_color = NULL,
  chart_popup_opacity = NULL,
  chart_popup_padding = NULL,
  chart_popup_border_radius = NULL,
  chart_popup_max_width = NULL,
  popup_type = NULL,
  popup_content = NULL,
  animation_speed = NULL,
  chart_number_localization_input_decimal_separator = NULL,
  chart_number_localization_output_separators = NULL,
  chart_label_number_date_format_prefix = NULL,
  chart_label_number_date_format_suffix = NULL,
  chart_label_number_date_format_n_dec = NULL,
  chart_label_number_date_format_advanced = NULL,
  chart_label_number_date_format_negative_sign = NULL,
  chart_label_number_date_format_strip_zeros = NULL,
  chart_label_number_date_format_strip_separator = NULL,
  chart_label_number_date_format_transform_labels = NULL,
  chart_label_number_date_format_transform = NULL,
  chart_label_number_date_format_multiply_divide_constant = NULL,
  chart_label_number_date_format_exponentiate_constant = NULL,
  chart_scale = NULL,
  allow_resize = NULL,
  chart_layout_body_font = NULL,
  chart_layout_font_color = NULL,
  chart_layout_background_color_enabled = NULL,
  chart_layout_background_image_enabled = NULL,
  chart_layout_background_color = NULL,
  chart_layout_background_image_src = NULL,
  chart_layout_background_image_size = NULL,
  chart_layout_background_image_position = NULL,
  chart_layout_max_width_target = NULL,
  chart_layout_max_width = NULL,
  chart_layout_max_width_align = NULL,
  chart_layout_layout_order = NULL,
  chart_layout_space_between_sections = NULL,
  chart_layout_space_between_sections_custom = NULL,
  chart_layout_margin_top = NULL,
  chart_layout_margin_right = NULL,
  chart_layout_margin_bottom = NULL,
  chart_layout_margin_left = NULL,
  chart_layout_border_enabled = NULL,
  chart_layout_border_top_width = NULL,
  chart_layout_border_top_style = NULL,
  chart_layout_border_top_color = NULL,
  chart_layout_border_right_width = NULL,
  chart_layout_border_right_style = NULL,
  chart_layout_border_right_color = NULL,
  chart_layout_border_bottom_width = NULL,
  chart_layout_border_bottom_style = NULL,
  chart_layout_border_bottom_color = NULL,
  chart_layout_border_left_width = NULL,
  chart_layout_border_left_style = NULL,
  chart_layout_border_left_color = NULL,
  chart_layout_read_direction = NULL,
  chart_layout_font_size_mobile_small = NULL,
  chart_layout_font_size_mobile_big = NULL,
  chart_layout_font_size_tablet = NULL,
  chart_layout_font_size_desktop = NULL,
  chart_layout_font_size_big_screen = NULL,
  chart_layout_breakpoint_mobile_small = NULL,
  chart_layout_breakpoint_mobile_big = NULL,
  chart_layout_breakpoint_tablet = NULL,
  chart_layout_breakpoint_desktop = NULL,
  chart_layout_breakpoint_big_screen = NULL,
  chart_layout_header_align = NULL,
  chart_layout_title = NULL,
  chart_layout_title_styling = NULL,
  chart_layout_title_font = NULL,
  chart_layout_title_size = NULL,
  chart_layout_title_size_custom = NULL,
  chart_layout_title_weight = NULL,
  chart_layout_title_color = NULL,
  chart_layout_title_line_height = NULL,
  chart_layout_title_space_above = NULL,
  chart_layout_title_space_above_custom = NULL,
  chart_layout_subtitle = NULL,
  chart_layout_subtitle_styling = NULL,
  chart_layout_subtitle_font = NULL,
  chart_layout_subtitle_size = NULL,
  chart_layout_subtitle_size_custom = NULL,
  chart_layout_subtitle_weight = NULL,
  chart_layout_subtitle_color = NULL,
  chart_layout_subtitle_line_height = NULL,
  chart_layout_subtitle_space_above = NULL,
  chart_layout_subtitle_space_above_custom = NULL,
  chart_layout_header_text = NULL,
  chart_layout_header_text_styling = NULL,
  chart_layout_header_text_size = NULL,
  chart_layout_header_text_size_custom = NULL,
  chart_layout_header_text_weight = NULL,
  chart_layout_header_text_color = NULL,
  chart_layout_header_text_line_height = NULL,
  chart_layout_header_text_space_above = NULL,
  chart_layout_header_text_space_above_custom = NULL,
  chart_layout_header_border = NULL,
  chart_layout_header_border_width = NULL,
  chart_layout_header_border_color = NULL,
  chart_layout_header_border_style = NULL,
  chart_layout_header_border_space = NULL,
  chart_layout_header_logo_enabled = NULL,
  chart_layout_header_logo_src = NULL,
  chart_layout_header_logo_alt = NULL,
  chart_layout_header_logo_link_url = NULL,
  chart_layout_header_logo_height = NULL,
  chart_layout_header_logo_align = NULL,
  chart_layout_header_logo_position_inside = NULL,
  chart_layout_header_logo_position_outside = NULL,
  chart_layout_header_logo_margin_top = NULL,
  chart_layout_header_logo_margin_right = NULL,
  chart_layout_header_logo_margin_bottom = NULL,
  chart_layout_header_logo_margin_left = NULL,
  chart_layout_footer_align = NULL,
  chart_layout_footer_text_size = NULL,
  chart_layout_footer_text_color = NULL,
  chart_layout_footer_styling = NULL,
  chart_layout_footer_font = NULL,
  chart_layout_footer_text_weight = NULL,
  chart_layout_source_name = NULL,
  chart_layout_source_url = NULL,
  chart_layout_multiple_sources = NULL,
  chart_layout_source_name_2 = NULL,
  chart_layout_source_url_2 = NULL,
  chart_layout_source_name_3 = NULL,
  chart_layout_source_url_3 = NULL,
  chart_layout_source_label = NULL,
  chart_layout_footer_note = NULL,
  chart_layout_footer_note_secondary = NULL,
  chart_layout_advanced_note_styling = NULL,
  chart_layout_footer_timestamp_format = NULL,
  chart_layout_footer_logo_enabled = NULL,
  chart_layout_footer_logo_src = NULL,
  chart_layout_footer_logo_src_light = NULL,
  chart_layout_footer_logo_alt = NULL,
  chart_layout_footer_logo_link_url = NULL,
  chart_layout_footer_logo_height = NULL,
  chart_layout_footer_logo_margin = NULL,
  chart_layout_footer_logo_order = NULL,
  chart_layout_footer_align_vertical = NULL,
  chart_layout_footer_border = NULL,
  chart_layout_footer_border_width = NULL,
  chart_layout_footer_border_color = NULL,
  chart_layout_footer_border_style = NULL,
  chart_layout_footer_border_space = NULL,
  chart_layout_screenreader_text_primary = NULL,
  chart_layout_screenreader_label = NULL,
  chart_layout_screenreader_hide_primary = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{opacity_default}{Default opacity. Opacity for the default view. Flourish type hint: number, defaults to \code{NULL}}

\item{opacity_focused}{Focused opacity. Opacity for the selected and/or hovered chords. Flourish type hint: number, defaults to \code{NULL}}

\item{opacity_defocused}{Defocused opacity. Opacity for the background chords when one is in focus. Flourish type hint: number, defaults to \code{NULL}}

\item{selected_chords}{Highlight chords. You can highlight all of an arcs chords by typing the name of the arc. Alternatively, you can highlight individual chords by typing the name of the two arcs that it connects using the format: Arc 1 >> Arc 2  If you wanted to highlight all chords from "Arc 3" and any chords between "Arc 1" and "Arc 2" you would write: Arc 3  Arc 1 >> Arc 2  If you wish to highlight/unhighlight in the story editor you can do so by clicking on chords. Flourish type hint: text, defaults to \code{NULL}}

\item{use_mouseover_label}{Show chord values on hover. Flourish type hint: boolean, defaults to \code{NULL}}

\item{show_selected_label}{Keep chord values on click. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chord_labels_reponsive}{Label text size. Choices: \code{TRUE}, \code{FALSE}. Auto changes the label size based on the window size, subject to the specified min and max. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chord_label_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{chord_label_font_size_min}{Min size. Flourish type hint: number, defaults to \code{NULL}}

\item{chord_label_font_size_max}{Max size. Flourish type hint: number, defaults to \code{NULL}}

\item{show_chord_stroke}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chord_stroke_color}{Stroke color. Flourish type hint: color, defaults to \code{NULL}}

\item{chord_stroke_opacity}{Stroke opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{arc_thickness}{Thickness. As a percentage of the radius. Flourish type hint: number, defaults to \code{NULL}}

\item{arc_spacing}{Space below. As a percentage of the radius. Flourish type hint: number, defaults to \code{NULL}}

\item{pad_angle}{Space between. In degrees. Flourish type hint: number, defaults to \code{NULL}}

\item{label_alignment}{Alignment. Choices: along_arcs, 90_rotated. Flourish type hint: string, defaults to \code{NULL}}

\item{label_responsive}{Label text size. Choices: \code{TRUE}, \code{FALSE}. Auto changes the label size based on the window size, subject to the specified min and max. Flourish type hint: boolean, defaults to \code{NULL}}

\item{label_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{label_font_size_min}{Min size. Flourish type hint: number, defaults to \code{NULL}}

\item{label_font_size_max}{Max size. Flourish type hint: number, defaults to \code{NULL}}

\item{label_color}{Label color. Flourish type hint: color, defaults to \code{NULL}}

\item{use_label_color_arc}{Make the arc label color the same as the arc color. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_categorical_palette}{Palette. Flourish type hint: colors, defaults to \code{NULL}}

\item{chart_color_scale_categorical_extend}{Extend. Automatically generate additional colours when needed to avoid the palette colours being used more than once. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_categorical_custom_palette}{Custom overrides. Enter the label name for which you wish to set the color, followed by a colon and the desired color value.  Colors can be set using Hex, RGB, color names or RGBA, if you want to set the opacity Multiple colors can be set using multiple lines. For example:  Party 1: red Party 2: #4455AA Party 3: rgb(30,168,26). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_color_scale_numeric_type}{Scale type. Choices: sequential, diverging. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_binning}{Choices: \code{FALSE}, \code{TRUE}. In linear mode, the color scale will run as a smooth gradient between 2 colors. In binned mode, the gradient will be divided in smaller blocks. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_bin_mode}{Binning mode. Choices: fixed, quantile, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_bin_count}{Number of bins. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_bin_thresholds}{Custom thresholds. Enter your desired thresholds, separating them with a ";". For instance, "5;10;15".  Bins form as follows:  From the datas minimum value (domain minimum) up to the first threshold.  Between consecutive thresholds.  From the last threshold to the datas maximum value (domain maximum).   For "5;10;15", youll get four bins based on your datas range. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_mid}{Midpoint color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_mid}{Mid. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{gradient_type}{Chord color mode. Choices: linear_gradient, largest_end, smallest_end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_show_popups}{Popups. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_is_custom}{Popup contents. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_custom_template}{Popup content. The text to appear in the popup. You can use \code{{{column_name}}} to add a value from your data. It must be in a selected column, but you can add columns to Metadata if you just want to include them for use in the popup. Advanced users can include HTML to apply layouts, formatting, images, etc. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_popup_show_pointer}{Pointer. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_show_shadow}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_style_popups}{Custom styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_text_color}{Text colour. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_align}{Alignment. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_text_size}{Font size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_fill_color}{Fill colour. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_opacity}{Fill opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_border_radius}{Radius. Corner radius of popup. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_max_width}{Max width. Flourish type hint: number, defaults to \code{NULL}}

\item{popup_type}{Position. Choices: directional, non_directional. Flourish type hint: string, defaults to \code{NULL}}

\item{popup_content}{Popup description. Add a title or a discription to help explain your tooltip. Flourish type hint: text, defaults to \code{NULL}}

\item{animation_speed}{Animation duration. The duration, in seconds, of transitions  for example between two slides in a story Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_localization_input_decimal_separator}{Decimal separator in data sheet. Choices: ., ,. Used for interpreting your data. Only change if data is not displaying on the chart as expected. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_localization_output_separators}{Number format to display. Choices: ,., .,, ., ,, ., ,. How the numbers should appear on chart labels. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_label_number_date_format_prefix}{Prefix. Text to place in front of number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_label_number_date_format_suffix}{Suffix. Text to place after number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_label_number_date_format_n_dec}{Decimal places. Enter a negative integer to round to a whole number with that many zeros. For example, -2 will round to the nearest hundred. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_label_number_date_format_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_label_number_date_format_negative_sign}{Styling of negative numbers. Choices: -$nk, $-nk, ($nk), $(n)k, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_label_number_date_format_strip_zeros}{Remove trailing zeros. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_label_number_date_format_strip_separator}{Hide thousands separator below 10,000. Turn off if you want four-digit numbers to include a separator, e.g. 1,234 rather than 1234. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_label_number_date_format_transform_labels}{Multiply/divide values. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_label_number_date_format_transform}{Choices: multiply, divide, exponentiate. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_label_number_date_format_multiply_divide_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_label_number_date_format_exponentiate_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_scale}{Chart scale. Flourish type hint: number, defaults to \code{NULL}}

\item{allow_resize}{Allow embeds to resize. Checking this option allows the visualization to resize so the chord diagram is always a square without white space to the sides or top/bottom. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_body_font}{Main font. This font will apply to the whole graphic by default. You can optionally change the font for the title, subtitle, footer, etc in the Header and Footer settings panels. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_font_color}{Text color. This color will apply to the whole graphic by default, You can optionally change the color for individual text elements, in other settings panels. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_color_enabled}{Color. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_image_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_background_image_size}{Size. Choices: cover, contain, auto, 100\% 100\%. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_background_image_position}{Position. Choices: top left, top center, top right, center left, center center, center right, bottom left, bottom center, bottom right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width_target}{Maximum width. Choices: none, wrapper, primary. Apply a maximum width to just the main graphic or everything (main graphic plus header, footer, etc). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width}{Maximum width. Leave blank to stretch to container width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_max_width_align}{Align. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_layout_order}{Layout order. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections}{Space between sections. Choices: 0.5, 1, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_enabled}{Show borders around visualisation. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_border_top_width}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_top_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_top_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_right_width}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_right_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_right_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_bottom_width}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_bottom_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_bottom_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_left_width}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_left_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_left_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_read_direction}{Read direction. Choices: ltr, rtl. This will change the reading direction of the main text elements on the page. Its not possible to adjust this on all elements, such as axes.  Note that when direction is set to right to left any alignment icons will be reversed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_align}{Alignment. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_title_font}{Title Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_title_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_title_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_subtitle_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_subtitle_font}{Subtitle Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_subtitle_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_subtitle_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_header_text_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_text_size}{Size. Choices: 1.2, 1.4, 1.6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_text_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_space}{Space. Space between border and header text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_enabled}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_header_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_align}{Align. Choices: inside, outside. Align logo inside either the header or the main visualization container. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_inside}{Position. Choices: top, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_outside}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_align}{Alignment. Choices: left, center, right, justify. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_styling}{Advanced footer styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_font}{Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_footer_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_multiple_sources}{Multiple sources. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_source_name_2}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_2}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name_3}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_3}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_label}{Source label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_note}{Note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_footer_note_secondary}{Note (secondary). The secondary note is placed below the source and primary note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_advanced_note_styling}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_timestamp_format}{Time/date stamp formatting. Choices: \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M, \%H:\%M \%p, \%H:\%M:\%S, \%d \%b \%Y, \%H:\%M, \%d \%b \%Y, \%H:\%M, \%d/\%m/\%Y, \%H:\%M, \%m/\%d/\%Y, \%H:\%M, \%d-\%m-\%Y, \%H:\%M, \%m-\%d-\%Y, \%d \%b \%Y, \%m/\%d/\%Y, \%d/\%m/\%Y, \%m-\%d-\%Y, \%d-\%m-\%Y, \%Y-\%m-\%d, \%Y-\%m-\%d \%H:\%M:\%S. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src_light}{Image (light version). If provided this version will be used whenever the background color is dark. Flourish type hint: hidden, defaults to \code{NULL}}

\item{chart_layout_footer_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_margin}{Margin. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_order}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_align_vertical}{V. align. Choices: flex-start, center, flex-end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_space}{Space. Space between border and footer text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_screenreader_text_primary}{Screenreader description. A text alternative to the visual content that will only be visible to screenreaders, e.g. The line chart shows China consistently higher than the other countries since 1990.  Do not replicate your title, since that will also be read by screenreaders. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_layout_screenreader_label}{Screenreader label. A short text label given to the main Flourish embed wrapper to provide an accessible name that is only visible to screenreaders Added in the form of an "aria-label". Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_screenreader_hide_primary}{Screenreader mode for main visual container. Choices: \code{TRUE}, \code{FALSE}. Whether the main visual container is visible to screenreaders (Text in the header and footer are always available to screenreaders.). Flourish type hint: boolean, defaults to \code{NULL}}
}
\value{
A Flourish chart
}
\description{
Set chord diagram chart details
}
\examples{
try(
  flourish(chart_type = "chord_diagram", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_chord_diagram_data(gapminder) |>
    set_chord_diagram_details()
)
}
