% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_draw_the_line_details.R
\name{set_draw_the_line_details}
\alias{set_draw_the_line_details}
\title{Set draw the line chart chart details}
\usage{
set_draw_the_line_details(
  .,
  guess_lines_start = NULL,
  guess_lines_width = NULL,
  guess_lines_dot_radius = NULL,
  guess_lines_dash = NULL,
  fixed_lines_width = NULL,
  fixed_lines_dot_radius = NULL,
  line_curve = NULL,
  guess_lines_panel_position = NULL,
  message_styling = NULL,
  message_background = NULL,
  message_opacity = NULL,
  message_color = NULL,
  message_font_size = NULL,
  message_border = NULL,
  message_border_width = NULL,
  message_border_radius = NULL,
  message_shadow = NULL,
  guess_lines_message_text = NULL,
  scores_3_min = NULL,
  scores_3_text = NULL,
  scores_2_min = NULL,
  scores_2_text = NULL,
  scores_1_min = NULL,
  scores_1_text = NULL,
  scores_0_text = NULL,
  guess_lines_submit_text = NULL,
  submit_styling = NULL,
  submit_background = NULL,
  submit_opacity = NULL,
  submit_color = NULL,
  submit_font_size = NULL,
  submit_border = NULL,
  submit_border_width = NULL,
  submit_border_radius = NULL,
  submit_shadow = NULL,
  chart_color_scale_categorical_palette = NULL,
  chart_color_scale_categorical_extend = NULL,
  chart_color_scale_categorical_custom_palette = NULL,
  chart_color_scale_numeric_type = NULL,
  chart_color_scale_binning = NULL,
  chart_color_scale_bin_mode = NULL,
  chart_color_scale_bin_count = NULL,
  chart_color_scale_bin_thresholds = NULL,
  chart_color_scale_sequential_palette = NULL,
  chart_color_scale_sequential_reverse = NULL,
  chart_color_scale_sequential_custom_min = NULL,
  chart_color_scale_sequential_custom_max = NULL,
  chart_color_scale_sequential_color_space = NULL,
  chart_color_scale_sequential_custom_domain = NULL,
  chart_color_scale_sequential_domain_min = NULL,
  chart_color_scale_sequential_domain_max = NULL,
  chart_color_scale_diverging_palette = NULL,
  chart_color_scale_diverging_reverse = NULL,
  chart_color_scale_diverging_custom_min = NULL,
  chart_color_scale_diverging_custom_mid = NULL,
  chart_color_scale_diverging_custom_max = NULL,
  chart_color_scale_diverging_color_space = NULL,
  chart_color_scale_diverging_custom_domain = NULL,
  chart_color_scale_diverging_domain_min = NULL,
  chart_color_scale_diverging_domain_mid = NULL,
  chart_color_scale_diverging_domain_max = NULL,
  chart_categorical_legend_show_legend = NULL,
  chart_categorical_legend_title_mode = NULL,
  chart_categorical_legend_title = NULL,
  chart_categorical_legend_swatch_width = NULL,
  chart_categorical_legend_swatch_height = NULL,
  chart_categorical_legend_swatch_radius = NULL,
  chart_categorical_legend_legend_items_padding = NULL,
  chart_categorical_legend_swatch_outline = NULL,
  chart_categorical_legend_swatch_outline_color = NULL,
  chart_categorical_legend_order_override = NULL,
  chart_categorical_legend_icon_height = NULL,
  chart_categorical_legend_icon_color = NULL,
  chart_categorical_legend_max_width = NULL,
  chart_categorical_legend_orientation = NULL,
  chart_legend_container_alignment = NULL,
  chart_legend_container_orientation = NULL,
  chart_legend_container_title_weight = NULL,
  chart_legend_container_text_weight = NULL,
  chart_legend_container_text_color = NULL,
  chart_legend_container_text_size = NULL,
  chart_x_axis_axis_position = NULL,
  chart_x_axis_y_axis_position = NULL,
  chart_x_axis_numeric_scale_type = NULL,
  chart_x_axis_linear_min = NULL,
  chart_x_axis_linear_max = NULL,
  chart_x_axis_log_min = NULL,
  chart_x_axis_log_max = NULL,
  chart_x_axis_datetime_min = NULL,
  chart_x_axis_datetime_max = NULL,
  chart_x_axis_flip = NULL,
  chart_x_axis_show_scale_settings = NULL,
  chart_x_axis_nice = NULL,
  chart_x_axis_zero_axis = NULL,
  chart_x_axis_title_mode = NULL,
  chart_x_axis_title = NULL,
  chart_x_axis_title_styling = NULL,
  chart_x_axis_title_weight = NULL,
  chart_x_axis_title_color = NULL,
  chart_x_axis_title_size = NULL,
  chart_x_axis_title_padding = NULL,
  chart_x_axis_tick_label_position = NULL,
  chart_x_axis_tick_label_styling = NULL,
  chart_x_axis_tick_label_size = NULL,
  chart_x_axis_tick_label_color = NULL,
  chart_x_axis_tick_padding = NULL,
  chart_x_axis_tick_label_angle = NULL,
  chart_x_axis_tick_label_weight = NULL,
  chart_x_axis_tick_label_max_lines = NULL,
  chart_x_axis_tick_label_line_height = NULL,
  chart_x_axis_tick_label_space_mode = NULL,
  chart_x_axis_tick_label_space = NULL,
  chart_x_axis_tick_mode = NULL,
  chart_x_axis_tick_number = NULL,
  chart_x_axis_tick_custom = NULL,
  chart_x_axis_line_and_tick_color = NULL,
  chart_x_axis_line_and_tick_width = NULL,
  chart_x_axis_tick_length = NULL,
  chart_x_axis_tick_side = NULL,
  chart_x_axis_line_visible = NULL,
  chart_x_axis_edge_padding = NULL,
  chart_x_axis_gridlines_visible = NULL,
  chart_x_axis_gridlines_styling = NULL,
  chart_x_axis_gridline_color = NULL,
  chart_x_axis_gridline_style = NULL,
  chart_x_axis_gridline_width = NULL,
  chart_x_axis_gridline_category_dividers = NULL,
  chart_x_axis_gridline_category_dividers_extend = NULL,
  chart_y_axis_axis_visible = NULL,
  chart_y_axis_numeric_scale_type = NULL,
  chart_y_axis_linear_min = NULL,
  chart_y_axis_linear_max = NULL,
  chart_y_axis_log_min = NULL,
  chart_y_axis_log_max = NULL,
  chart_y_axis_datetime_min = NULL,
  chart_y_axis_datetime_max = NULL,
  chart_y_axis_flip = NULL,
  chart_y_axis_show_scale_settings = NULL,
  chart_y_axis_nice = NULL,
  chart_y_axis_zero_axis = NULL,
  chart_y_axis_title_mode = NULL,
  chart_y_axis_title = NULL,
  chart_y_axis_title_styling = NULL,
  chart_y_axis_title_position = NULL,
  chart_y_axis_title_weight = NULL,
  chart_y_axis_title_color = NULL,
  chart_y_axis_title_size = NULL,
  chart_y_axis_title_padding = NULL,
  chart_y_axis_tick_label_position = NULL,
  chart_y_axis_tick_label_styling = NULL,
  chart_y_axis_tick_label_size = NULL,
  chart_y_axis_tick_label_color = NULL,
  chart_y_axis_tick_padding = NULL,
  chart_y_axis_tick_label_angle = NULL,
  chart_y_axis_tick_label_weight = NULL,
  chart_y_axis_tick_label_max_lines = NULL,
  chart_y_axis_tick_label_line_height = NULL,
  chart_y_axis_tick_label_space_mode = NULL,
  chart_y_axis_tick_label_space = NULL,
  chart_y_axis_tick_mode = NULL,
  chart_y_axis_tick_number = NULL,
  chart_y_axis_tick_custom = NULL,
  chart_y_axis_line_and_tick_color = NULL,
  chart_y_axis_line_and_tick_width = NULL,
  chart_y_axis_tick_length = NULL,
  chart_y_axis_tick_side = NULL,
  chart_y_axis_line_visible = NULL,
  chart_y_axis_edge_padding = NULL,
  chart_y_axis_gridlines_visible = NULL,
  chart_y_axis_gridlines_styling = NULL,
  chart_y_axis_gridline_color = NULL,
  chart_y_axis_gridline_style = NULL,
  chart_y_axis_gridline_width = NULL,
  chart_y_axis_gridline_category_dividers = NULL,
  chart_y_axis_gridline_category_dividers_extend = NULL,
  chart_background_background_color_enabled = NULL,
  chart_background_background_image_enabled = NULL,
  chart_background_background_color = NULL,
  chart_background_background_image_src = NULL,
  chart_background_background_image_size = NULL,
  chart_background_background_image_position = NULL,
  datetime_input_format = NULL,
  date_format = NULL,
  chart_number_localization_input_decimal_separator = NULL,
  chart_number_localization_output_separators = NULL,
  chart_number_date_format_prefix = NULL,
  chart_number_date_format_suffix = NULL,
  chart_number_date_format_n_dec = NULL,
  chart_number_date_format_advanced = NULL,
  chart_number_date_format_negative_sign = NULL,
  chart_number_date_format_strip_zeros = NULL,
  chart_number_date_format_strip_separator = NULL,
  chart_number_date_format_transform_labels = NULL,
  chart_number_date_format_transform = NULL,
  chart_number_date_format_multiply_divide_constant = NULL,
  chart_number_date_format_exponentiate_constant = NULL,
  chart_layout_body_font = NULL,
  chart_layout_font_color = NULL,
  chart_layout_background_color_enabled = NULL,
  chart_layout_background_image_enabled = NULL,
  chart_layout_background_color = NULL,
  chart_layout_background_image_src = NULL,
  chart_layout_background_image_size = NULL,
  chart_layout_background_image_position = NULL,
  chart_layout_max_width_target = NULL,
  chart_layout_max_width = NULL,
  chart_layout_max_width_align = NULL,
  chart_layout_layout_order = NULL,
  chart_layout_space_between_sections = NULL,
  chart_layout_space_between_sections_custom = NULL,
  chart_layout_margin_top = NULL,
  chart_layout_margin_right = NULL,
  chart_layout_margin_bottom = NULL,
  chart_layout_margin_left = NULL,
  chart_layout_border_enabled = NULL,
  chart_layout_border_top_width = NULL,
  chart_layout_border_top_style = NULL,
  chart_layout_border_top_color = NULL,
  chart_layout_border_right_width = NULL,
  chart_layout_border_right_style = NULL,
  chart_layout_border_right_color = NULL,
  chart_layout_border_bottom_width = NULL,
  chart_layout_border_bottom_style = NULL,
  chart_layout_border_bottom_color = NULL,
  chart_layout_border_left_width = NULL,
  chart_layout_border_left_style = NULL,
  chart_layout_border_left_color = NULL,
  chart_layout_read_direction = NULL,
  chart_layout_font_size_mobile_small = NULL,
  chart_layout_font_size_mobile_big = NULL,
  chart_layout_font_size_tablet = NULL,
  chart_layout_font_size_desktop = NULL,
  chart_layout_font_size_big_screen = NULL,
  chart_layout_breakpoint_mobile_small = NULL,
  chart_layout_breakpoint_mobile_big = NULL,
  chart_layout_breakpoint_tablet = NULL,
  chart_layout_breakpoint_desktop = NULL,
  chart_layout_breakpoint_big_screen = NULL,
  chart_layout_header_align = NULL,
  chart_layout_title = NULL,
  chart_layout_title_styling = NULL,
  chart_layout_title_font = NULL,
  chart_layout_title_size = NULL,
  chart_layout_title_size_custom = NULL,
  chart_layout_title_weight = NULL,
  chart_layout_title_color = NULL,
  chart_layout_title_line_height = NULL,
  chart_layout_title_space_above = NULL,
  chart_layout_title_space_above_custom = NULL,
  chart_layout_subtitle = NULL,
  chart_layout_subtitle_styling = NULL,
  chart_layout_subtitle_font = NULL,
  chart_layout_subtitle_size = NULL,
  chart_layout_subtitle_size_custom = NULL,
  chart_layout_subtitle_weight = NULL,
  chart_layout_subtitle_color = NULL,
  chart_layout_subtitle_line_height = NULL,
  chart_layout_subtitle_space_above = NULL,
  chart_layout_subtitle_space_above_custom = NULL,
  chart_layout_header_text = NULL,
  chart_layout_header_text_styling = NULL,
  chart_layout_header_text_size = NULL,
  chart_layout_header_text_size_custom = NULL,
  chart_layout_header_text_weight = NULL,
  chart_layout_header_text_color = NULL,
  chart_layout_header_text_line_height = NULL,
  chart_layout_header_text_space_above = NULL,
  chart_layout_header_text_space_above_custom = NULL,
  chart_layout_header_border = NULL,
  chart_layout_header_border_width = NULL,
  chart_layout_header_border_color = NULL,
  chart_layout_header_border_style = NULL,
  chart_layout_header_border_space = NULL,
  chart_layout_header_logo_enabled = NULL,
  chart_layout_header_logo_src = NULL,
  chart_layout_header_logo_alt = NULL,
  chart_layout_header_logo_link_url = NULL,
  chart_layout_header_logo_height = NULL,
  chart_layout_header_logo_align = NULL,
  chart_layout_header_logo_position_inside = NULL,
  chart_layout_header_logo_position_outside = NULL,
  chart_layout_header_logo_margin_top = NULL,
  chart_layout_header_logo_margin_right = NULL,
  chart_layout_header_logo_margin_bottom = NULL,
  chart_layout_header_logo_margin_left = NULL,
  chart_layout_footer_align = NULL,
  chart_layout_footer_text_size = NULL,
  chart_layout_footer_text_color = NULL,
  chart_layout_footer_styling = NULL,
  chart_layout_footer_font = NULL,
  chart_layout_footer_text_weight = NULL,
  chart_layout_source_name = NULL,
  chart_layout_source_url = NULL,
  chart_layout_multiple_sources = NULL,
  chart_layout_source_name_2 = NULL,
  chart_layout_source_url_2 = NULL,
  chart_layout_source_name_3 = NULL,
  chart_layout_source_url_3 = NULL,
  chart_layout_source_label = NULL,
  chart_layout_footer_note = NULL,
  chart_layout_footer_note_secondary = NULL,
  chart_layout_advanced_note_styling = NULL,
  chart_layout_footer_timestamp_format = NULL,
  chart_layout_footer_logo_enabled = NULL,
  chart_layout_footer_logo_src = NULL,
  chart_layout_footer_logo_src_light = NULL,
  chart_layout_footer_logo_alt = NULL,
  chart_layout_footer_logo_link_url = NULL,
  chart_layout_footer_logo_height = NULL,
  chart_layout_footer_logo_margin = NULL,
  chart_layout_footer_logo_order = NULL,
  chart_layout_footer_align_vertical = NULL,
  chart_layout_footer_border = NULL,
  chart_layout_footer_border_width = NULL,
  chart_layout_footer_border_color = NULL,
  chart_layout_footer_border_style = NULL,
  chart_layout_footer_border_space = NULL,
  chart_layout_screenreader_text_primary = NULL,
  chart_layout_screenreader_label = NULL,
  chart_layout_screenreader_hide_primary = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{guess_lines_start}{Percent to reveal. What percentage of the line to show before a user starts drawing. Flourish type hint: number, defaults to \code{NULL}}

\item{guess_lines_width}{Width. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{guess_lines_dot_radius}{Dot radius. Flourish type hint: number, defaults to \code{NULL}}

\item{guess_lines_dash}{Drawing dash size. Only applies to the drawn section of the line. Set to zero for a solid line. Flourish type hint: number, defaults to \code{NULL}}

\item{fixed_lines_width}{Width. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{fixed_lines_dot_radius}{Dot radius. Flourish type hint: number, defaults to \code{NULL}}

\item{line_curve}{Live curve. Choices: linear, step, curve. Flourish type hint: string, defaults to \code{NULL}}

\item{guess_lines_panel_position}{Position. Choices: top, middle, bottom. Flourish type hint: string, defaults to \code{NULL}}

\item{message_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{message_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{message_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{message_color}{Font color. Flourish type hint: color, defaults to \code{NULL}}

\item{message_font_size}{Font size. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{message_border}{Border color. Flourish type hint: color, defaults to \code{NULL}}

\item{message_border_width}{Border width. In pixels. Flourish type hint: number, defaults to \code{NULL}}

\item{message_border_radius}{Border radius. In rems, a multiple of the panels font size. Flourish type hint: number, defaults to \code{NULL}}

\item{message_shadow}{Drop shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{guess_lines_message_text}{Text. Include \code{{{LINE}}} to show the lines name, in color. Flourish type hint: string, defaults to \code{NULL}}

\item{scores_3_min}{\% error. Specified as the maximum percentage error allowed, relative to the total Y axis. Flourish type hint: number, defaults to \code{NULL}}

\item{scores_3_text}{Text. Use \code{{{DIFF}}} to show the average gap from the users guess to the real data point. Flourish type hint: string, defaults to \code{NULL}}

\item{scores_2_min}{\% error. Specified as the maximum percentage error allowed, relative to the total Y axis. Flourish type hint: number, defaults to \code{NULL}}

\item{scores_2_text}{Text. Flourish type hint: string, defaults to \code{NULL}}

\item{scores_1_min}{\% error. Specified as the maximum percentage error allowed, relative to the total Y axis. Flourish type hint: number, defaults to \code{NULL}}

\item{scores_1_text}{Text. Flourish type hint: string, defaults to \code{NULL}}

\item{scores_0_text}{Text. Flourish type hint: string, defaults to \code{NULL}}

\item{guess_lines_submit_text}{Text. In rems, a multiple of the pages base font size. Flourish type hint: string, defaults to \code{NULL}}

\item{submit_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{submit_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{submit_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_color}{Font color. Flourish type hint: color, defaults to \code{NULL}}

\item{submit_font_size}{Font size. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_border}{Border colour. Flourish type hint: color, defaults to \code{NULL}}

\item{submit_border_width}{Border width. In pixels. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_border_radius}{Border radius. In ems a multiple of the panels font size. Flourish type hint: number, defaults to \code{NULL}}

\item{submit_shadow}{Drop shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_categorical_palette}{Palette. Flourish type hint: colors, defaults to \code{NULL}}

\item{chart_color_scale_categorical_extend}{Extend. Automatically generate additional colours when needed to avoid the palette colours being used more than once. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_categorical_custom_palette}{Custom overrides. Enter the label name for which you wish to set the color, followed by a colon and the desired color value.  Colors can be set using Hex, RGB, color names or RGBA, if you want to set the opacity Multiple colors can be set using multiple lines. For example:  Party 1: red Party 2: #4455AA Party 3: rgb(30,168,26). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_color_scale_numeric_type}{Scale type. Choices: sequential, diverging. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_binning}{Choices: \code{FALSE}, \code{TRUE}. In linear mode, the color scale will run as a smooth gradient between 2 colors. In binned mode, the gradient will be divided in smaller blocks. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_bin_mode}{Binning mode. Choices: fixed, quantile, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_bin_count}{Number of bins. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_bin_thresholds}{Custom thresholds. Enter your desired thresholds, separating them with a ";". For instance, "5;10;15".  Bins form as follows:  From the datas minimum value (domain minimum) up to the first threshold.  Between consecutive thresholds.  From the last threshold to the datas maximum value (domain maximum).   For "5;10;15", youll get four bins based on your datas range. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_mid}{Midpoint color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_mid}{Mid. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_show_legend}{Legend mode. Choices: \code{TRUE}, \code{FALSE}. A legend will not show with a single entry. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_categorical_legend_title_mode}{Title mode. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_categorical_legend_title}{Title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_radius}{Roundness. The radius of the corners of the swatch (in pixels). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_legend_items_padding}{Padding. Padding between legend items ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_outline}{Outline. An optional outline for the swatch in the legend. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_outline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_categorical_legend_order_override}{Custom order override. Manually specify the order of legend entries (one entry per line). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_categorical_legend_icon_height}{Height. Height of icon ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_icon_color}{Color. Fallback color (icon color if not determined by template). Flourish type hint: color, defaults to \code{NULL}}

\item{chart_categorical_legend_max_width}{Max width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_alignment}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_title_weight}{Title weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_text_weight}{Text weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_legend_container_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_axis_position}{Position. Choices: bottom, float-below, float-above, top, off. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_y_axis_position}{Y value. Vertical position of the X axis on the Y scale. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_numeric_scale_type}{Type. Choices: linear, log. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_linear_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_linear_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_log_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_log_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_datetime_min}{Min (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_datetime_max}{Max (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_flip}{Flip axis. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_show_scale_settings}{Configure default min/max. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_nice}{Round min/max. Choices: \code{TRUE}, \code{FALSE}. Enabling this option rounds the X axis start and end values to tidy numbers, such as 0, 50, 100 or whole years. Specified min and max values will override this feature. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_zero_axis}{Include zero. Choices: auto, on, off. Enable to extend the axis to include zero, even when data is all positive or all negative. This is ignored if you set specific min and max scale values. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_mode}{Type. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title}{Text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_title_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_title_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_position}{Position. Choices: default, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_tick_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_angle}{Angle. Choices: 0, 30, 45, 60, 90. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_max_lines}{Max lines. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_space_mode}{Space mode. Choices: auto, max, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_space}{Space. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_mode}{Mode. Choices: auto, number, custom. Sets the method for selecting tick values on the axis:  Auto automatically selects the ticks  Number allows you to choose the number of ticks to display.  Custom lets you select specific ticks to show.  Note: If the axis is categorical, Number mode is ignored. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_number}{Number. Approximate number of ticks or gridlines. The actual number will depend on the range of values, chart size, etc. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_custom}{One tick label per line. Specify the ticks to show, one per line. For dates, use the input format as specified in the datasheet.  You can also use \code{{{FIRST}}} and \code{{{LAST}}} to add ticks at the minimum and maximum of the axis. For example:  \code{{{FIRST}}} 25000 50000 75000 \code{{{LAST}}}. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_x_axis_line_and_tick_color}{Line color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_line_and_tick_width}{Line width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_length}{Tick length. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_side}{Tick side. Choices: out, in. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_line_visible}{Axis line. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_edge_padding}{Edge padding (\%). Space between start/end of axis line and first/last category tick as a percentage of the spacing between ticks. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_gridlines_visible}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridlines_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_gridline_style}{Style. Choices: solid, dash, dot, dot_dash. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_gridline_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_gridline_category_dividers}{Put lines between categories. On a categorical axis, this option disables the gridline on each tick and instead adds rules between each category to give a table-like design. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridline_category_dividers_extend}{Extend. Extends the dividers into the axis margins. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_axis_visible}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_numeric_scale_type}{Type. Choices: linear, log. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_linear_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_linear_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_log_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_log_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_datetime_min}{Min (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_datetime_max}{Max (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_flip}{Flip axis. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_show_scale_settings}{Configure default min/max. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_nice}{Round min/max. Choices: \code{TRUE}, \code{FALSE}. When enabled, scales automatically extend to "nice" rounded start/end values, such as hundreds or whole years. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_zero_axis}{Include zero. Choices: auto, on, off. Whether to extend the axis to include zero if the data values are all positive or all negative. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_mode}{Type. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title}{Text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_title_position}{Position. Choices: side, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_title_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_title_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_position}{Position. Choices: default, above, below. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_tick_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_angle}{Angle. Choices: 0, 30, 45, 60, 90. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_max_lines}{Max lines. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_space_mode}{Space mode. Choices: auto, max, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_space}{Space. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_mode}{Mode. Choices: auto, number, custom. Sets the method for selecting tick values on the axis:  Auto automatically selects the ticks  Number allows you to choose the number of ticks to display.  Custom lets you select specific ticks to show.  Note: If the axis is categorical, Number mode is ignored. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_number}{Number. Approximate number of ticks or gridlines. The actual number will depend on the range of values, chart size, etc. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_custom}{One tick label per line. Specify the ticks to show, one per line. For dates, use the input format as specified in the datasheet.  You can also use \code{{{FIRST}}} and \code{{{LAST}}} to add ticks at the minimum and maximum of the axis. For example:  \code{{{FIRST}}} 25000 50000 75000 \code{{{LAST}}}. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_y_axis_line_and_tick_color}{Line color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_line_and_tick_width}{Line width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_length}{Tick length. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_side}{Side. Choices: out, in. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_line_visible}{Axis line. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_edge_padding}{Edge padding (\%). Space between start/end of axis line and first/last category tick as a percentage of the spacing between ticks. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_gridlines_visible}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_gridlines_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_gridline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_gridline_style}{Style. Choices: solid, dash, dot, dot_dash. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_gridline_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_gridline_category_dividers}{Put lines between categories. On a categorical axis, this option disables the gridline on each tick and instead adds rules between each category to give a table-like design. Ignored for date and numeric axes. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_gridline_category_dividers_extend}{Extend. Extends the dividers into the axis margins. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_background_background_color_enabled}{Color. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_background_background_image_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_background_background_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_background_background_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_background_background_image_size}{Size. Choices: stretch, slice, meet. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_background_background_image_position}{Position. Choices: x_min_y_min, x_mid_y_min, x_max_y_min, x_max_y_mid, x_mid_y_mid, x_min_y_max, x_mid_y_max, x_max_y_max. Flourish type hint: string, defaults to \code{NULL}}

\item{datetime_input_format}{Format in data sheet. Choices: auto, \%Y-\%m-\%dT\%H:\%M:\%S.\%LZ, \%Y-\%m-\%d, \%m/\%d/\%Y, \%d/\%m/\%Y, \%d-\%b-\%y, \%m/\%Y, \%m/\%y, \%b \%Y, \%b \%y, \%B \%d, \%d \%b, \%Y, \%y, \%B, \%b, \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M. The date/time format in the datasheet. Leave as Auto for Flourish to make a best guess. If it isnt in the list, you can enter a custom format using d3-time-format syntax. See NPM d3-data-format for details. Flourish type hint: string, defaults to \code{NULL}}

\item{date_format}{Display format. Choices: auto, \%Y-\%m-\%dT\%H:\%M:\%S.\%LZ, \%Y-\%m-\%d, \%m/\%d/\%Y, \%d/\%m/\%Y, \%d-\%b-\%y, \%m/\%Y, \%m/\%y, \%b \%Y, \%b \%y, \%B \%d, \%d \%b, \%Y, \%y, \%B, \%b, \%A, \%a, \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M. The date/time format to show in the graphic. If it isnt in the list, you can enter a custom format using d3-time-format syntax. See NPM d3-data-format for details. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_localization_input_decimal_separator}{Decimal separator in data sheet. Choices: ., ,. Used for interpreting your data. Only change if data is not displaying on the chart as expected. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_localization_output_separators}{Number format to display. Choices: ,., .,, ., ,, ., ,. How the numbers should appear on chart labels. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_prefix}{Prefix. Text to place in front of number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_suffix}{Suffix. Text to place after number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_n_dec}{Decimal places. Enter a negative integer to round to a whole number with that many zeros. For example, -2 will round to the nearest hundred. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_negative_sign}{Styling of negative numbers. Choices: -$nk, $-nk, ($nk), $(n)k, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_strip_zeros}{Remove trailing zeros. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_strip_separator}{Hide thousands separator below 10,000. Turn off if you want four-digit numbers to include a separator, e.g. 1,234 rather than 1234. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_transform_labels}{Multiply/divide values. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_transform}{Choices: multiply, divide, exponentiate. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_multiply_divide_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_exponentiate_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_body_font}{Main font. This font will apply to the whole graphic by default. You can optionally change the font for the title, subtitle, footer, etc in the Header and Footer settings panels. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_font_color}{Text color. This color will apply to the whole graphic by default, You can optionally change the color for individual text elements, in other settings panels. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_color_enabled}{Color. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_image_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_background_image_size}{Size. Choices: cover, contain, auto, 100\% 100\%. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_background_image_position}{Position. Choices: top left, top center, top right, center left, center center, center right, bottom left, bottom center, bottom right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width_target}{Maximum width. Choices: none, wrapper, primary. Apply a maximum width to just the main graphic or everything (main graphic plus header, footer, etc). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width}{Maximum width. Leave blank to stretch to container width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_max_width_align}{Align. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_layout_order}{Layout order. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections}{Space between sections. Choices: 0.5, 1, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_enabled}{Show borders around visualisation. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_border_top_width}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_top_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_top_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_right_width}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_right_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_right_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_bottom_width}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_bottom_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_bottom_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_left_width}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_left_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_left_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_read_direction}{Read direction. Choices: ltr, rtl. This will change the reading direction of the main text elements on the page. Its not possible to adjust this on all elements, such as axes.  Note that when direction is set to right to left any alignment icons will be reversed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_align}{Alignment. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_title_font}{Title Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_title_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_title_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_subtitle_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_subtitle_font}{Subtitle Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_subtitle_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_subtitle_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_header_text_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_text_size}{Size. Choices: 1.2, 1.4, 1.6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_text_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_space}{Space. Space between border and header text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_enabled}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_header_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_align}{Align. Choices: inside, outside. Align logo inside either the header or the main visualization container. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_inside}{Position. Choices: top, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_outside}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_align}{Alignment. Choices: left, center, right, justify. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_styling}{Advanced footer styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_font}{Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_footer_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_multiple_sources}{Multiple sources. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_source_name_2}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_2}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name_3}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_3}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_label}{Source label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_note}{Note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_footer_note_secondary}{Note (secondary). The secondary note is placed below the source and primary note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_advanced_note_styling}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_timestamp_format}{Time/date stamp formatting. Choices: \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M, \%H:\%M \%p, \%H:\%M:\%S, \%d \%b \%Y, \%H:\%M, \%d \%b \%Y, \%H:\%M, \%d/\%m/\%Y, \%H:\%M, \%m/\%d/\%Y, \%H:\%M, \%d-\%m-\%Y, \%H:\%M, \%m-\%d-\%Y, \%d \%b \%Y, \%m/\%d/\%Y, \%d/\%m/\%Y, \%m-\%d-\%Y, \%d-\%m-\%Y, \%Y-\%m-\%d, \%Y-\%m-\%d \%H:\%M:\%S. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src_light}{Image (light version). If provided this version will be used whenever the background color is dark. Flourish type hint: hidden, defaults to \code{NULL}}

\item{chart_layout_footer_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_margin}{Margin. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_order}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_align_vertical}{V. align. Choices: flex-start, center, flex-end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_space}{Space. Space between border and footer text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_screenreader_text_primary}{Screenreader description. A text alternative to the visual content that will only be visible to screenreaders, e.g. The line chart shows China consistently higher than the other countries since 1990.  Do not replicate your title, since that will also be read by screenreaders. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_layout_screenreader_label}{Screenreader label. A short text label given to the main Flourish embed wrapper to provide an accessible name that is only visible to screenreaders Added in the form of an "aria-label". Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_screenreader_hide_primary}{Screenreader mode for main visual container. Choices: \code{TRUE}, \code{FALSE}. Whether the main visual container is visible to screenreaders (Text in the header and footer are always available to screenreaders.). Flourish type hint: boolean, defaults to \code{NULL}}
}
\value{
A Flourish chart
}
\description{
Set draw the line chart chart details
}
\examples{
try(
  flourish(chart_type = "draw_the_line", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_draw_the_line_data(gapminder) |>
    set_draw_the_line_details()
)
}
