% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_radar_details.R
\name{set_radar_details}
\alias{set_radar_details}
\title{Set radar chart chart details}
\usage{
set_radar_details(
  .,
  radars_curve = NULL,
  radars_star = NULL,
  radars_stroke_width = NULL,
  radars_stroke_opacity = NULL,
  radars_fill_opacity = NULL,
  radars_star_inner_radius = NULL,
  chart_color_scale_categorical_palette = NULL,
  chart_color_scale_categorical_extend = NULL,
  chart_color_scale_categorical_custom_palette = NULL,
  chart_color_scale_numeric_type = NULL,
  chart_color_scale_binning = NULL,
  chart_color_scale_bin_mode = NULL,
  chart_color_scale_bin_count = NULL,
  chart_color_scale_bin_thresholds = NULL,
  chart_color_scale_sequential_palette = NULL,
  chart_color_scale_sequential_reverse = NULL,
  chart_color_scale_sequential_custom_min = NULL,
  chart_color_scale_sequential_custom_max = NULL,
  chart_color_scale_sequential_color_space = NULL,
  chart_color_scale_sequential_custom_domain = NULL,
  chart_color_scale_sequential_domain_min = NULL,
  chart_color_scale_sequential_domain_max = NULL,
  chart_color_scale_diverging_palette = NULL,
  chart_color_scale_diverging_reverse = NULL,
  chart_color_scale_diverging_custom_min = NULL,
  chart_color_scale_diverging_custom_mid = NULL,
  chart_color_scale_diverging_custom_max = NULL,
  chart_color_scale_diverging_color_space = NULL,
  chart_color_scale_diverging_custom_domain = NULL,
  chart_color_scale_diverging_domain_min = NULL,
  chart_color_scale_diverging_domain_mid = NULL,
  chart_color_scale_diverging_domain_max = NULL,
  facet_comparisons = NULL,
  facet_comparison_stroke = NULL,
  facet_comparison_stroke_width = NULL,
  facet_comparison_stroke_opacity = NULL,
  facet_comparison_fill = NULL,
  facet_comparison_fill_opacity = NULL,
  chart_facet_columns_fixed = NULL,
  chart_facet_columns = NULL,
  chart_facet_min_width = NULL,
  chart_facet_min_width_fixed = NULL,
  chart_facet_custom_grid = NULL,
  chart_facet_custom_positions = NULL,
  chart_facet_max_width = NULL,
  chart_facet_gutter_w = NULL,
  chart_facet_gutter_h = NULL,
  chart_facet_animation = NULL,
  chart_facet_titles = NULL,
  chart_facet_title_align = NULL,
  chart_facet_title_size = NULL,
  chart_facet_title_unit = NULL,
  chart_facet_title_color = NULL,
  chart_facet_title_padding_top = NULL,
  chart_facet_title_padding_bottom = NULL,
  chart_facet_title_bg = NULL,
  chart_facet_title_bg_color = NULL,
  chart_facet_title_bg_padding = NULL,
  chart_radial_axis_axis_show = NULL,
  chart_radial_axis_x_line_color = NULL,
  chart_radial_axis_x_line_width = NULL,
  chart_radial_axis_x_tick_label_color = NULL,
  chart_radial_axis_x_tick_label_size = NULL,
  chart_radial_axis_x_tick_label_outline = NULL,
  chart_radial_axis_x_tick_label_weight = NULL,
  chart_radial_axis_y_min = NULL,
  chart_radial_axis_y_max = NULL,
  chart_radial_axis_y_tick_number = NULL,
  chart_radial_axis_y_line_color = NULL,
  chart_radial_axis_y_line_width = NULL,
  chart_radial_axis_y_line_dash = NULL,
  chart_radial_axis_y_line_dash_gap = NULL,
  chart_radial_axis_y_tick_label_color = NULL,
  chart_radial_axis_y_tick_label_size = NULL,
  chart_radial_axis_y_tick_label_outline = NULL,
  chart_radial_axis_y_tick_label_space = NULL,
  chart_radial_axis_y_tick_label_weight = NULL,
  chart_radial_axis_margin_top = NULL,
  chart_radial_axis_margin_right = NULL,
  chart_radial_axis_margin_bottom = NULL,
  chart_radial_axis_margin_left = NULL,
  chart_radial_axis_group_labels = NULL,
  chart_radial_axis_group_labels_color = NULL,
  chart_radial_axis_group_labels_size = NULL,
  chart_radial_axis_group_labels_letter_spacing = NULL,
  chart_radial_axis_group_labels_space_above = NULL,
  chart_radial_axis_group_labels_space_below = NULL,
  chart_legend_container_alignment = NULL,
  chart_legend_container_orientation = NULL,
  chart_legend_container_title_weight = NULL,
  chart_legend_container_text_weight = NULL,
  chart_legend_container_text_color = NULL,
  chart_legend_container_text_size = NULL,
  chart_legend_show_legend = NULL,
  chart_legend_title_mode = NULL,
  chart_legend_title = NULL,
  chart_legend_swatch_width = NULL,
  chart_legend_swatch_height = NULL,
  chart_legend_swatch_radius = NULL,
  chart_legend_legend_items_padding = NULL,
  chart_legend_swatch_outline = NULL,
  chart_legend_swatch_outline_color = NULL,
  chart_legend_order_override = NULL,
  chart_legend_icon_height = NULL,
  chart_legend_icon_color = NULL,
  chart_legend_max_width = NULL,
  chart_legend_orientation = NULL,
  facet_control_show = NULL,
  facet_control_default = NULL,
  facet_control_none = NULL,
  facets_label_combined = NULL,
  facet_control_individual = NULL,
  facets_label_separate = NULL,
  chart_facet_filter_controls_control_type = NULL,
  chart_facet_filter_controls_control_styles = NULL,
  chart_facet_filter_controls_control_title = NULL,
  chart_facet_filter_controls_button_group = NULL,
  chart_facet_filter_controls_button_group_width_mode = NULL,
  chart_facet_filter_controls_button_group_width_fixed = NULL,
  chart_facet_filter_controls_dropdown_width_mode = NULL,
  chart_facet_filter_controls_dropdown_width_fixed = NULL,
  chart_facet_filter_controls_slider_width = NULL,
  chart_facet_filter_controls_slider_max_label_width = NULL,
  chart_facet_filter_controls_slider_step_time = NULL,
  chart_facet_filter_controls_slider_loop = NULL,
  chart_facet_filter_controls_slider_restart_pause = NULL,
  chart_facet_filter_controls_sort = NULL,
  chart_filter_controls_control_type = NULL,
  chart_filter_controls_control_styles = NULL,
  chart_filter_controls_control_title = NULL,
  chart_filter_controls_button_group = NULL,
  chart_filter_controls_button_group_width_mode = NULL,
  chart_filter_controls_button_group_width_fixed = NULL,
  chart_filter_controls_dropdown_width_mode = NULL,
  chart_filter_controls_dropdown_width_fixed = NULL,
  chart_filter_controls_slider_width = NULL,
  chart_filter_controls_slider_max_label_width = NULL,
  chart_filter_controls_slider_step_time = NULL,
  chart_filter_controls_slider_loop = NULL,
  chart_filter_controls_slider_restart_pause = NULL,
  chart_filter_controls_sort = NULL,
  filter_control_all = NULL,
  filter_y_max = NULL,
  chart_controls_style_font_size = NULL,
  chart_controls_style_font_weight = NULL,
  chart_controls_style_height = NULL,
  chart_button_control_style_background = NULL,
  chart_button_control_style_background_selected = NULL,
  chart_button_control_style_background_hover = NULL,
  chart_button_control_style_font_color = NULL,
  chart_button_control_style_font_color_selected = NULL,
  chart_button_control_style_font_color_hover = NULL,
  chart_button_control_style_button_styles_advanced = NULL,
  chart_button_control_style_border_width = NULL,
  chart_button_control_style_border_color = NULL,
  chart_button_control_style_border_transparency = NULL,
  chart_button_control_style_border_radius = NULL,
  chart_dropdown_control_style_background = NULL,
  chart_dropdown_control_style_font_color = NULL,
  chart_dropdown_control_style_border_styles_advanced = NULL,
  chart_dropdown_control_style_border_style = NULL,
  chart_dropdown_control_style_border_width = NULL,
  chart_dropdown_control_style_border_color = NULL,
  chart_dropdown_control_style_border_transparency = NULL,
  chart_dropdown_control_style_border_radius = NULL,
  chart_slider_control_style_track_color = NULL,
  chart_slider_control_style_font_color = NULL,
  chart_slider_control_style_handle_color = NULL,
  chart_slider_control_style_play_color = NULL,
  chart_slider_control_style_track_height = NULL,
  chart_slider_control_style_handle_height = NULL,
  chart_slider_control_style_play_button = NULL,
  chart_filter_controls_container_alignment = NULL,
  chart_filter_controls_container_controls_spacing = NULL,
  chart_popup_mode = NULL,
  chart_popup_popup_is_custom = NULL,
  chart_popup_show_popup_styles = NULL,
  chart_popup_popup_custom_header = NULL,
  chart_popup_popup_custom_main = NULL,
  chart_popup_popup_custom_accessible = NULL,
  chart_popup_popup_width = NULL,
  chart_popup_popup_font_size = NULL,
  chart_popup_popup_text_color = NULL,
  chart_popup_popup_align = NULL,
  chart_popup_popup_background = NULL,
  chart_popup_popup_background_opacity = NULL,
  chart_popup_popup_padding = NULL,
  chart_popup_popup_radius = NULL,
  chart_popup_popup_header_type = NULL,
  chart_popup_popup_header_background = NULL,
  chart_popup_popup_header_text_color = NULL,
  chart_popup_popup_list_type = NULL,
  chart_popup_popup_list_label_weight = NULL,
  chart_popup_popup_list_separator = NULL,
  chart_popup_popup_shadow = NULL,
  chart_popup_popup_pointer = NULL,
  chart_popup_panel_position = NULL,
  chart_popup_panel_open_duration = NULL,
  chart_popup_panel_side_fixed = NULL,
  chart_popup_panel_side_default_content = NULL,
  chart_popup_panel_is_custom = NULL,
  chart_popup_show_panel_styles = NULL,
  chart_popup_panel_custom_header = NULL,
  chart_popup_panel_custom_main = NULL,
  chart_popup_panel_width_overlay = NULL,
  chart_popup_panel_height_overlay = NULL,
  chart_popup_panel_width_side = NULL,
  chart_popup_panel_height_side = NULL,
  chart_popup_panel_max_width = NULL,
  chart_popup_panel_max_height = NULL,
  chart_popup_panel_vertical_position = NULL,
  chart_popup_panel_text_color = NULL,
  chart_popup_panel_font_size = NULL,
  chart_popup_panel_align = NULL,
  chart_popup_panel_background = NULL,
  chart_popup_panel_background_opacity = NULL,
  chart_popup_panel_padding = NULL,
  chart_popup_panel_radius = NULL,
  chart_popup_panel_list_type = NULL,
  chart_popup_panel_list_label_weight = NULL,
  chart_popup_panel_list_separator = NULL,
  chart_popup_panel_shadow = NULL,
  popup_color_headings = NULL,
  duration = NULL,
  chart_number_date_format_prefix = NULL,
  chart_number_date_format_suffix = NULL,
  chart_number_date_format_n_dec = NULL,
  chart_number_date_format_advanced = NULL,
  chart_number_date_format_negative_sign = NULL,
  chart_number_date_format_strip_zeros = NULL,
  chart_number_date_format_strip_separator = NULL,
  chart_number_date_format_transform_labels = NULL,
  chart_number_date_format_transform = NULL,
  chart_number_date_format_multiply_divide_constant = NULL,
  chart_number_date_format_exponentiate_constant = NULL,
  chart_layout_body_font = NULL,
  chart_layout_font_color = NULL,
  chart_layout_background_color_enabled = NULL,
  chart_layout_background_image_enabled = NULL,
  chart_layout_background_color = NULL,
  chart_layout_background_image_src = NULL,
  chart_layout_background_image_size = NULL,
  chart_layout_background_image_position = NULL,
  chart_layout_max_width_target = NULL,
  chart_layout_max_width = NULL,
  chart_layout_max_width_align = NULL,
  chart_layout_layout_order = NULL,
  chart_layout_space_between_sections = NULL,
  chart_layout_space_between_sections_custom = NULL,
  chart_layout_margin_top = NULL,
  chart_layout_margin_right = NULL,
  chart_layout_margin_bottom = NULL,
  chart_layout_margin_left = NULL,
  chart_layout_border_enabled = NULL,
  chart_layout_border_top_width = NULL,
  chart_layout_border_top_style = NULL,
  chart_layout_border_top_color = NULL,
  chart_layout_border_right_width = NULL,
  chart_layout_border_right_style = NULL,
  chart_layout_border_right_color = NULL,
  chart_layout_border_bottom_width = NULL,
  chart_layout_border_bottom_style = NULL,
  chart_layout_border_bottom_color = NULL,
  chart_layout_border_left_width = NULL,
  chart_layout_border_left_style = NULL,
  chart_layout_border_left_color = NULL,
  chart_layout_read_direction = NULL,
  chart_layout_font_size_mobile_small = NULL,
  chart_layout_font_size_mobile_big = NULL,
  chart_layout_font_size_tablet = NULL,
  chart_layout_font_size_desktop = NULL,
  chart_layout_font_size_big_screen = NULL,
  chart_layout_breakpoint_mobile_small = NULL,
  chart_layout_breakpoint_mobile_big = NULL,
  chart_layout_breakpoint_tablet = NULL,
  chart_layout_breakpoint_desktop = NULL,
  chart_layout_breakpoint_big_screen = NULL,
  chart_layout_header_align = NULL,
  chart_layout_title = NULL,
  chart_layout_title_styling = NULL,
  chart_layout_title_font = NULL,
  chart_layout_title_size = NULL,
  chart_layout_title_size_custom = NULL,
  chart_layout_title_weight = NULL,
  chart_layout_title_color = NULL,
  chart_layout_title_line_height = NULL,
  chart_layout_title_space_above = NULL,
  chart_layout_title_space_above_custom = NULL,
  chart_layout_subtitle = NULL,
  chart_layout_subtitle_styling = NULL,
  chart_layout_subtitle_font = NULL,
  chart_layout_subtitle_size = NULL,
  chart_layout_subtitle_size_custom = NULL,
  chart_layout_subtitle_weight = NULL,
  chart_layout_subtitle_color = NULL,
  chart_layout_subtitle_line_height = NULL,
  chart_layout_subtitle_space_above = NULL,
  chart_layout_subtitle_space_above_custom = NULL,
  chart_layout_header_text = NULL,
  chart_layout_header_text_styling = NULL,
  chart_layout_header_text_size = NULL,
  chart_layout_header_text_size_custom = NULL,
  chart_layout_header_text_weight = NULL,
  chart_layout_header_text_color = NULL,
  chart_layout_header_text_line_height = NULL,
  chart_layout_header_text_space_above = NULL,
  chart_layout_header_text_space_above_custom = NULL,
  chart_layout_header_border = NULL,
  chart_layout_header_border_width = NULL,
  chart_layout_header_border_color = NULL,
  chart_layout_header_border_style = NULL,
  chart_layout_header_border_space = NULL,
  chart_layout_header_logo_enabled = NULL,
  chart_layout_header_logo_src = NULL,
  chart_layout_header_logo_alt = NULL,
  chart_layout_header_logo_link_url = NULL,
  chart_layout_header_logo_height = NULL,
  chart_layout_header_logo_align = NULL,
  chart_layout_header_logo_position_inside = NULL,
  chart_layout_header_logo_position_outside = NULL,
  chart_layout_header_logo_margin_top = NULL,
  chart_layout_header_logo_margin_right = NULL,
  chart_layout_header_logo_margin_bottom = NULL,
  chart_layout_header_logo_margin_left = NULL,
  chart_layout_footer_align = NULL,
  chart_layout_footer_text_size = NULL,
  chart_layout_footer_text_color = NULL,
  chart_layout_footer_styling = NULL,
  chart_layout_footer_font = NULL,
  chart_layout_footer_text_weight = NULL,
  chart_layout_source_name = NULL,
  chart_layout_source_url = NULL,
  chart_layout_multiple_sources = NULL,
  chart_layout_source_name_2 = NULL,
  chart_layout_source_url_2 = NULL,
  chart_layout_source_name_3 = NULL,
  chart_layout_source_url_3 = NULL,
  chart_layout_source_label = NULL,
  chart_layout_footer_note = NULL,
  chart_layout_footer_note_secondary = NULL,
  chart_layout_advanced_note_styling = NULL,
  chart_layout_footer_timestamp_format = NULL,
  chart_layout_footer_logo_enabled = NULL,
  chart_layout_footer_logo_src = NULL,
  chart_layout_footer_logo_src_light = NULL,
  chart_layout_footer_logo_alt = NULL,
  chart_layout_footer_logo_link_url = NULL,
  chart_layout_footer_logo_height = NULL,
  chart_layout_footer_logo_margin = NULL,
  chart_layout_footer_logo_order = NULL,
  chart_layout_footer_align_vertical = NULL,
  chart_layout_footer_border = NULL,
  chart_layout_footer_border_width = NULL,
  chart_layout_footer_border_color = NULL,
  chart_layout_footer_border_style = NULL,
  chart_layout_footer_border_space = NULL,
  chart_layout_screenreader_text_primary = NULL,
  chart_layout_screenreader_label = NULL,
  chart_layout_screenreader_hide_primary = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{radars_curve}{Line curve. Choices: linear, cardinal. Flourish type hint: string, defaults to \code{NULL}}

\item{radars_star}{Line shape. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{radars_stroke_width}{Line width. Flourish type hint: number, defaults to \code{NULL}}

\item{radars_stroke_opacity}{Line opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{radars_fill_opacity}{Fill opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{radars_star_inner_radius}{Star inner. The star shapes inner radius, as a percent of the max radius. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_categorical_palette}{Flourish type hint: colors, defaults to \code{NULL}}

\item{chart_color_scale_categorical_extend}{Extend. Automatically generate additional colours when needed to avoid the palette colours being used more than once. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_categorical_custom_palette}{Custom overrides. Enter the label name for which you wish to set the color, followed by a colon and the desired color value.  Colors can be set using Hex, RGB, color names or RGBA, if you want to set the opacity Multiple colors can be set using multiple lines. For example:  Party 1: red Party 2: #4455AA Party 3: rgb(30,168,26). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_color_scale_numeric_type}{Scale type. Choices: sequential, diverging. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_binning}{Choices: \code{FALSE}, \code{TRUE}. In linear mode, the color scale will run as a smooth gradient between 2 colors. In binned mode, the gradient will be divided in smaller blocks. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_bin_mode}{Binning mode. Choices: fixed, quantile, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_bin_count}{Number of bins. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_bin_thresholds}{Custom thresholds. Enter your desired thresholds, separating them with a ";". For instance, "5;10;15".  Bins form as follows:  From the datas minimum value (domain minimum) up to the first threshold.  Between consecutive thresholds.  From the last threshold to the datas maximum value (domain maximum).   For "5;10;15", youll get four bins based on your datas range. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_mid}{Midpoint color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_mid}{Mid. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{facet_comparisons}{Show comparison lines in grid mode. Flourish type hint: boolean, defaults to \code{NULL}}

\item{facet_comparison_stroke}{Line. Flourish type hint: color, defaults to \code{NULL}}

\item{facet_comparison_stroke_width}{Line width. Flourish type hint: number, defaults to \code{NULL}}

\item{facet_comparison_stroke_opacity}{Line opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{facet_comparison_fill}{Fill. Flourish type hint: color, defaults to \code{NULL}}

\item{facet_comparison_fill_opacity}{Fill opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_columns_fixed}{Grid mode. Choices: \code{FALSE}, \code{TRUE}. In Auto mode, the number of columns will change with screen size. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_facet_columns}{Columns. The number of columns in the grid. (If specifying custom positions, extra columns may be added automatically.). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_min_width}{Min chart width. The minimum width of a chart or cell in the grid. Determines the number of columns. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_min_width_fixed}{Min chart width. Specify the minimum width of a chart or cell in the grid, below which the Fixed grid will be ignored and it will fall back to responsive Auto mode. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_custom_grid}{Add custom positions. Specify positions for individual charts in the grid useful for making a map layout. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_facet_custom_positions}{Custom positions. Enter the name of a facet, one per line, using the format: Facet name: row, column where row and column are numbers indicating the grid position where a facet should be positioned.  For example, Alaska : 1, 1 Maine : 1, 11 Wisconsin : 2, 6 Flourish type hint: text, defaults to \code{NULL}}

\item{chart_facet_max_width}{Max chart width. The maximum width of a chart or cell in the grid. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_gutter_w}{Horizontal margin. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_gutter_h}{Vertical margin. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_animation}{On data change. Choices: auto, name, reuse. If a chart in the grid is removed, that cell can either be animated out (with another cell moving to take its place) or reused for the new data in that cell. Auto decides for you based on the data. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_facet_titles}{Show titles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_facet_title_align}{Alignment. Choices: auto, left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_facet_title_size}{Size. Relative to the width of the chart. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_title_unit}{Scale with chart. Choices: rem, facet. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_facet_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_facet_title_padding_top}{Space above. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_title_padding_bottom}{Space below. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_title_bg}{Background. Choices: none, full, text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_facet_title_bg_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_facet_title_bg_padding}{Background padding. Horizontal space between text and background. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_axis_show}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_radial_axis_x_line_color}{Line. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_radial_axis_x_line_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_x_tick_label_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_radial_axis_x_tick_label_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_x_tick_label_outline}{Outline. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_x_tick_label_weight}{Choices: 600, 400. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_radial_axis_y_min}{Min value. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_y_max}{Max value. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_y_tick_number}{Number. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_y_line_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_radial_axis_y_line_width}{Width. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_radial_axis_y_line_dash}{Dash. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_radial_axis_y_line_dash_gap}{Gap. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_radial_axis_y_tick_label_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_radial_axis_y_tick_label_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_y_tick_label_outline}{Outline. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_y_tick_label_space}{Space. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_y_tick_label_weight}{Choices: 600, 400. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_radial_axis_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_group_labels}{Outer labels. Enter one per line, with an angle and a label separated by a colon, such as: 50: My Label. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_radial_axis_group_labels_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_radial_axis_group_labels_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_group_labels_letter_spacing}{Letter spacing. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_group_labels_space_above}{Space above. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_radial_axis_group_labels_space_below}{Space below. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_container_alignment}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_title_weight}{Title weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_text_weight}{Text weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_legend_container_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_show_legend}{Legend mode. Choices: \code{TRUE}, \code{FALSE}. A legend will not show with a single entry. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_legend_title_mode}{Title mode. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_title}{Title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_swatch_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_swatch_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_swatch_radius}{Roundness. The radius of the corners of the swatch (in pixels). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_legend_items_padding}{Padding. Padding between legend items ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_swatch_outline}{Outline. An optional outline for the swatch in the legend. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_legend_swatch_outline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_legend_order_override}{Custom order override. Manually specify the order of legend entries (one entry per line). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_legend_icon_height}{Height. Height of icon ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_icon_color}{Color. Fallback color (icon color if not determined by template). Flourish type hint: color, defaults to \code{NULL}}

\item{chart_legend_max_width}{Max width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{facet_control_show}{Show charts grid control. Flourish type hint: boolean, defaults to \code{NULL}}

\item{facet_control_default}{Option to show on load. If you want to specify a different facet view to show on load either make a story or enter the label from the relevant button here. Case sensitive. Flourish type hint: string, defaults to \code{NULL}}

\item{facet_control_none}{Include combined option. Flourish type hint: boolean, defaults to \code{NULL}}

\item{facets_label_combined}{Combined label. Flourish type hint: string, defaults to \code{NULL}}

\item{facet_control_individual}{Include separate option. Flourish type hint: boolean, defaults to \code{NULL}}

\item{facets_label_separate}{Separate label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_facet_filter_controls_control_type}{Control type. Choices: dropdown, buttons, slider. Choose between a dropdown, buttons or a  time slider Flourish type hint: string, defaults to \code{NULL}}

\item{chart_facet_filter_controls_control_styles}{Adv. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_facet_filter_controls_control_title}{Control title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_facet_filter_controls_button_group}{Grouped. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_facet_filter_controls_button_group_width_mode}{Button group width. Choices: auto, full, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_facet_filter_controls_button_group_width_fixed}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_filter_controls_dropdown_width_mode}{Dropdown width. Choices: auto, full, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_facet_filter_controls_dropdown_width_fixed}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_filter_controls_slider_width}{Slider width. Width of the slider track in rems Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_filter_controls_slider_max_label_width}{Max label width. In rems, multiples of base font size. If larger than the longest label, only the space required to fit labels will be used. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_filter_controls_slider_step_time}{Time between steps. Measured in seconds, positive values move the slider left to right, negative values move the slider right to left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_filter_controls_slider_loop}{Loop. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_facet_filter_controls_slider_restart_pause}{Pause before loop. Measured in seconds and in addition to the regular step time displayed above. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_facet_filter_controls_sort}{Sorting. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_control_type}{Control type. Choices: dropdown, buttons, slider. Choose between a dropdown, buttons or a  time slider Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_control_styles}{Adv. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_control_title}{Control title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_button_group}{Grouped. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_button_group_width_mode}{Button group width. Choices: auto, full, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_button_group_width_fixed}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_dropdown_width_mode}{Dropdown width. Choices: auto, full, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_dropdown_width_fixed}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_width}{Slider width. Width of the slider track in rems Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_max_label_width}{Max label width. In rems, multiples of base font size. If larger than the longest label, only the space required to fit labels will be used. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_step_time}{Time between steps. Measured in seconds, positive values move the slider left to right, negative values move the slider right to left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_loop}{Loop. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_slider_restart_pause}{Pause before loop. Measured in seconds and in addition to the regular step time displayed above. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_sort}{Sorting. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{filter_control_all}{Name for "All". Leave blank to not include an "All" option. Flourish type hint: string, defaults to \code{NULL}}

\item{filter_y_max}{Update axis y max when filtering. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_controls_style_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_controls_style_font_weight}{Text weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_controls_style_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_background_selected}{Selected. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_background_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color_selected}{Selected. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_button_styles_advanced}{Button border styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_button_control_style_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_border_transparency}{Transparency. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_font_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_styles_advanced}{Dropdown border styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_style}{Border style. Choices: all, bottom. Show border on all sides, or only at the bottom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_transparency}{Transparency. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_control_style_track_color}{Track. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_font_color}{Text. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_handle_color}{Handle. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_play_color}{Play/pause. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_track_height}{Track height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_control_style_handle_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_control_style_play_button}{Play button. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_container_alignment}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_container_controls_spacing}{Spacing. The space between controls, like dropdowns and adjacent button sets, in em units. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_mode}{Choices: none, popup, panel, both. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_is_custom}{Popup content. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_show_popup_styles}{Show popup styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_popup_custom_header}{Popup custom header. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for popup use.  Advanced users can add HTML styling or add category colors to popup headers.  To hide parts of the popup when a column is empty using \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_popup_custom_main}{Popup custom main content. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for popup use.  Advanced users can add HTML for designs, styles, and images.  To hide parts of the popup when a column is empty use the syntax \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_popup_custom_accessible}{Popup custom accessible content. A text alternative to visual custom popup content that will be read out by screenreaders. This content is added as an aria-label attribute to data point elements. You can still use \code{{{column_name}}} to add values from your data, but do not use HTML. For example: \verb{\{\{Name\}\}, Population: \{\{Population\}\}, GDP: \{\{GDP\}\}}. If left empty, accessible content will fallback to the default popup content. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_popup_popup_width}{Max width. In rems, multiples of base font size. Leave blank to let the width adjust based on the contents. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_text_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_align}{Align. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_background}{Fill color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_background_opacity}{Fill opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_radius}{Radius. Corner radius of popup. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_header_type}{Header style. Choices: block, inline. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_header_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_header_text_color}{Text. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_list_type}{List type. Choices: table, ul. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_list_label_weight}{Label weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_list_separator}{List separator. Choices: :, , , \verb{<br />}. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_shadow}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_popup_pointer}{Pointer. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_panel_position}{Choices: overlay, left, right, top, bottom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_open_duration}{Open duration. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_side_fixed}{Always show. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_panel_side_default_content}{Placeholder content. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_popup_panel_is_custom}{Panel content. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_show_panel_styles}{Show panel styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_panel_custom_header}{Panel custom header. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for panel use.  Advanced users can add HTML for styling or add category colors to panel headers.  To hide parts of the panel when a column is empty using \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_panel_custom_main}{Panel custom main content. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for panel use.  Advanced users can add HTML for designs, styles, images, videos and embedded Flourish charts,  To hide parts of the panel when a column is empty using \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_panel_width_overlay}{Width (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_height_overlay}{Height (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_width_side}{Width (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_height_side}{Height (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_max_width}{Max width (px). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_max_height}{Max height (px). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_vertical_position}{Vertical position. Choices: click, center. Choosing Click position is especially useful when your visualization is really tall, this will make sure the panel content is always on top of the element youve clicked on. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_text_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_panel_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_align}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_background}{Fill color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_panel_background_opacity}{Fill opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_radius}{Radius. Corner radius of panel. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_list_type}{List type. Choices: table, ul. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_list_label_weight}{Label weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_list_separator}{List separator. Choices: :, , , \verb{<br />}. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_shadow}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{popup_color_headings}{Color popup headings. Flourish type hint: boolean, defaults to \code{NULL}}

\item{duration}{Duration. The duration, in seconds, of transitions  for example between two slides in a story. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_prefix}{Prefix. Text to place in front of a number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_suffix}{Suffix. Text to place after a number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_n_dec}{Decimal places. Decimal places. You can also enter a negative integer to round to a whole number with that many zeros. For example, -2 will round to the nearest hundred. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_negative_sign}{Styling of negative numbers. Choices: -$nk, $-nk, ($nk), $(n)k, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_strip_zeros}{Remove trailing zeros. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_strip_separator}{Hide thousands separator below 10,000. Turn off if you want four-digit numbers to include a separator, e.g. 1,234 rather than 1234. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_transform_labels}{Multiply/divide values. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_transform}{Choices: multiply, divide, exponentiate. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_multiply_divide_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_exponentiate_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_body_font}{Main font. This font will apply to the whole graphic by default. You can optionally change the font for the title, subtitle, footer, etc in the Header and Footer settings panels. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_font_color}{Text color. This color will apply to the whole graphic by default, You can optionally change the color for individual text elements, in other settings panels. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_color_enabled}{Color. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_image_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_background_image_size}{Size. Choices: cover, contain, auto, 100\% 100\%. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_background_image_position}{Position. Choices: top left, top center, top right, center left, center center, center right, bottom left, bottom center, bottom right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width_target}{Maximum width. Choices: none, wrapper, primary. Apply a maximum width to just the main graphic or everything (main graphic plus header, footer, etc). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width}{Maximum width. Leave blank to stretch to container width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_max_width_align}{Align. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_layout_order}{Layout order. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections}{Space between sections. Choices: 0.5, 1, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_enabled}{Show borders around visualisation. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_border_top_width}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_top_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_top_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_right_width}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_right_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_right_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_bottom_width}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_bottom_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_bottom_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_left_width}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_left_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_left_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_read_direction}{Read direction. Choices: ltr, rtl. This will change the reading direction of the main text elements on the page. Its not possible to adjust this on all elements, such as axes.  Note that when direction is set to right to left any alignment icons will be reversed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_align}{Alignment. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_title_font}{Title Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_title_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_title_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_subtitle_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_subtitle_font}{Subtitle Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_subtitle_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_subtitle_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_header_text_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_text_size}{Size. Choices: 1.2, 1.4, 1.6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_text_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_space}{Space. Space between border and header text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_enabled}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_header_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_align}{Align. Choices: inside, outside. Align logo inside either the header or the main visualization container. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_inside}{Position. Choices: top, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_outside}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_align}{Alignment. Choices: left, center, right, justify. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_styling}{Advanced footer styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_font}{Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_footer_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_multiple_sources}{Multiple sources. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_source_name_2}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_2}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name_3}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_3}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_label}{Source label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_note}{Note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_footer_note_secondary}{Note (secondary). The secondary note is placed below the source and primary note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_advanced_note_styling}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_timestamp_format}{Time/date stamp formatting. Choices: \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M, \%H:\%M \%p, \%H:\%M:\%S, \%d \%b \%Y, \%H:\%M, \%d \%b \%Y, \%H:\%M, \%d/\%m/\%Y, \%H:\%M, \%m/\%d/\%Y, \%H:\%M, \%d-\%m-\%Y, \%H:\%M, \%m-\%d-\%Y, \%d \%b \%Y, \%m/\%d/\%Y, \%d/\%m/\%Y, \%m-\%d-\%Y, \%d-\%m-\%Y, \%Y-\%m-\%d, \%Y-\%m-\%d \%H:\%M:\%S. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src_light}{Image (light version). If provided this version will be used whenever the background color is dark. Flourish type hint: hidden, defaults to \code{NULL}}

\item{chart_layout_footer_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_margin}{Margin. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_order}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_align_vertical}{V. align. Choices: flex-start, center, flex-end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_space}{Space. Space between border and footer text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_screenreader_text_primary}{Screenreader description. A text alternative to the visual content that will only be visible to screenreaders, e.g. The line chart shows China consistently higher than the other countries since 1990.  Do not replicate your title, since that will also be read by screenreaders. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_layout_screenreader_label}{Screenreader label. A short text label given to the main Flourish embed wrapper to provide an accessible name that is only visible to screenreaders Added in the form of an "aria-label". Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_screenreader_hide_primary}{Screenreader mode for main visual container. Choices: \code{TRUE}, \code{FALSE}. Whether the main visual container is visible to screenreaders (Text in the header and footer are always available to screenreaders.). Flourish type hint: boolean, defaults to \code{NULL}}
}
\value{
A Flourish chart
}
\description{
Set radar chart chart details
}
\examples{
try(
  flourish(chart_type = "radar", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_radar_data(gapminder) |>
    set_radar_details()
)
}
