% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_survey_details.R
\name{set_survey_details}
\alias{set_survey_details}
\title{Set survey chart details}
\usage{
set_survey_details(
  .,
  dot_fill = NULL,
  dot_opacity = NULL,
  dot_opacity_unhighlighted = NULL,
  chart_color_scale_categorical_palette = NULL,
  chart_color_scale_categorical_extend = NULL,
  chart_color_scale_categorical_custom_palette = NULL,
  chart_color_scale_numeric_type = NULL,
  chart_color_scale_binning = NULL,
  chart_color_scale_bin_mode = NULL,
  chart_color_scale_bin_count = NULL,
  chart_color_scale_bin_thresholds = NULL,
  chart_color_scale_sequential_palette = NULL,
  chart_color_scale_sequential_reverse = NULL,
  chart_color_scale_sequential_custom_min = NULL,
  chart_color_scale_sequential_custom_max = NULL,
  chart_color_scale_sequential_color_space = NULL,
  chart_color_scale_sequential_custom_domain = NULL,
  chart_color_scale_sequential_domain_min = NULL,
  chart_color_scale_sequential_domain_max = NULL,
  chart_color_scale_diverging_palette = NULL,
  chart_color_scale_diverging_reverse = NULL,
  chart_color_scale_diverging_custom_min = NULL,
  chart_color_scale_diverging_custom_mid = NULL,
  chart_color_scale_diverging_custom_max = NULL,
  chart_color_scale_diverging_color_space = NULL,
  chart_color_scale_diverging_custom_domain = NULL,
  chart_color_scale_diverging_domain_min = NULL,
  chart_color_scale_diverging_domain_mid = NULL,
  chart_color_scale_diverging_domain_max = NULL,
  dot_shape = NULL,
  dot_sides = NULL,
  outline_width = NULL,
  outline_color = NULL,
  dot_label = NULL,
  dot_label_color = NULL,
  dot_label_bold = NULL,
  dot_label_min = NULL,
  dot_label_auto_size = NULL,
  dot_label_fixed = NULL,
  dot_label_scale = NULL,
  dot_label_size_value = NULL,
  all_group_label = NULL,
  include_grouped_questions_on_shade_menu = NULL,
  include_grouped_questions_on_compare_menu = NULL,
  order_groups_by_size = NULL,
  order_groups_by_size_mode = NULL,
  group_colors = NULL,
  exclusions = NULL,
  exclude_blanks = NULL,
  bundle_small_groups = NULL,
  small_groups_threshold = NULL,
  small_groups_other_label = NULL,
  label_mode = NULL,
  label_tight_fit = NULL,
  label_color = NULL,
  label_min_font_size = NULL,
  label_max_font_size = NULL,
  label_group_padding_lines = NULL,
  label_shadow_size = NULL,
  label_shadow_color = NULL,
  label_group_count = NULL,
  label_group_percent = NULL,
  label_group_percent_decimals = NULL,
  always_show_x_label = NULL,
  show_xy_axis_labels_when_comparing = NULL,
  compare_y_axis_width = NULL,
  compare_x_axis_height = NULL,
  chart_height_mode = NULL,
  chart_aspect_ratio_wide = NULL,
  chart_aspect_ratio_narrow = NULL,
  map_opacity_map_mode = NULL,
  map_opacity_non_map_mode = NULL,
  map_max_group_size = NULL,
  map_dimensions = NULL,
  map_aspect_ratio = NULL,
  map_region = NULL,
  map_region_url = NULL,
  map_region_bounds_x1 = NULL,
  map_region_bounds_x2 = NULL,
  map_region_bounds_y2 = NULL,
  map_region_bounds_y1 = NULL,
  chart_popup_mode = NULL,
  chart_popup_popup_is_custom = NULL,
  chart_popup_show_popup_styles = NULL,
  chart_popup_popup_custom_header = NULL,
  chart_popup_popup_custom_main = NULL,
  chart_popup_popup_custom_accessible = NULL,
  chart_popup_popup_width = NULL,
  chart_popup_popup_font_size = NULL,
  chart_popup_popup_text_color = NULL,
  chart_popup_popup_align = NULL,
  chart_popup_popup_background = NULL,
  chart_popup_popup_background_opacity = NULL,
  chart_popup_popup_padding = NULL,
  chart_popup_popup_radius = NULL,
  chart_popup_popup_header_type = NULL,
  chart_popup_popup_header_background = NULL,
  chart_popup_popup_header_text_color = NULL,
  chart_popup_popup_list_type = NULL,
  chart_popup_popup_list_label_weight = NULL,
  chart_popup_popup_list_separator = NULL,
  chart_popup_popup_shadow = NULL,
  chart_popup_popup_pointer = NULL,
  chart_popup_panel_position = NULL,
  chart_popup_panel_open_duration = NULL,
  chart_popup_panel_side_fixed = NULL,
  chart_popup_panel_side_default_content = NULL,
  chart_popup_panel_is_custom = NULL,
  chart_popup_show_panel_styles = NULL,
  chart_popup_panel_custom_header = NULL,
  chart_popup_panel_custom_main = NULL,
  chart_popup_panel_width_overlay = NULL,
  chart_popup_panel_height_overlay = NULL,
  chart_popup_panel_width_side = NULL,
  chart_popup_panel_height_side = NULL,
  chart_popup_panel_max_width = NULL,
  chart_popup_panel_max_height = NULL,
  chart_popup_panel_vertical_position = NULL,
  chart_popup_panel_text_color = NULL,
  chart_popup_panel_font_size = NULL,
  chart_popup_panel_align = NULL,
  chart_popup_panel_background = NULL,
  chart_popup_panel_background_opacity = NULL,
  chart_popup_panel_padding = NULL,
  chart_popup_panel_radius = NULL,
  chart_popup_panel_list_type = NULL,
  chart_popup_panel_list_label_weight = NULL,
  chart_popup_panel_list_separator = NULL,
  chart_popup_panel_shadow = NULL,
  panel_background_color = NULL,
  panel_background_opacity = NULL,
  panel_text_color = NULL,
  panel_text_size = NULL,
  panel_icon_color = NULL,
  panel_width = NULL,
  panel_cog_color = NULL,
  panel_overlay = NULL,
  panel_hide_cog = NULL,
  select_group_show = NULL,
  select_color_show = NULL,
  select_size_show = NULL,
  select_compare_show = NULL,
  histogram_show = NULL,
  select_group_text = NULL,
  select_color_text = NULL,
  select_size_text = NULL,
  select_compare_text = NULL,
  select_filter_text = NULL,
  select_map_text = NULL,
  chart_filter_controls_control_type = NULL,
  chart_filter_controls_control_styles = NULL,
  chart_filter_controls_control_title = NULL,
  chart_filter_controls_button_group = NULL,
  chart_filter_controls_button_group_width_mode = NULL,
  chart_filter_controls_button_group_width_fixed = NULL,
  chart_filter_controls_dropdown_width_mode = NULL,
  chart_filter_controls_dropdown_width_fixed = NULL,
  chart_filter_controls_slider_width = NULL,
  chart_filter_controls_slider_max_label_width = NULL,
  chart_filter_controls_slider_step_time = NULL,
  chart_filter_controls_slider_loop = NULL,
  chart_filter_controls_slider_restart_pause = NULL,
  chart_filter_controls_sort = NULL,
  column_filter_show_all = NULL,
  column_filter_show_all_label = NULL,
  chart_controls_style_font_size = NULL,
  chart_controls_style_font_weight = NULL,
  chart_controls_style_height = NULL,
  chart_button_control_style_background = NULL,
  chart_button_control_style_background_selected = NULL,
  chart_button_control_style_background_hover = NULL,
  chart_button_control_style_font_color = NULL,
  chart_button_control_style_font_color_selected = NULL,
  chart_button_control_style_font_color_hover = NULL,
  chart_button_control_style_button_styles_advanced = NULL,
  chart_button_control_style_border_width = NULL,
  chart_button_control_style_border_color = NULL,
  chart_button_control_style_border_transparency = NULL,
  chart_button_control_style_border_radius = NULL,
  chart_slider_control_style_track_color = NULL,
  chart_slider_control_style_font_color = NULL,
  chart_slider_control_style_handle_color = NULL,
  chart_slider_control_style_play_color = NULL,
  chart_slider_control_style_track_height = NULL,
  chart_slider_control_style_handle_height = NULL,
  chart_slider_control_style_play_button = NULL,
  chart_dropdown_control_style_background = NULL,
  chart_dropdown_control_style_font_color = NULL,
  chart_dropdown_control_style_border_styles_advanced = NULL,
  chart_dropdown_control_style_border_style = NULL,
  chart_dropdown_control_style_border_width = NULL,
  chart_dropdown_control_style_border_color = NULL,
  chart_dropdown_control_style_border_transparency = NULL,
  chart_dropdown_control_style_border_radius = NULL,
  chart_categorical_legend_show_legend = NULL,
  chart_categorical_legend_title_mode = NULL,
  chart_categorical_legend_title = NULL,
  chart_categorical_legend_swatch_width = NULL,
  chart_categorical_legend_swatch_height = NULL,
  chart_categorical_legend_swatch_radius = NULL,
  chart_categorical_legend_legend_items_padding = NULL,
  chart_categorical_legend_swatch_outline = NULL,
  chart_categorical_legend_swatch_outline_color = NULL,
  chart_categorical_legend_order_override = NULL,
  chart_categorical_legend_icon_height = NULL,
  chart_categorical_legend_icon_color = NULL,
  chart_categorical_legend_max_width = NULL,
  chart_categorical_legend_orientation = NULL,
  legend_position = NULL,
  legend_hide_space = NULL,
  chart_legend_container_alignment = NULL,
  chart_legend_container_orientation = NULL,
  chart_legend_container_title_weight = NULL,
  chart_legend_container_text_weight = NULL,
  chart_legend_container_text_color = NULL,
  chart_legend_container_text_size = NULL,
  zoom_on_double_click = NULL,
  zoom_on_mousewheel = NULL,
  transition_duration = NULL,
  transition_stagger = NULL,
  chart_layout_body_font = NULL,
  chart_layout_font_color = NULL,
  chart_layout_background_color_enabled = NULL,
  chart_layout_background_image_enabled = NULL,
  chart_layout_background_color = NULL,
  chart_layout_background_image_src = NULL,
  chart_layout_background_image_size = NULL,
  chart_layout_background_image_position = NULL,
  chart_layout_max_width_target = NULL,
  chart_layout_max_width = NULL,
  chart_layout_max_width_align = NULL,
  chart_layout_layout_order = NULL,
  chart_layout_space_between_sections = NULL,
  chart_layout_space_between_sections_custom = NULL,
  chart_layout_margin_top = NULL,
  chart_layout_margin_right = NULL,
  chart_layout_margin_bottom = NULL,
  chart_layout_margin_left = NULL,
  chart_layout_border_enabled = NULL,
  chart_layout_border_top_width = NULL,
  chart_layout_border_top_style = NULL,
  chart_layout_border_top_color = NULL,
  chart_layout_border_right_width = NULL,
  chart_layout_border_right_style = NULL,
  chart_layout_border_right_color = NULL,
  chart_layout_border_bottom_width = NULL,
  chart_layout_border_bottom_style = NULL,
  chart_layout_border_bottom_color = NULL,
  chart_layout_border_left_width = NULL,
  chart_layout_border_left_style = NULL,
  chart_layout_border_left_color = NULL,
  chart_layout_read_direction = NULL,
  chart_layout_font_size_mobile_small = NULL,
  chart_layout_font_size_mobile_big = NULL,
  chart_layout_font_size_tablet = NULL,
  chart_layout_font_size_desktop = NULL,
  chart_layout_font_size_big_screen = NULL,
  chart_layout_breakpoint_mobile_small = NULL,
  chart_layout_breakpoint_mobile_big = NULL,
  chart_layout_breakpoint_tablet = NULL,
  chart_layout_breakpoint_desktop = NULL,
  chart_layout_breakpoint_big_screen = NULL,
  chart_layout_header_align = NULL,
  chart_layout_title = NULL,
  chart_layout_title_styling = NULL,
  chart_layout_title_font = NULL,
  chart_layout_title_size = NULL,
  chart_layout_title_size_custom = NULL,
  chart_layout_title_weight = NULL,
  chart_layout_title_color = NULL,
  chart_layout_title_line_height = NULL,
  chart_layout_title_space_above = NULL,
  chart_layout_title_space_above_custom = NULL,
  chart_layout_subtitle = NULL,
  chart_layout_subtitle_styling = NULL,
  chart_layout_subtitle_font = NULL,
  chart_layout_subtitle_size = NULL,
  chart_layout_subtitle_size_custom = NULL,
  chart_layout_subtitle_weight = NULL,
  chart_layout_subtitle_color = NULL,
  chart_layout_subtitle_line_height = NULL,
  chart_layout_subtitle_space_above = NULL,
  chart_layout_subtitle_space_above_custom = NULL,
  chart_layout_header_text = NULL,
  chart_layout_header_text_styling = NULL,
  chart_layout_header_text_size = NULL,
  chart_layout_header_text_size_custom = NULL,
  chart_layout_header_text_weight = NULL,
  chart_layout_header_text_color = NULL,
  chart_layout_header_text_line_height = NULL,
  chart_layout_header_text_space_above = NULL,
  chart_layout_header_text_space_above_custom = NULL,
  chart_layout_header_border = NULL,
  chart_layout_header_border_width = NULL,
  chart_layout_header_border_color = NULL,
  chart_layout_header_border_style = NULL,
  chart_layout_header_border_space = NULL,
  chart_layout_header_logo_enabled = NULL,
  chart_layout_header_logo_src = NULL,
  chart_layout_header_logo_alt = NULL,
  chart_layout_header_logo_link_url = NULL,
  chart_layout_header_logo_height = NULL,
  chart_layout_header_logo_align = NULL,
  chart_layout_header_logo_position_inside = NULL,
  chart_layout_header_logo_position_outside = NULL,
  chart_layout_header_logo_margin_top = NULL,
  chart_layout_header_logo_margin_right = NULL,
  chart_layout_header_logo_margin_bottom = NULL,
  chart_layout_header_logo_margin_left = NULL,
  chart_layout_footer_align = NULL,
  chart_layout_footer_text_size = NULL,
  chart_layout_footer_text_color = NULL,
  chart_layout_footer_styling = NULL,
  chart_layout_footer_font = NULL,
  chart_layout_footer_text_weight = NULL,
  chart_layout_source_name = NULL,
  chart_layout_source_url = NULL,
  chart_layout_multiple_sources = NULL,
  chart_layout_source_name_2 = NULL,
  chart_layout_source_url_2 = NULL,
  chart_layout_source_name_3 = NULL,
  chart_layout_source_url_3 = NULL,
  chart_layout_source_label = NULL,
  chart_layout_footer_note = NULL,
  chart_layout_footer_note_secondary = NULL,
  chart_layout_advanced_note_styling = NULL,
  chart_layout_footer_timestamp_format = NULL,
  chart_layout_footer_logo_enabled = NULL,
  chart_layout_footer_logo_src = NULL,
  chart_layout_footer_logo_src_light = NULL,
  chart_layout_footer_logo_alt = NULL,
  chart_layout_footer_logo_link_url = NULL,
  chart_layout_footer_logo_height = NULL,
  chart_layout_footer_logo_margin = NULL,
  chart_layout_footer_logo_order = NULL,
  chart_layout_footer_align_vertical = NULL,
  chart_layout_footer_border = NULL,
  chart_layout_footer_border_width = NULL,
  chart_layout_footer_border_color = NULL,
  chart_layout_footer_border_style = NULL,
  chart_layout_footer_border_space = NULL,
  chart_layout_screenreader_text_primary = NULL,
  chart_layout_screenreader_label = NULL,
  chart_layout_screenreader_hide_primary = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{dot_fill}{Default color. If left blank, the first color in the color palette will be used. Flourish type hint: color, defaults to \code{NULL}}

\item{dot_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{dot_opacity_unhighlighted}{Opacity (unhighlighted). The opacity of the unhighlighted dots when one of the dots is highlighted. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_categorical_palette}{Color when shading. The colors to use when shading by the answer to a questions. Flourish type hint: colors, defaults to \code{NULL}}

\item{chart_color_scale_categorical_extend}{Extend. Automatically generate additional colours when needed to avoid the palette colours being used more than once. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_categorical_custom_palette}{Custom overrides. Enter the name of the category that you wish to set the color, followed by a colon and the desired color value.  Colors can be set using Hex, RGB, color names or RGBA, if you want to set the opacity Multiple colors can be set using multiple lines. For example:  Party 1: red Party 2: #4455AA Party 3: rgb(30,168,26) Note that overrides from the "Color overrides" sheet take priority over the overrides in this setting. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_color_scale_numeric_type}{Scale type. Choices: sequential, diverging. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_binning}{Choices: \code{FALSE}, \code{TRUE}. In linear mode, the color scale will run as a smooth gradient between 2 colors. In binned mode, the gradient will be divided in smaller blocks. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_bin_mode}{Binning mode. Choices: fixed, quantile, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_bin_count}{Number of bins. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_bin_thresholds}{Custom thresholds. Enter your desired thresholds, separating them with a ";". For instance, "5;10;15".  Bins form as follows:  From the datas minimum value (domain minimum) up to the first threshold.  Between consecutive thresholds.  From the last threshold to the datas maximum value (domain maximum).   For "5;10;15", youll get four bins based on your datas range. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_mid}{Midpoint color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_mid}{Mid. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{dot_shape}{Shape. Choices: circle, square, polygon. Flourish type hint: string, defaults to \code{NULL}}

\item{dot_sides}{Sides. Flourish type hint: number, defaults to \code{NULL}}

\item{outline_width}{Stroke width. In pixels. The width of the outline stroke. Flourish type hint: number, defaults to \code{NULL}}

\item{outline_color}{Stroke color. Outline color for your dots/squares/polygons. If not set, the layout background color will be used. Flourish type hint: color, defaults to \code{NULL}}

\item{dot_label}{Default label or emoji. A label or emoji to show on each dot. Overriden by any label specified using the Label column setting. Flourish type hint: string, defaults to \code{NULL}}

\item{dot_label_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{dot_label_bold}{Bold. Flourish type hint: boolean, defaults to \code{NULL}}

\item{dot_label_min}{Min size to display. The minimum text size in pixels for dot labels. Below this level they will not render. Flourish type hint: number, defaults to \code{NULL}}

\item{dot_label_auto_size}{Sizing mode. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{dot_label_fixed}{Size in \%. Specifies font size across the graphic as a percentage of dot height. Flourish type hint: number, defaults to \code{NULL}}

\item{dot_label_scale}{Scale. Scales the labels up or down from their default size. Flourish type hint: number, defaults to \code{NULL}}

\item{dot_label_size_value}{Show numerical value when sizing dots. Flourish type hint: boolean, defaults to \code{NULL}}

\item{all_group_label}{Name for "all" group. Flourish type hint: string, defaults to \code{NULL}}

\item{include_grouped_questions_on_shade_menu}{Include grouped questions on shading menu. Flourish type hint: boolean, defaults to \code{NULL}}

\item{include_grouped_questions_on_compare_menu}{Include grouped questions on compare menu. Flourish type hint: boolean, defaults to \code{NULL}}

\item{order_groups_by_size}{Order groups by size. Flourish type hint: boolean, defaults to \code{NULL}}

\item{order_groups_by_size_mode}{Choices: ascending, descending. Flourish type hint: string, defaults to \code{NULL}}

\item{group_colors}{Arrange groups by color. If selected, dots of the same color are grouped together; if not the dots are placed in spreadsheet order. Flourish type hint: boolean, defaults to \code{NULL}}

\item{exclusions}{Exclude groups with the following names. Case-sensitive comma-separated list. Flourish type hint: string, defaults to \code{NULL}}

\item{exclude_blanks}{Exclude groups with blank names. Flourish type hint: boolean, defaults to \code{NULL}}

\item{bundle_small_groups}{Bundle small groups. Flourish type hint: boolean, defaults to \code{NULL}}

\item{small_groups_threshold}{Min size. Flourish type hint: number, defaults to \code{NULL}}

\item{small_groups_other_label}{"Other" group label. Flourish type hint: string, defaults to \code{NULL}}

\item{label_mode}{Label position. Choices: atop, around, above. Flourish type hint: string, defaults to \code{NULL}}

\item{label_tight_fit}{Position close to group. Flourish type hint: boolean, defaults to \code{NULL}}

\item{label_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{label_min_font_size}{Min font size. In pixels; the actual font size in any view is set automatically between the min and max. Flourish type hint: number, defaults to \code{NULL}}

\item{label_max_font_size}{Max font size. In pixels; the actual font size in any view is set automatically between the min and max. Flourish type hint: number, defaults to \code{NULL}}

\item{label_group_padding_lines}{Space for labels. The amount of space to leave above the groups, specified in lines of text. Flourish type hint: number, defaults to \code{NULL}}

\item{label_shadow_size}{Size. As percent of font size. Flourish type hint: number, defaults to \code{NULL}}

\item{label_shadow_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{label_group_count}{Show number of dots in group. Adds the number in brackets after the main label; ignored in chart, map and compare modes. Flourish type hint: boolean, defaults to \code{NULL}}

\item{label_group_percent}{Show group size as \% of visible dots. Adds group size as a percent of visible dots, in brackets after the main label.  This setting is ignored in chart, map and compare modes and when showing a grouped question. Flourish type hint: boolean, defaults to \code{NULL}}

\item{label_group_percent_decimals}{Decimal places on percentage labels. Flourish type hint: number, defaults to \code{NULL}}

\item{always_show_x_label}{Show X label when not comparing. Show the groups label at the bottom of chart when not comparing two metrics. Flourish type hint: boolean, defaults to \code{NULL}}

\item{show_xy_axis_labels_when_comparing}{Show X/Y labels when comparing. Show groups labels on both axes when comparing two metrics. Flourish type hint: boolean, defaults to \code{NULL}}

\item{compare_y_axis_width}{Y axis width. Flourish type hint: number, defaults to \code{NULL}}

\item{compare_x_axis_height}{X axis height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_height_mode}{Height mode. Choices: auto, aspect_ratio. How the graphic height will be set when embedded with the responsive script embed without a specified height. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_aspect_ratio_wide}{Desktop. The chart aspect ratio on screens 640 pixels wide and higher (1 for square - a smaller value means a bigger height). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_aspect_ratio_narrow}{Mobile. The chart aspect ratio on screens narrower than 640 pixels (1 for square - a smaller value means a bigger height). Flourish type hint: number, defaults to \code{NULL}}

\item{map_opacity_map_mode}{Map opacity (in map mode). Flourish type hint: number, defaults to \code{NULL}}

\item{map_opacity_non_map_mode}{Map opacity (in other modes). Flourish type hint: number, defaults to \code{NULL}}

\item{map_max_group_size}{Max group size in map view. In relative units. Flourish type hint: number, defaults to \code{NULL}}

\item{map_dimensions}{Map dimensions. Choices: auto, aspect_ratio. Flourish type hint: string, defaults to \code{NULL}}

\item{map_aspect_ratio}{Map aspect ratio. Aspect ratio of the map. 1 will make the map square. Flourish type hint: number, defaults to \code{NULL}}

\item{map_region}{Region. Choices: world, usa, europe, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{map_region_url}{Url to equirectangular projection of map. Flourish type hint: url, defaults to \code{NULL}}

\item{map_region_bounds_x1}{Minimum longitude. Flourish type hint: number, defaults to \code{NULL}}

\item{map_region_bounds_x2}{Maximum longitude. Flourish type hint: number, defaults to \code{NULL}}

\item{map_region_bounds_y2}{Minimum latitude. Flourish type hint: number, defaults to \code{NULL}}

\item{map_region_bounds_y1}{Maximum latitude. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_mode}{Choices: none, popup, panel, both. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_is_custom}{Popup content. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_show_popup_styles}{Show popup styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_popup_custom_header}{Popup custom header. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for popup use.  Advanced users can add HTML styling or add category colors to popup headers.  To hide parts of the popup when a column is empty using \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_popup_custom_main}{Popup custom main content. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for popup use.  Advanced users can add HTML for designs, styles, and images.  To hide parts of the popup when a column is empty use the syntax \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_popup_custom_accessible}{Popup custom accessible content. A text alternative to visual custom popup content that will be read out by screenreaders. This content is added as an aria-label attribute to data point elements. You can still use \code{{{column_name}}} to add values from your data, but do not use HTML. For example: \verb{\{\{Name\}\}, Population: \{\{Population\}\}, GDP: \{\{GDP\}\}}. If left empty, accessible content will fallback to the default popup content. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_popup_popup_width}{Max width. In rems, multiples of base font size. Leave blank to let the width adjust based on the contents. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_text_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_align}{Align. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_background}{Fill color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_background_opacity}{Fill opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_radius}{Radius. Corner radius of popup. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_header_type}{Header style. Choices: block, inline. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_header_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_header_text_color}{Text. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_list_type}{List type. Choices: table, ul. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_list_label_weight}{Label weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_list_separator}{List separator. Choices: :, , , \verb{<br />}. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_shadow}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_popup_pointer}{Pointer. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_panel_position}{Choices: overlay, left, right, top, bottom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_open_duration}{Open duration. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_side_fixed}{Always show. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_panel_side_default_content}{Placeholder content. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_popup_panel_is_custom}{Panel content. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_show_panel_styles}{Show panel styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_panel_custom_header}{Panel custom header. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for panel use.  Advanced users can add HTML for styling or add category colors to panel headers.  To hide parts of the panel when a column is empty using \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_panel_custom_main}{Panel custom main content. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for panel use.  Advanced users can add HTML for designs, styles, images, videos and embedded Flourish charts,  To hide parts of the panel when a column is empty using \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_panel_width_overlay}{Width (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_height_overlay}{Height (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_width_side}{Width (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_height_side}{Height (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_max_width}{Max width (px). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_max_height}{Max height (px). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_vertical_position}{Vertical position. Choices: click, center. Choosing Click position is especially useful when your visualization is really tall, this will make sure the panel content is always on top of the element youve clicked on. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_text_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_panel_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_align}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_background}{Fill color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_panel_background_opacity}{Fill opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_radius}{Radius. Corner radius of panel. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_list_type}{List type. Choices: table, ul. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_list_label_weight}{Label weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_list_separator}{List separator. Choices: :, , , \verb{<br />}. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_shadow}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{panel_background_color}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{panel_background_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{panel_text_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{panel_text_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{panel_icon_color}{Icon color. Choices: light, dark. Flourish type hint: string, defaults to \code{NULL}}

\item{panel_width}{Panel width. Width of the controls panels on wide screens, ignored on mobile. Flourish type hint: number, defaults to \code{NULL}}

\item{panel_cog_color}{Cog icon color. Flourish type hint: color, defaults to \code{NULL}}

\item{panel_overlay}{Overlay panel over graphic. Flourish type hint: boolean, defaults to \code{NULL}}

\item{panel_hide_cog}{Disable controls toggle in published graphic. When checked, the controls icon for showing/ hiding the controls panel is visible only in the visualization and story editors  not in the published visualization or story.  When making a story, you can still lock the panel to be visible or hidden by clicking the cog icon in the story editor. Flourish type hint: boolean, defaults to \code{NULL}}

\item{select_group_show}{Group menu. Populated by all category columns. Flourish type hint: boolean, defaults to \code{NULL}}

\item{select_color_show}{Color menu. Populated by all category columns. Flourish type hint: boolean, defaults to \code{NULL}}

\item{select_size_show}{Size menu. Populated by all continuous number columns. Flourish type hint: boolean, defaults to \code{NULL}}

\item{select_compare_show}{Compare menu. Populated by all category columns; appears only when Group by is selected. Flourish type hint: boolean, defaults to \code{NULL}}

\item{histogram_show}{Bar mode. Flourish type hint: boolean, defaults to \code{NULL}}

\item{select_group_text}{Group by. Flourish type hint: string, defaults to \code{NULL}}

\item{select_color_text}{Shade by. Flourish type hint: string, defaults to \code{NULL}}

\item{select_size_text}{Size by. Flourish type hint: string, defaults to \code{NULL}}

\item{select_compare_text}{Compare. Flourish type hint: string, defaults to \code{NULL}}

\item{select_filter_text}{Filter. Flourish type hint: string, defaults to \code{NULL}}

\item{select_map_text}{Group location. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_control_type}{Control type. Choices: dropdown, buttons, slider. Choose between a dropdown, buttons or a  time slider Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_control_styles}{Adv. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_control_title}{Control title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_button_group}{Grouped. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_button_group_width_mode}{Button group width. Choices: auto, full, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_button_group_width_fixed}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_dropdown_width_mode}{Dropdown width. Choices: auto, full, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_dropdown_width_fixed}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_width}{Slider width. Width of the slider track in rems Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_max_label_width}{Max label width. In rems, multiples of base font size. If larger than the longest label, only the space required to fit labels will be used. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_step_time}{Time between steps. Measured in seconds, positive values move the slider left to right, negative values move the slider right to left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_loop}{Loop. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_slider_restart_pause}{Pause before loop. Measured in seconds and in addition to the regular step time displayed above. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_sort}{Sorting. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{column_filter_show_all}{Include "All" in slider/menu. Flourish type hint: boolean, defaults to \code{NULL}}

\item{column_filter_show_all_label}{Label for All. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_controls_style_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_controls_style_font_weight}{Text weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_controls_style_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_background_selected}{Selected. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_background_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color_selected}{Selected. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_button_styles_advanced}{Button border styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_button_control_style_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_border_transparency}{Transparency. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_control_style_track_color}{Track. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_font_color}{Text. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_handle_color}{Handle. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_play_color}{Play/pause. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_track_height}{Track height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_control_style_handle_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_control_style_play_button}{Play button. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_dropdown_control_style_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_font_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_styles_advanced}{Dropdown border styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_style}{Border style. Choices: all, bottom. Show border on all sides, or only at the bottom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_transparency}{Transparency. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_show_legend}{Legend mode. Choices: \code{TRUE}, \code{FALSE}. A legend will not show with a single entry. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_categorical_legend_title_mode}{Title mode. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_categorical_legend_title}{Title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_radius}{Roundness. The radius of the corners of the swatch (in pixels). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_legend_items_padding}{Padding. Padding between legend items ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_outline}{Outline. An optional outline for the swatch in the legend. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_outline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_categorical_legend_order_override}{Custom order override. Manually specify the order of legend entries (one entry per line). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_categorical_legend_icon_height}{Height. Height of icon ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_icon_color}{Color. Fallback color (icon color if not determined by template). Flourish type hint: color, defaults to \code{NULL}}

\item{chart_categorical_legend_max_width}{Max width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{legend_position}{Position. Choices: above, below. Flourish type hint: string, defaults to \code{NULL}}

\item{legend_hide_space}{Remove space when empty. Turn off to avoid the legend taking up space when empty.  Leave on to avoid the graphic jumping up and down a little when the legend is revealed or hidden. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_legend_container_alignment}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_title_weight}{Title weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_text_weight}{Text weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_legend_container_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{zoom_on_double_click}{Zoom on double click. Flourish type hint: boolean, defaults to \code{NULL}}

\item{zoom_on_mousewheel}{Zoom on pinch/mousewheel. Flourish type hint: boolean, defaults to \code{NULL}}

\item{transition_duration}{Duration. Time in milliseconds of each dots animation. Flourish type hint: number, defaults to \code{NULL}}

\item{transition_stagger}{Stagger. The spread of time over which the dots start animating. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_body_font}{Main font. This font will apply to the whole graphic by default. You can optionally change the font for the title, subtitle, footer, etc in the Header and Footer settings panels. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_font_color}{Text color. This color will apply to the whole graphic by default, You can optionally change the color for individual text elements, in other settings panels. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_color_enabled}{Color. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_image_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_background_image_size}{Size. Choices: cover, contain, auto, 100\% 100\%. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_background_image_position}{Position. Choices: top left, top center, top right, center left, center center, center right, bottom left, bottom center, bottom right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width_target}{Maximum width. Choices: none, wrapper, primary. Apply a maximum width to just the main graphic or everything (main graphic plus header, footer, etc). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width}{Maximum width. Leave blank to stretch to container width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_max_width_align}{Align. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_layout_order}{Layout order. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections}{Space between sections. Choices: 0.5, 1, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_enabled}{Show borders around visualisation. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_border_top_width}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_top_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_top_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_right_width}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_right_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_right_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_bottom_width}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_bottom_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_bottom_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_left_width}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_left_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_left_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_read_direction}{Read direction. Choices: ltr, rtl. This will change the reading direction of the main text elements on the page. Its not possible to adjust this on all elements, such as axes.  Note that when direction is set to right to left any alignment icons will be reversed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_align}{Alignment. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_title_font}{Title Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_title_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_title_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_subtitle_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_subtitle_font}{Subtitle Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_subtitle_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_subtitle_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_header_text_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_text_size}{Size. Choices: 1.2, 1.4, 1.6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_text_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_space}{Space. Space between border and header text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_enabled}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_header_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_align}{Align. Choices: inside, outside. Align logo inside either the header or the main visualization container. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_inside}{Position. Choices: top, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_outside}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_align}{Alignment. Choices: left, center, right, justify. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_styling}{Advanced footer styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_font}{Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_footer_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_multiple_sources}{Multiple sources. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_source_name_2}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_2}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name_3}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_3}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_label}{Source label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_note}{Note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_footer_note_secondary}{Note (secondary). The secondary note is placed below the source and primary note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_advanced_note_styling}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_timestamp_format}{Time/date stamp formatting. Choices: \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M, \%H:\%M \%p, \%H:\%M:\%S, \%d \%b \%Y, \%H:\%M, \%d \%b \%Y, \%H:\%M, \%d/\%m/\%Y, \%H:\%M, \%m/\%d/\%Y, \%H:\%M, \%d-\%m-\%Y, \%H:\%M, \%m-\%d-\%Y, \%d \%b \%Y, \%m/\%d/\%Y, \%d/\%m/\%Y, \%m-\%d-\%Y, \%d-\%m-\%Y, \%Y-\%m-\%d, \%Y-\%m-\%d \%H:\%M:\%S. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src_light}{Image (light version). If provided this version will be used whenever the background color is dark. Flourish type hint: hidden, defaults to \code{NULL}}

\item{chart_layout_footer_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_margin}{Margin. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_order}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_align_vertical}{V. align. Choices: flex-start, center, flex-end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_space}{Space. Space between border and footer text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_screenreader_text_primary}{Screenreader description. A text alternative to the visual content that will only be visible to screenreaders, e.g. The line chart shows China consistently higher than the other countries since 1990.  Do not replicate your title, since that will also be read by screenreaders. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_layout_screenreader_label}{Screenreader label. A short text label given to the main Flourish embed wrapper to provide an accessible name that is only visible to screenreaders Added in the form of an "aria-label". Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_screenreader_hide_primary}{Screenreader mode for main visual container. Choices: \code{TRUE}, \code{FALSE}. Whether the main visual container is visible to screenreaders (Text in the header and footer are always available to screenreaders.). Flourish type hint: boolean, defaults to \code{NULL}}
}
\value{
A Flourish chart
}
\description{
Set survey chart details
}
\examples{
try(
  flourish(chart_type = "survey", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_survey_data(gapminder) |>
    set_survey_details()
)
}
