% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_table_details.R
\name{set_table_details}
\alias{set_table_details}
\title{Set table chart details}
\usage{
set_table_details(
  .,
  table_min_width = NULL,
  table_border_width = NULL,
  table_border_color = NULL,
  sorting_enabled = NULL,
  sorting_custom_columns = NULL,
  pagination_amount = NULL,
  markdown_enabled = NULL,
  cell_height = NULL,
  cell_padding_vertical = NULL,
  cell_padding_horizontal = NULL,
  cell_wrap = NULL,
  cell_horizontal_alignment = NULL,
  cell_numeric_horizontal_alignment = NULL,
  cell_vertical_alignment = NULL,
  cell_fill_1 = NULL,
  cell_fill_2 = NULL,
  cell_fill_direction = NULL,
  cell_fill_custom_enabled = NULL,
  cell_fill_custom_categorical = NULL,
  cell_fill_custom_numeric = NULL,
  cell_fill_margin_vertical = NULL,
  cell_fill_margin_horizontal = NULL,
  cell_border_mode = NULL,
  cell_border_color = NULL,
  cell_border_width = NULL,
  cell_border_style = NULL,
  cell_font_color = NULL,
  cell_link_color = NULL,
  cell_font_size = NULL,
  cell_font_size_custom = NULL,
  column_width_mode = NULL,
  column_widths = NULL,
  header_font_weight = NULL,
  header_fill = NULL,
  header_font_color = NULL,
  header_underline_sorting = NULL,
  header_style_default = NULL,
  header_height = NULL,
  header_padding = NULL,
  header_horizontal_alignment = NULL,
  header_font_default = NULL,
  header_font_size = NULL,
  header_border_default = NULL,
  header_border_color = NULL,
  header_border_width = NULL,
  header_border_style = NULL,
  cell_image_height = NULL,
  cell_image_radius = NULL,
  bar_columns_enabled = NULL,
  bar_columns_bar_1_columns = NULL,
  bar_columns_bar_1_column_name = NULL,
  bar_columns_type = NULL,
  bar_columns_bar_1_height = NULL,
  bar_columns_bar_1_width = NULL,
  bar_columns_bar_1_column_padding = NULL,
  bar_columns_bar_1_proportional = NULL,
  bar_columns_bar_1_same_scale = NULL,
  bar_columns_bar_1_zero_line = NULL,
  bar_columns_bar_1_zero_line_color = NULL,
  chart_color_scale_categorical_palette = NULL,
  chart_color_scale_categorical_extend = NULL,
  chart_color_scale_categorical_custom_palette = NULL,
  chart_color_scale_numeric_type = NULL,
  chart_color_scale_binning = NULL,
  chart_color_scale_bin_mode = NULL,
  chart_color_scale_bin_count = NULL,
  chart_color_scale_bin_thresholds = NULL,
  chart_color_scale_sequential_palette = NULL,
  chart_color_scale_sequential_reverse = NULL,
  chart_color_scale_sequential_custom_min = NULL,
  chart_color_scale_sequential_custom_max = NULL,
  chart_color_scale_sequential_color_space = NULL,
  chart_color_scale_sequential_custom_domain = NULL,
  chart_color_scale_sequential_domain_min = NULL,
  chart_color_scale_sequential_domain_max = NULL,
  chart_color_scale_diverging_palette = NULL,
  chart_color_scale_diverging_reverse = NULL,
  chart_color_scale_diverging_custom_min = NULL,
  chart_color_scale_diverging_custom_mid = NULL,
  chart_color_scale_diverging_custom_max = NULL,
  chart_color_scale_diverging_color_space = NULL,
  chart_color_scale_diverging_custom_domain = NULL,
  chart_color_scale_diverging_domain_min = NULL,
  chart_color_scale_diverging_domain_mid = NULL,
  chart_color_scale_diverging_domain_max = NULL,
  bar_columns_bar_1_labels = NULL,
  bar_columns_bar_1_label_size = NULL,
  chart_legend_show_legend = NULL,
  chart_legend_title_mode = NULL,
  chart_legend_title = NULL,
  chart_legend_swatch_width = NULL,
  chart_legend_swatch_height = NULL,
  chart_legend_swatch_radius = NULL,
  chart_legend_legend_items_padding = NULL,
  chart_legend_swatch_outline = NULL,
  chart_legend_swatch_outline_color = NULL,
  chart_legend_order_override = NULL,
  chart_legend_icon_height = NULL,
  chart_legend_icon_color = NULL,
  chart_legend_max_width = NULL,
  chart_legend_orientation = NULL,
  legend_position = NULL,
  chart_legend_container_alignment = NULL,
  chart_legend_container_orientation = NULL,
  chart_legend_container_title_weight = NULL,
  chart_legend_container_text_weight = NULL,
  chart_legend_container_text_color = NULL,
  chart_legend_container_text_size = NULL,
  line_columns_enabled = NULL,
  line_columns_line_1_columns = NULL,
  line_columns_line_1_column_name = NULL,
  line_columns_line_1_width = NULL,
  line_columns_line_1_height = NULL,
  line_columns_line_1_color = NULL,
  line_columns_line_1_area = NULL,
  line_columns_line_1_min_enabled = NULL,
  line_columns_line_1_min = NULL,
  line_columns_line_1_max_enabled = NULL,
  line_columns_line_1_max = NULL,
  mobile_view = NULL,
  mobile_breakpoint = NULL,
  mobile_cell_font_size = NULL,
  mobile_cell_font_size_custom = NULL,
  mobile_pagination_amount = NULL,
  search_enabled = NULL,
  search_mode = NULL,
  search_width = NULL,
  search_background_color = NULL,
  search_text_color = NULL,
  search_placeholder = NULL,
  no_results_text = NULL,
  search_hide_table = NULL,
  search_hide_published = NULL,
  search_columns = NULL,
  search_columns_custom = NULL,
  search_resize = NULL,
  search_dropdown = NULL,
  search_dropdown_persistent = NULL,
  search_val = NULL,
  chart_popup_show_popups = NULL,
  chart_popup_is_custom = NULL,
  chart_popup_custom_template = NULL,
  chart_popup_show_pointer = NULL,
  chart_popup_show_shadow = NULL,
  chart_popup_style_popups = NULL,
  chart_popup_text_color = NULL,
  chart_popup_align = NULL,
  chart_popup_text_size = NULL,
  chart_popup_fill_color = NULL,
  chart_popup_opacity = NULL,
  chart_popup_padding = NULL,
  chart_popup_border_radius = NULL,
  chart_popup_max_width = NULL,
  formatted_columns = NULL,
  chart_number_date_format_prefix = NULL,
  chart_number_date_format_suffix = NULL,
  chart_number_date_format_n_dec = NULL,
  chart_number_date_format_advanced = NULL,
  chart_number_date_format_negative_sign = NULL,
  chart_number_date_format_strip_zeros = NULL,
  chart_number_date_format_strip_separator = NULL,
  chart_number_date_format_transform_labels = NULL,
  chart_number_date_format_transform = NULL,
  chart_number_date_format_multiply_divide_constant = NULL,
  chart_number_date_format_exponentiate_constant = NULL,
  chart_layout_body_font = NULL,
  chart_layout_font_color = NULL,
  chart_layout_background_color_enabled = NULL,
  chart_layout_background_image_enabled = NULL,
  chart_layout_background_color = NULL,
  chart_layout_background_image_src = NULL,
  chart_layout_background_image_size = NULL,
  chart_layout_background_image_position = NULL,
  chart_layout_max_width_target = NULL,
  chart_layout_max_width = NULL,
  chart_layout_max_width_align = NULL,
  chart_layout_layout_order = NULL,
  chart_layout_space_between_sections = NULL,
  chart_layout_space_between_sections_custom = NULL,
  chart_layout_margin_top = NULL,
  chart_layout_margin_right = NULL,
  chart_layout_margin_bottom = NULL,
  chart_layout_margin_left = NULL,
  chart_layout_border_enabled = NULL,
  chart_layout_border_top_width = NULL,
  chart_layout_border_top_style = NULL,
  chart_layout_border_top_color = NULL,
  chart_layout_border_right_width = NULL,
  chart_layout_border_right_style = NULL,
  chart_layout_border_right_color = NULL,
  chart_layout_border_bottom_width = NULL,
  chart_layout_border_bottom_style = NULL,
  chart_layout_border_bottom_color = NULL,
  chart_layout_border_left_width = NULL,
  chart_layout_border_left_style = NULL,
  chart_layout_border_left_color = NULL,
  chart_layout_read_direction = NULL,
  chart_layout_font_size_mobile_small = NULL,
  chart_layout_font_size_mobile_big = NULL,
  chart_layout_font_size_tablet = NULL,
  chart_layout_font_size_desktop = NULL,
  chart_layout_font_size_big_screen = NULL,
  chart_layout_breakpoint_mobile_small = NULL,
  chart_layout_breakpoint_mobile_big = NULL,
  chart_layout_breakpoint_tablet = NULL,
  chart_layout_breakpoint_desktop = NULL,
  chart_layout_breakpoint_big_screen = NULL,
  chart_layout_header_align = NULL,
  chart_layout_title = NULL,
  chart_layout_title_styling = NULL,
  chart_layout_title_font = NULL,
  chart_layout_title_size = NULL,
  chart_layout_title_size_custom = NULL,
  chart_layout_title_weight = NULL,
  chart_layout_title_color = NULL,
  chart_layout_title_line_height = NULL,
  chart_layout_title_space_above = NULL,
  chart_layout_title_space_above_custom = NULL,
  chart_layout_subtitle = NULL,
  chart_layout_subtitle_styling = NULL,
  chart_layout_subtitle_font = NULL,
  chart_layout_subtitle_size = NULL,
  chart_layout_subtitle_size_custom = NULL,
  chart_layout_subtitle_weight = NULL,
  chart_layout_subtitle_color = NULL,
  chart_layout_subtitle_line_height = NULL,
  chart_layout_subtitle_space_above = NULL,
  chart_layout_subtitle_space_above_custom = NULL,
  chart_layout_header_text = NULL,
  chart_layout_header_text_styling = NULL,
  chart_layout_header_text_size = NULL,
  chart_layout_header_text_size_custom = NULL,
  chart_layout_header_text_weight = NULL,
  chart_layout_header_text_color = NULL,
  chart_layout_header_text_line_height = NULL,
  chart_layout_header_text_space_above = NULL,
  chart_layout_header_text_space_above_custom = NULL,
  chart_layout_header_border = NULL,
  chart_layout_header_border_width = NULL,
  chart_layout_header_border_color = NULL,
  chart_layout_header_border_style = NULL,
  chart_layout_header_border_space = NULL,
  chart_layout_header_logo_enabled = NULL,
  chart_layout_header_logo_src = NULL,
  chart_layout_header_logo_alt = NULL,
  chart_layout_header_logo_link_url = NULL,
  chart_layout_header_logo_height = NULL,
  chart_layout_header_logo_align = NULL,
  chart_layout_header_logo_position_inside = NULL,
  chart_layout_header_logo_position_outside = NULL,
  chart_layout_header_logo_margin_top = NULL,
  chart_layout_header_logo_margin_right = NULL,
  chart_layout_header_logo_margin_bottom = NULL,
  chart_layout_header_logo_margin_left = NULL,
  chart_layout_footer_align = NULL,
  chart_layout_footer_text_size = NULL,
  chart_layout_footer_text_color = NULL,
  chart_layout_footer_styling = NULL,
  chart_layout_footer_font = NULL,
  chart_layout_footer_text_weight = NULL,
  chart_layout_source_name = NULL,
  chart_layout_source_url = NULL,
  chart_layout_multiple_sources = NULL,
  chart_layout_source_name_2 = NULL,
  chart_layout_source_url_2 = NULL,
  chart_layout_source_name_3 = NULL,
  chart_layout_source_url_3 = NULL,
  chart_layout_source_label = NULL,
  chart_layout_footer_note = NULL,
  chart_layout_footer_note_secondary = NULL,
  chart_layout_advanced_note_styling = NULL,
  chart_layout_footer_timestamp_format = NULL,
  chart_layout_footer_logo_enabled = NULL,
  chart_layout_footer_logo_src = NULL,
  chart_layout_footer_logo_src_light = NULL,
  chart_layout_footer_logo_alt = NULL,
  chart_layout_footer_logo_link_url = NULL,
  chart_layout_footer_logo_height = NULL,
  chart_layout_footer_logo_margin = NULL,
  chart_layout_footer_logo_order = NULL,
  chart_layout_footer_align_vertical = NULL,
  chart_layout_footer_border = NULL,
  chart_layout_footer_border_width = NULL,
  chart_layout_footer_border_color = NULL,
  chart_layout_footer_border_style = NULL,
  chart_layout_footer_border_space = NULL,
  chart_layout_screenreader_text_primary = NULL,
  chart_layout_screenreader_label = NULL,
  chart_layout_screenreader_hide_primary = NULL,
  chart_url_settings_enabled = NULL,
  chart_url_settings_whitelist = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{table_min_width}{Min width. Flourish type hint: number, defaults to \code{NULL}}

\item{table_border_width}{Border. Flourish type hint: number, defaults to \code{NULL}}

\item{table_border_color}{Border color. Flourish type hint: color, defaults to \code{NULL}}

\item{sorting_enabled}{Sortable columns. Choices: all, none, custom. Determines whether a viewer can sort the table by clicking on the column headers. Custom allows the user only to sort listed columns. Flourish type hint: string, defaults to \code{NULL}}

\item{sorting_custom_columns}{Columns to make sortable. Write each column name you want to make sortable on a new line. Flourish type hint: text, defaults to \code{NULL}}

\item{pagination_amount}{Rows per page. If you dont want pagination, set this number to a value greater than the number of rows in your data. Flourish type hint: number, defaults to \code{NULL}}

\item{markdown_enabled}{Cell text styling. Choices: \code{TRUE}, \code{FALSE}. Enables Markdown for your cells. For example for \emph{italic}, \strong{bold} text. Flourish type hint: boolean, defaults to \code{NULL}}

\item{cell_height}{Min height. Flourish type hint: number, defaults to \code{NULL}}

\item{cell_padding_vertical}{Padding vertical. Flourish type hint: number, defaults to \code{NULL}}

\item{cell_padding_horizontal}{Padding horizontal. Flourish type hint: number, defaults to \code{NULL}}

\item{cell_wrap}{Wrap text in cell. Flourish type hint: boolean, defaults to \code{NULL}}

\item{cell_horizontal_alignment}{Text. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{cell_numeric_horizontal_alignment}{Numbers. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{cell_vertical_alignment}{Vertical alignment. Choices: top, center, bottom. Flourish type hint: string, defaults to \code{NULL}}

\item{cell_fill_1}{Main color. Flourish type hint: color, defaults to \code{NULL}}

\item{cell_fill_2}{Stripe color. Flourish type hint: color, defaults to \code{NULL}}

\item{cell_fill_direction}{Stripe direction. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{cell_fill_custom_enabled}{Color cells based on content. Flourish type hint: boolean, defaults to \code{NULL}}

\item{cell_fill_custom_categorical}{Categorical coloring. To color cells based on their content, use the format:  cell-color/text-color : cell-text : column name OR row-number To specify a row, write "row-" followed by the row number from the datasheet.  Its also possible to leave out the column name or row number if you want to target all the cells in the table For example:   red/white : Chelsea  Use * in place of cell-text to select an entire column or row. For example:  #0000dd/#ffffff : Alabama #ffffff/#dddddd : * : State #0000dd/#ffffff : * : row-2 #ffffff/#dddddd : n/a : Capital Flourish type hint: text, defaults to \code{NULL}}

\item{cell_fill_custom_numeric}{Numeric coloring. Color cells numerically based on their content. For sequential coloring, use the format:   min_color >> max_color : column_name : domain_min >> domain_max  The domain is optional, for example:  #ffffff >> #00ff00 : Population #ffffff >> #00ff00 : Population : 0 >> 1000000000  For diverging coloring, use the format:  min_color >> midpt_color >> max_color : col_name : domain_min >> domain_mid >> domain_max  The domain is optional, for example:  #ffffff >> #00ff00 >> #ff00ff : Vote share #ffffff >> #00ff00 >> #ff00ff : Vote share : 0 >> 50 >> 100  Where both numeric and categorical coloring are used on a cell, categorical will override numeric. Flourish type hint: text, defaults to \code{NULL}}

\item{cell_fill_margin_vertical}{Vertical margin. Flourish type hint: number, defaults to \code{NULL}}

\item{cell_fill_margin_horizontal}{Horizontal margin. Flourish type hint: number, defaults to \code{NULL}}

\item{cell_border_mode}{Mode. Choices: none, horizontal, vertical, both. Flourish type hint: string, defaults to \code{NULL}}

\item{cell_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{cell_border_width}{Thickness. Flourish type hint: number, defaults to \code{NULL}}

\item{cell_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{cell_font_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{cell_link_color}{Link color. Flourish type hint: color, defaults to \code{NULL}}

\item{cell_font_size}{Size. Choices: 1, 1.25, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{cell_font_size_custom}{Custom. Specify a custom responsive font size in rems Best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{column_width_mode}{Choices: auto, equal, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{column_widths}{Column widths. A comma separated lists of column widths. The first column will be the first number in the list. You can use "px" to set fixed widths or "\%" to use percentage of total width. (e.g. 200px,200px,25\%). Flourish type hint: string, defaults to \code{NULL}}

\item{header_font_weight}{Text weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{header_fill}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{header_font_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{header_underline_sorting}{Underline sorting header. Flourish type hint: boolean, defaults to \code{NULL}}

\item{header_style_default}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{header_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{header_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{header_horizontal_alignment}{Horizontal alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{header_font_default}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{header_font_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{header_border_default}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{header_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{header_border_width}{Thickness. Flourish type hint: number, defaults to \code{NULL}}

\item{header_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{cell_image_height}{Height. as a percentage of the cell height. Flourish type hint: number, defaults to \code{NULL}}

\item{cell_image_radius}{Rounded corners. Sets the radius of the corner roundness (in \%). A value of 50 means your image will show as a circle. Flourish type hint: number, defaults to \code{NULL}}

\item{bar_columns_enabled}{Choices: \code{TRUE}, \code{FALSE}. This setting will convert numerical values in your data to a simple visual bar chart. Flourish type hint: boolean, defaults to \code{NULL}}

\item{bar_columns_bar_1_columns}{Which columns should be part of the bar chart?. Just write down the column names on a new line. Flourish type hint: text, defaults to \code{NULL}}

\item{bar_columns_bar_1_column_name}{New column name. Flourish type hint: string, defaults to \code{NULL}}

\item{bar_columns_type}{Bar type. Choices: bars, columns. Flourish type hint: string, defaults to \code{NULL}}

\item{bar_columns_bar_1_height}{Max height. Flourish type hint: number, defaults to \code{NULL}}

\item{bar_columns_bar_1_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{bar_columns_bar_1_column_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{bar_columns_bar_1_proportional}{Mode. Choices: \code{FALSE}, \code{TRUE}. Display data in a proportional bar chart or a standard stacked bar chart with a fixed min and max, e.g. for negative numbers. Flourish type hint: boolean, defaults to \code{NULL}}

\item{bar_columns_bar_1_same_scale}{Consistent scale. If enabled, all the bar charts in the table will use the same X scale. Flourish type hint: boolean, defaults to \code{NULL}}

\item{bar_columns_bar_1_zero_line}{Show zero line. Flourish type hint: boolean, defaults to \code{NULL}}

\item{bar_columns_bar_1_zero_line_color}{Zero line color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_categorical_palette}{Palette. Flourish type hint: colors, defaults to \code{NULL}}

\item{chart_color_scale_categorical_extend}{Extend. Automatically generate additional colours when needed to avoid the palette colours being used more than once. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_categorical_custom_palette}{Custom overrides. Enter the label name for which you wish to set the color, followed by a colon and the desired color value.  Colors can be set using Hex, RGB, color names or RGBA, if you want to set the opacity Multiple colors can be set using multiple lines. For example:  Party 1: red Party 2: #4455AA Party 3: rgb(30,168,26). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_color_scale_numeric_type}{Scale type. Choices: sequential, diverging. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_binning}{Choices: \code{FALSE}, \code{TRUE}. In linear mode, the color scale will run as a smooth gradient between 2 colors. In binned mode, the gradient will be divided in smaller blocks. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_bin_mode}{Binning mode. Choices: fixed, quantile, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_bin_count}{Number of bins. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_bin_thresholds}{Custom thresholds. Enter your desired thresholds, separating them with a ";". For instance, "5;10;15".  Bins form as follows:  From the datas minimum value (domain minimum) up to the first threshold.  Between consecutive thresholds.  From the last threshold to the datas maximum value (domain maximum).   For "5;10;15", youll get four bins based on your datas range. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_mid}{Midpoint color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_mid}{Mid. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{bar_columns_bar_1_labels}{Labels. Choices: none, popup, inline. Flourish type hint: string, defaults to \code{NULL}}

\item{bar_columns_bar_1_label_size}{Inline label size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_show_legend}{Legend. Choices: \code{TRUE}, \code{FALSE}. A legend will not show with a single entry. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_legend_title_mode}{Title mode. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_title}{Title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_swatch_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_swatch_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_swatch_radius}{Roundness. The radius of the corners of the swatch (in pixels). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_legend_items_padding}{Padding. Padding between legend items ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_swatch_outline}{Outline. An optional outline for the swatch in the legend. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_legend_swatch_outline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_legend_order_override}{Custom order override. Manually specify the order of legend entries (one entry per line). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_legend_icon_height}{Height. Height of icon ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_icon_color}{Color. Fallback color (icon color if not determined by template). Flourish type hint: color, defaults to \code{NULL}}

\item{chart_legend_max_width}{Max width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_legend_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{legend_position}{Legend position. Choices: above, below. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_alignment}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_title_weight}{Title weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_text_weight}{Text weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_legend_container_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{line_columns_enabled}{Choices: \code{TRUE}, \code{FALSE}. This setting will convert numerical values in your data to a simple visual line chart. Flourish type hint: boolean, defaults to \code{NULL}}

\item{line_columns_line_1_columns}{Which columns should be part of the line chart?. Write down the column names on a new line. Flourish type hint: text, defaults to \code{NULL}}

\item{line_columns_line_1_column_name}{New column name. Flourish type hint: string, defaults to \code{NULL}}

\item{line_columns_line_1_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{line_columns_line_1_height}{Max height. Flourish type hint: number, defaults to \code{NULL}}

\item{line_columns_line_1_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{line_columns_line_1_area}{Area chart. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{line_columns_line_1_min_enabled}{Min. Choices: \code{TRUE}, \code{FALSE}. Use a fixed min value. Flourish type hint: boolean, defaults to \code{NULL}}

\item{line_columns_line_1_min}{Min value. Flourish type hint: number, defaults to \code{NULL}}

\item{line_columns_line_1_max_enabled}{Max. Choices: \code{TRUE}, \code{FALSE}. Use a fixed max value. Flourish type hint: boolean, defaults to \code{NULL}}

\item{line_columns_line_1_max}{Max value. Flourish type hint: number, defaults to \code{NULL}}

\item{mobile_view}{Mobile view. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{mobile_breakpoint}{Mobile breakpoint width. Flourish type hint: number, defaults to \code{NULL}}

\item{mobile_cell_font_size}{Text size. Choices: 1, 1.25, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{mobile_cell_font_size_custom}{Custom. Specify a custom responsive font size in rems Best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{mobile_pagination_amount}{Rows per page. If you dont want pagination, set this number to a value greater than the number of rows in your data. Flourish type hint: number, defaults to \code{NULL}}

\item{search_enabled}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{search_mode}{Search mode. Choices: anywhere, start, whole. Anywhere : Search for the specified text anywhere in the cell  Start : Search for the specified text at the beginning of the cell  Whole : Search for the exact specified text in the cell. Flourish type hint: string, defaults to \code{NULL}}

\item{search_width}{Width. Width of search container. Flourish type hint: number, defaults to \code{NULL}}

\item{search_background_color}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{search_text_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{search_placeholder}{Placeholder text. Flourish type hint: string, defaults to \code{NULL}}

\item{no_results_text}{Text when no results. Flourish type hint: string, defaults to \code{NULL}}

\item{search_hide_table}{Hide table before search. Flourish type hint: boolean, defaults to \code{NULL}}

\item{search_hide_published}{Hide search when published. Keeps the search box visible in the editor (e.g. for filtering in a story) but hides it in the published visualization or storyFlourish type hint: boolean, defaults to \code{NULL}}

\item{search_columns}{Columns to search. Choices: all, first, custom. Choose which columns you want to search through. You might want to narrow this down with big datasets to speed it up You can select specific columns by writing them down on a new line, like this:  Column A Column B. Flourish type hint: string, defaults to \code{NULL}}

\item{search_columns_custom}{Flourish type hint: text, defaults to \code{NULL}}

\item{search_resize}{Resize table after searching. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{search_dropdown}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{search_dropdown_persistent}{Always show dropdown suggestions. Flourish type hint: boolean, defaults to \code{NULL}}

\item{search_val}{Search value. The current string value of the search box. Flourish type hint: hidden, defaults to \code{NULL}}

\item{chart_popup_show_popups}{Popups. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_is_custom}{Popup contents. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_custom_template}{Popup content. The text to appear in the popup. You can use \code{{{column_name}}} to add a value from your data. It must be in a selected column, but you can add columns to Metadata if you just want to include them for use in the popup. Advanced users can include HTML to apply layouts, formatting, images, etc. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_popup_show_pointer}{Pointer. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_show_shadow}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_style_popups}{Custom styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_text_color}{Text colour. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_align}{Alignment. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_text_size}{Font size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_fill_color}{Fill colour. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_opacity}{Fill opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_border_radius}{Radius. Corner radius of popup. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_max_width}{Max width. Flourish type hint: number, defaults to \code{NULL}}

\item{formatted_columns}{Columns to format. Write down the column names on a new line that you want to apply formatting to. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_number_date_format_prefix}{Prefix. Text to place in front of a number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_suffix}{Suffix. Text to place after a number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_n_dec}{Decimal places. Decimal places. You can also enter a negative integer to round to a whole number with that many zeros. For example, -2 will round to the nearest hundred. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_negative_sign}{Styling of negative numbers. Choices: -$nk, $-nk, ($nk), $(n)k, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_strip_zeros}{Remove trailing zeros. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_strip_separator}{Hide thousands separator below 10,000. Turn off if you want four-digit numbers to include a separator, e.g. 1,234 rather than 1234. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_transform_labels}{Multiply/divide values. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_transform}{Choices: multiply, divide, exponentiate. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_multiply_divide_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_exponentiate_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_body_font}{Main font. This font will apply to the whole graphic by default. You can optionally change the font for the title, subtitle, footer, etc in the Header and Footer settings panels. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_font_color}{Text color. This color will apply to the whole graphic by default, You can optionally change the color for individual text elements, in other settings panels. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_color_enabled}{Color. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_image_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_background_image_size}{Size. Choices: cover, contain, auto, 100\% 100\%. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_background_image_position}{Position. Choices: top left, top center, top right, center left, center center, center right, bottom left, bottom center, bottom right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width_target}{Maximum width. Choices: none, wrapper, primary. Apply a maximum width to just the main graphic or everything (main graphic plus header, footer, etc). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width}{Maximum width. Leave blank to stretch to container width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_max_width_align}{Align. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_layout_order}{Layout order. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections}{Space between sections. Choices: 0.5, 1, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_enabled}{Show borders around visualisation. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_border_top_width}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_top_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_top_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_right_width}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_right_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_right_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_bottom_width}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_bottom_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_bottom_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_left_width}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_left_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_left_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_read_direction}{Read direction. Choices: ltr, rtl. This will change the reading direction of the main text elements on the page. Its not possible to adjust this on all elements, such as axes.  Note that when direction is set to right to left any alignment icons will be reversed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_align}{Alignment. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_title_font}{Title Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_title_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_title_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_subtitle_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_subtitle_font}{Subtitle Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_subtitle_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_subtitle_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_header_text_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_text_size}{Size. Choices: 1.2, 1.4, 1.6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_text_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_space}{Space. Space between border and header text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_enabled}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_header_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_align}{Align. Choices: inside, outside. Align logo inside either the header or the main visualization container. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_inside}{Position. Choices: top, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_outside}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_align}{Alignment. Choices: left, center, right, justify. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_styling}{Advanced footer styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_font}{Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_footer_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_multiple_sources}{Multiple sources. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_source_name_2}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_2}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name_3}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_3}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_label}{Source label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_note}{Note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_footer_note_secondary}{Note (secondary). The secondary note is placed below the source and primary note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_advanced_note_styling}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_timestamp_format}{Time/date stamp formatting. Choices: \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M, \%H:\%M \%p, \%H:\%M:\%S, \%d \%b \%Y, \%H:\%M, \%d \%b \%Y, \%H:\%M, \%d/\%m/\%Y, \%H:\%M, \%m/\%d/\%Y, \%H:\%M, \%d-\%m-\%Y, \%H:\%M, \%m-\%d-\%Y, \%d \%b \%Y, \%m/\%d/\%Y, \%d/\%m/\%Y, \%m-\%d-\%Y, \%d-\%m-\%Y, \%Y-\%m-\%d, \%Y-\%m-\%d \%H:\%M:\%S. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src_light}{Image (light version). If provided this version will be used whenever the background color is dark. Flourish type hint: hidden, defaults to \code{NULL}}

\item{chart_layout_footer_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_margin}{Margin. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_order}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_align_vertical}{V. align. Choices: flex-start, center, flex-end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_space}{Space. Space between border and footer text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_screenreader_text_primary}{Screenreader description. A text alternative to the visual content that will only be visible to screenreaders, e.g. The line chart shows China consistently higher than the other countries since 1990.  Do not replicate your title, since that will also be read by screenreaders. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_layout_screenreader_label}{Screenreader label. A short text label given to the main Flourish embed wrapper to provide an accessible name that is only visible to screenreaders Added in the form of an "aria-label". Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_screenreader_hide_primary}{Screenreader mode for main visual container. Choices: \code{TRUE}, \code{FALSE}. Whether the main visual container is visible to screenreaders (Text in the header and footer are always available to screenreaders.). Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_url_settings_enabled}{Update project from URL. Make it possible to change the visualizations settings/state from the URL hash, for example when embedding in an iframe or popup panel. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_url_settings_whitelist}{Allowed state properties. List each state property you want to make changeable with the URL hash See the template page for a full list of state properties.  Add one property name per line. Use dot format for nested properties, e.g. \code{layout.background_color}. Flourish type hint: text, defaults to \code{NULL}}
}
\value{
A Flourish chart
}
\description{
Set table chart details
}
\examples{
try(
  flourish(chart_type = "table", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_table_data(gapminder) |>
    set_table_details()
)
}
