% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_timeline_details.R
\name{set_timeline_details}
\alias{set_timeline_details}
\title{Set timeline chart details}
\usage{
set_timeline_details(
  .,
  widescreen_orientation = NULL,
  height_mode = NULL,
  facet_aspect = NULL,
  facet_aspect_custom = NULL,
  facet_aspect_breakpoint = NULL,
  facet_aspect_mobile = NULL,
  facet_aspect_mobile_custom = NULL,
  line_color = NULL,
  line_width = NULL,
  line_cap = NULL,
  events_scale_type = NULL,
  background_opacity = NULL,
  background_sizing = NULL,
  background_zoom = NULL,
  background_gradient_size = NULL,
  background_gradient_opacity = NULL,
  background_gradient_blur = NULL,
  chart_color_scale_categorical_palette = NULL,
  chart_color_scale_categorical_extend = NULL,
  chart_color_scale_categorical_custom_palette = NULL,
  chart_color_scale_numeric_type = NULL,
  chart_color_scale_binning = NULL,
  chart_color_scale_bin_mode = NULL,
  chart_color_scale_bin_count = NULL,
  chart_color_scale_bin_thresholds = NULL,
  chart_color_scale_sequential_palette = NULL,
  chart_color_scale_sequential_reverse = NULL,
  chart_color_scale_sequential_custom_min = NULL,
  chart_color_scale_sequential_custom_max = NULL,
  chart_color_scale_sequential_color_space = NULL,
  chart_color_scale_sequential_custom_domain = NULL,
  chart_color_scale_sequential_domain_min = NULL,
  chart_color_scale_sequential_domain_max = NULL,
  chart_color_scale_diverging_palette = NULL,
  chart_color_scale_diverging_reverse = NULL,
  chart_color_scale_diverging_custom_min = NULL,
  chart_color_scale_diverging_custom_mid = NULL,
  chart_color_scale_diverging_custom_max = NULL,
  chart_color_scale_diverging_color_space = NULL,
  chart_color_scale_diverging_custom_domain = NULL,
  chart_color_scale_diverging_domain_min = NULL,
  chart_color_scale_diverging_domain_mid = NULL,
  chart_color_scale_diverging_domain_max = NULL,
  marker_size = NULL,
  marker_background_color = NULL,
  marker_background_color_custom = NULL,
  marker_border_width = NULL,
  marker_border_color = NULL,
  marker_border_color_custom = NULL,
  connector_size = NULL,
  connector_space = NULL,
  connector_color = NULL,
  connector_opacity = NULL,
  content_background_color = NULL,
  content_padding = NULL,
  content_space_between = NULL,
  content_consistent_size = NULL,
  content_border_radius = NULL,
  individual_borders = NULL,
  content_border_width = NULL,
  content_border_opacity = NULL,
  content_border_color = NULL,
  content_border_color_custom = NULL,
  content_border_top_width = NULL,
  content_border_top_opacity = NULL,
  content_border_top_color = NULL,
  content_border_top_color_custom = NULL,
  content_border_right_width = NULL,
  content_border_right_opacity = NULL,
  content_border_right_color = NULL,
  content_border_right_color_custom = NULL,
  content_border_bottom_width = NULL,
  content_border_bottom_opacity = NULL,
  content_border_bottom_color = NULL,
  content_border_bottom_color_custom = NULL,
  content_border_left_width = NULL,
  content_border_left_opacity = NULL,
  content_border_left_color = NULL,
  content_border_left_color_custom = NULL,
  content_title_alignment = NULL,
  content_title_styling = NULL,
  content_title_line_height = NULL,
  content_title_size = NULL,
  content_title_size_custom = NULL,
  content_title_weight = NULL,
  content_title_color = NULL,
  content_title_color_custom = NULL,
  content_title_space_above = NULL,
  content_title_space_above_custom = NULL,
  content_subtitle_alignment = NULL,
  content_subtitle_styling = NULL,
  content_subtitle_line_height = NULL,
  content_subtitle_size = NULL,
  content_subtitle_size_custom = NULL,
  content_subtitle_weight = NULL,
  content_subtitle_color = NULL,
  content_subtitle_color_custom = NULL,
  content_subtitle_space_above = NULL,
  content_subtitle_space_above_custom = NULL,
  content_text_alignment = NULL,
  content_text_styling = NULL,
  content_text_line_height = NULL,
  content_text_size = NULL,
  content_text_size_custom = NULL,
  content_text_weight = NULL,
  content_text_color = NULL,
  content_text_color_custom = NULL,
  content_text_space_above = NULL,
  content_text_space_above_custom = NULL,
  link_weight = NULL,
  link_color = NULL,
  link_underline = NULL,
  image_location = NULL,
  image_opacity = NULL,
  image_display = NULL,
  image_sizing = NULL,
  image_zoom = NULL,
  image_content_sizing = NULL,
  image_content_zoom = NULL,
  image_float_position = NULL,
  image_inline_position = NULL,
  image_alignment = NULL,
  content_banner_depth = NULL,
  content_banner_depth_custom = NULL,
  image_width = NULL,
  image_margin = NULL,
  image_aspect = NULL,
  navigation_step_type = NULL,
  navigation_screen_proportion = NULL,
  navigation_event_number = NULL,
  navigation_icon_type = NULL,
  navigation_icon_size = NULL,
  navigation_icon_opacity = NULL,
  navigation_icon_custom = NULL,
  navigation_icon_color = NULL,
  navigation_gradient_color = NULL,
  navigation_gradient_width = NULL,
  navigation_gradient_opacity = NULL,
  chart_x_axis_axis_position = NULL,
  chart_x_axis_y_axis_position = NULL,
  chart_x_axis_numeric_scale_type = NULL,
  chart_x_axis_linear_min = NULL,
  chart_x_axis_linear_max = NULL,
  chart_x_axis_log_min = NULL,
  chart_x_axis_log_max = NULL,
  chart_x_axis_datetime_min = NULL,
  chart_x_axis_datetime_max = NULL,
  chart_x_axis_flip = NULL,
  chart_x_axis_show_scale_settings = NULL,
  chart_x_axis_nice = NULL,
  chart_x_axis_zero_axis = NULL,
  chart_x_axis_title_mode = NULL,
  chart_x_axis_title = NULL,
  chart_x_axis_title_styling = NULL,
  chart_x_axis_title_weight = NULL,
  chart_x_axis_title_color = NULL,
  chart_x_axis_title_size = NULL,
  chart_x_axis_title_padding = NULL,
  chart_x_axis_tick_label_position = NULL,
  chart_x_axis_tick_label_styling = NULL,
  chart_x_axis_tick_label_size = NULL,
  chart_x_axis_tick_label_color = NULL,
  chart_x_axis_tick_padding = NULL,
  chart_x_axis_tick_label_angle = NULL,
  chart_x_axis_tick_label_weight = NULL,
  chart_x_axis_tick_label_max_lines = NULL,
  chart_x_axis_tick_label_line_height = NULL,
  chart_x_axis_tick_label_space_mode = NULL,
  chart_x_axis_tick_label_space = NULL,
  chart_x_axis_tick_mode = NULL,
  chart_x_axis_tick_number = NULL,
  chart_x_axis_tick_custom = NULL,
  chart_x_axis_line_and_tick_color = NULL,
  chart_x_axis_line_and_tick_width = NULL,
  chart_x_axis_tick_length = NULL,
  chart_x_axis_tick_side = NULL,
  chart_x_axis_line_visible = NULL,
  chart_x_axis_edge_padding = NULL,
  chart_x_axis_gridlines_visible = NULL,
  chart_x_axis_gridlines_styling = NULL,
  chart_x_axis_gridline_color = NULL,
  chart_x_axis_gridline_style = NULL,
  chart_x_axis_gridline_width = NULL,
  chart_x_axis_gridline_category_dividers = NULL,
  chart_x_axis_gridline_category_dividers_extend = NULL,
  chart_y_axis_axis_visible = NULL,
  chart_y_axis_numeric_scale_type = NULL,
  chart_y_axis_linear_min = NULL,
  chart_y_axis_linear_max = NULL,
  chart_y_axis_log_min = NULL,
  chart_y_axis_log_max = NULL,
  chart_y_axis_datetime_min = NULL,
  chart_y_axis_datetime_max = NULL,
  chart_y_axis_flip = NULL,
  chart_y_axis_show_scale_settings = NULL,
  chart_y_axis_nice = NULL,
  chart_y_axis_zero_axis = NULL,
  chart_y_axis_title_mode = NULL,
  chart_y_axis_title = NULL,
  chart_y_axis_title_styling = NULL,
  chart_y_axis_title_position = NULL,
  chart_y_axis_title_weight = NULL,
  chart_y_axis_title_color = NULL,
  chart_y_axis_title_size = NULL,
  chart_y_axis_title_padding = NULL,
  chart_y_axis_tick_label_position = NULL,
  chart_y_axis_tick_label_styling = NULL,
  chart_y_axis_tick_label_size = NULL,
  chart_y_axis_tick_label_color = NULL,
  chart_y_axis_tick_padding = NULL,
  chart_y_axis_tick_label_angle = NULL,
  chart_y_axis_tick_label_weight = NULL,
  chart_y_axis_tick_label_max_lines = NULL,
  chart_y_axis_tick_label_line_height = NULL,
  chart_y_axis_tick_label_space_mode = NULL,
  chart_y_axis_tick_label_space = NULL,
  chart_y_axis_tick_mode = NULL,
  chart_y_axis_tick_number = NULL,
  chart_y_axis_tick_custom = NULL,
  chart_y_axis_line_and_tick_color = NULL,
  chart_y_axis_line_and_tick_width = NULL,
  chart_y_axis_tick_length = NULL,
  chart_y_axis_tick_side = NULL,
  chart_y_axis_line_visible = NULL,
  chart_y_axis_edge_padding = NULL,
  chart_y_axis_gridlines_visible = NULL,
  chart_y_axis_gridlines_styling = NULL,
  chart_y_axis_gridline_color = NULL,
  chart_y_axis_gridline_style = NULL,
  chart_y_axis_gridline_width = NULL,
  chart_y_axis_gridline_category_dividers = NULL,
  chart_y_axis_gridline_category_dividers_extend = NULL,
  chart_categorical_legend_show_legend = NULL,
  chart_categorical_legend_title_mode = NULL,
  chart_categorical_legend_title = NULL,
  chart_categorical_legend_swatch_width = NULL,
  chart_categorical_legend_swatch_height = NULL,
  chart_categorical_legend_swatch_radius = NULL,
  chart_categorical_legend_legend_items_padding = NULL,
  chart_categorical_legend_swatch_outline = NULL,
  chart_categorical_legend_swatch_outline_color = NULL,
  chart_categorical_legend_order_override = NULL,
  chart_categorical_legend_icon_height = NULL,
  chart_categorical_legend_icon_color = NULL,
  chart_categorical_legend_max_width = NULL,
  chart_categorical_legend_orientation = NULL,
  chart_layout_body_font = NULL,
  chart_layout_font_color = NULL,
  chart_layout_background_color_enabled = NULL,
  chart_layout_background_image_enabled = NULL,
  chart_layout_background_color = NULL,
  chart_layout_background_image_src = NULL,
  chart_layout_background_image_size = NULL,
  chart_layout_background_image_position = NULL,
  chart_layout_max_width_target = NULL,
  chart_layout_max_width = NULL,
  chart_layout_max_width_align = NULL,
  chart_layout_layout_order = NULL,
  chart_layout_space_between_sections = NULL,
  chart_layout_space_between_sections_custom = NULL,
  chart_layout_margin_top = NULL,
  chart_layout_margin_right = NULL,
  chart_layout_margin_bottom = NULL,
  chart_layout_margin_left = NULL,
  chart_layout_border_enabled = NULL,
  chart_layout_border_top_width = NULL,
  chart_layout_border_top_style = NULL,
  chart_layout_border_top_color = NULL,
  chart_layout_border_right_width = NULL,
  chart_layout_border_right_style = NULL,
  chart_layout_border_right_color = NULL,
  chart_layout_border_bottom_width = NULL,
  chart_layout_border_bottom_style = NULL,
  chart_layout_border_bottom_color = NULL,
  chart_layout_border_left_width = NULL,
  chart_layout_border_left_style = NULL,
  chart_layout_border_left_color = NULL,
  chart_layout_read_direction = NULL,
  chart_layout_font_size_mobile_small = NULL,
  chart_layout_font_size_mobile_big = NULL,
  chart_layout_font_size_tablet = NULL,
  chart_layout_font_size_desktop = NULL,
  chart_layout_font_size_big_screen = NULL,
  chart_layout_breakpoint_mobile_small = NULL,
  chart_layout_breakpoint_mobile_big = NULL,
  chart_layout_breakpoint_tablet = NULL,
  chart_layout_breakpoint_desktop = NULL,
  chart_layout_breakpoint_big_screen = NULL,
  chart_layout_header_align = NULL,
  chart_layout_title = NULL,
  chart_layout_title_styling = NULL,
  chart_layout_title_font = NULL,
  chart_layout_title_size = NULL,
  chart_layout_title_size_custom = NULL,
  chart_layout_title_weight = NULL,
  chart_layout_title_color = NULL,
  chart_layout_title_line_height = NULL,
  chart_layout_title_space_above = NULL,
  chart_layout_title_space_above_custom = NULL,
  chart_layout_subtitle = NULL,
  chart_layout_subtitle_styling = NULL,
  chart_layout_subtitle_font = NULL,
  chart_layout_subtitle_size = NULL,
  chart_layout_subtitle_size_custom = NULL,
  chart_layout_subtitle_weight = NULL,
  chart_layout_subtitle_color = NULL,
  chart_layout_subtitle_line_height = NULL,
  chart_layout_subtitle_space_above = NULL,
  chart_layout_subtitle_space_above_custom = NULL,
  chart_layout_header_text = NULL,
  chart_layout_header_text_styling = NULL,
  chart_layout_header_text_size = NULL,
  chart_layout_header_text_size_custom = NULL,
  chart_layout_header_text_weight = NULL,
  chart_layout_header_text_color = NULL,
  chart_layout_header_text_line_height = NULL,
  chart_layout_header_text_space_above = NULL,
  chart_layout_header_text_space_above_custom = NULL,
  chart_layout_header_border = NULL,
  chart_layout_header_border_width = NULL,
  chart_layout_header_border_color = NULL,
  chart_layout_header_border_style = NULL,
  chart_layout_header_border_space = NULL,
  chart_layout_header_logo_enabled = NULL,
  chart_layout_header_logo_src = NULL,
  chart_layout_header_logo_alt = NULL,
  chart_layout_header_logo_link_url = NULL,
  chart_layout_header_logo_height = NULL,
  chart_layout_header_logo_align = NULL,
  chart_layout_header_logo_position_inside = NULL,
  chart_layout_header_logo_position_outside = NULL,
  chart_layout_header_logo_margin_top = NULL,
  chart_layout_header_logo_margin_right = NULL,
  chart_layout_header_logo_margin_bottom = NULL,
  chart_layout_header_logo_margin_left = NULL,
  chart_layout_footer_align = NULL,
  chart_layout_footer_text_size = NULL,
  chart_layout_footer_text_color = NULL,
  chart_layout_footer_styling = NULL,
  chart_layout_footer_font = NULL,
  chart_layout_footer_text_weight = NULL,
  chart_layout_source_name = NULL,
  chart_layout_source_url = NULL,
  chart_layout_multiple_sources = NULL,
  chart_layout_source_name_2 = NULL,
  chart_layout_source_url_2 = NULL,
  chart_layout_source_name_3 = NULL,
  chart_layout_source_url_3 = NULL,
  chart_layout_source_label = NULL,
  chart_layout_footer_note = NULL,
  chart_layout_footer_note_secondary = NULL,
  chart_layout_advanced_note_styling = NULL,
  chart_layout_footer_timestamp_format = NULL,
  chart_layout_footer_logo_enabled = NULL,
  chart_layout_footer_logo_src = NULL,
  chart_layout_footer_logo_src_light = NULL,
  chart_layout_footer_logo_alt = NULL,
  chart_layout_footer_logo_link_url = NULL,
  chart_layout_footer_logo_height = NULL,
  chart_layout_footer_logo_margin = NULL,
  chart_layout_footer_logo_order = NULL,
  chart_layout_footer_align_vertical = NULL,
  chart_layout_footer_border = NULL,
  chart_layout_footer_border_width = NULL,
  chart_layout_footer_border_color = NULL,
  chart_layout_footer_border_style = NULL,
  chart_layout_footer_border_space = NULL,
  chart_layout_screenreader_text_primary = NULL,
  chart_layout_screenreader_label = NULL,
  chart_layout_screenreader_hide_primary = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{widescreen_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{height_mode}{Height mode. Choices: auto, standard, aspect. Auto: sets the height based on the chart type, data and width.  Standard: uses the default Flourish responsive sizing.  Aspect ratio: sets the aspect of each plot.  This setting is ignored when embedded in a fixed-height context, such as a simple iframe embedFlourish type hint: string, defaults to \code{NULL}}

\item{facet_aspect}{Desktop aspect. Choices: 0.75, 1, 1.5, custom. Aspect ratio of the visualization in desktop mode. Flourish type hint: string, defaults to \code{NULL}}

\item{facet_aspect_custom}{Custom. 1 is square, above 1 is wider, and below 1 is taller. Flourish type hint: number, defaults to \code{NULL}}

\item{facet_aspect_breakpoint}{Breakpoint. The width at which mobile mode ends and desktop begins. Flourish type hint: number, defaults to \code{NULL}}

\item{facet_aspect_mobile}{Mobile aspect. Choices: 0.75, 1, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{facet_aspect_mobile_custom}{Custom. 1 is square, above 1 is wider, and below 1 is taller. Flourish type hint: number, defaults to \code{NULL}}

\item{line_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{line_width}{Width. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{line_cap}{Cap. Choices: butt, round, square. The shape used to draw the end points. Flourish type hint: string, defaults to \code{NULL}}

\item{events_scale_type}{Choices: linear, timescale. Flourish type hint: string, defaults to \code{NULL}}

\item{background_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{background_sizing}{Sizing. Choices: contain, cover, zoom. Flourish type hint: string, defaults to \code{NULL}}

\item{background_zoom}{Zoom. Sets what size the image should fill the container as a percentage of the container size. Flourish type hint: number, defaults to \code{NULL}}

\item{background_gradient_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{background_gradient_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{background_gradient_blur}{Blur. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_categorical_palette}{Palette. Flourish type hint: colors, defaults to \code{NULL}}

\item{chart_color_scale_categorical_extend}{Extend. Automatically generate additional colours when needed to avoid the palette colours being used more than once. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_categorical_custom_palette}{Custom overrides. Enter the label name for which you wish to set the color, followed by a colon and the desired color value.  Colors can be set using Hex, RGB, color names or RGBA, if you want to set the opacity Multiple colors can be set using multiple lines. For example:  Party 1: red Party 2: #4455AA Party 3: rgb(30,168,26). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_color_scale_numeric_type}{Scale type. Choices: sequential, diverging. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_binning}{Choices: \code{FALSE}, \code{TRUE}. In linear mode, the color scale will run as a smooth gradient between 2 colors. In binned mode, the gradient will be divided in smaller blocks. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_bin_mode}{Binning mode. Choices: fixed, quantile, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_bin_count}{Number of bins. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_bin_thresholds}{Custom thresholds. Enter your desired thresholds, separating them with a ";". For instance, "5;10;15".  Bins form as follows:  From the datas minimum value (domain minimum) up to the first threshold.  Between consecutive thresholds.  From the last threshold to the datas maximum value (domain maximum).   For "5;10;15", youll get four bins based on your datas range. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_mid}{Midpoint color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_mid}{Mid. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{marker_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{marker_background_color}{Color. Choices: data, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{marker_background_color_custom}{Flourish type hint: color, defaults to \code{NULL}}

\item{marker_border_width}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{marker_border_color}{Color. Choices: data, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{marker_border_color_custom}{Flourish type hint: color, defaults to \code{NULL}}

\item{connector_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{connector_space}{Space. Sets the area around the markers to keep clear of content containers. Flourish type hint: number, defaults to \code{NULL}}

\item{connector_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{connector_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{content_background_color}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{content_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{content_space_between}{Space between. Defines the minimum amount of space between each content container. Flourish type hint: number, defaults to \code{NULL}}

\item{content_consistent_size}{Keep sizes consistent. Flourish type hint: boolean, defaults to \code{NULL}}

\item{content_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{individual_borders}{Style borders individually. Flourish type hint: boolean, defaults to \code{NULL}}

\item{content_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{content_border_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{content_border_color}{Color. Choices: data, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_border_color_custom}{Flourish type hint: color, defaults to \code{NULL}}

\item{content_border_top_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{content_border_top_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{content_border_top_color}{Color. Choices: data, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_border_top_color_custom}{Flourish type hint: color, defaults to \code{NULL}}

\item{content_border_right_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{content_border_right_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{content_border_right_color}{Color. Choices: data, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_border_right_color_custom}{Flourish type hint: color, defaults to \code{NULL}}

\item{content_border_bottom_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{content_border_bottom_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{content_border_bottom_color}{Color. Choices: data, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_border_bottom_color_custom}{Flourish type hint: color, defaults to \code{NULL}}

\item{content_border_left_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{content_border_left_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{content_border_left_color}{Color. Choices: data, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_border_left_color_custom}{Flourish type hint: color, defaults to \code{NULL}}

\item{content_title_alignment}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{content_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{content_title_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{content_title_size}{Size. Choices: 1.0, 1.2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_title_size_custom}{Custom. Specify a custom responsive font size. Best results will be with values between 0.8 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{content_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{content_title_color}{Color. Choices: data, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_title_color_custom}{Flourish type hint: color, defaults to \code{NULL}}

\item{content_title_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_title_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{content_subtitle_alignment}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{content_subtitle_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{content_subtitle_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{content_subtitle_size}{Size. Choices: 1.0, 1.2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_subtitle_size_custom}{Custom. Specify a custom responsive font size. Best results will be with values between 0.8 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{content_subtitle_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{content_subtitle_color}{Color. Choices: data, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_subtitle_color_custom}{Flourish type hint: color, defaults to \code{NULL}}

\item{content_subtitle_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_subtitle_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{content_text_alignment}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{content_text_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{content_text_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{content_text_size}{Size. Choices: 1.0, 1.2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_text_size_custom}{Custom. Specify a custom responsive font size. Best results will be with values between 0.8 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{content_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{content_text_color}{Color. Choices: data, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_text_color_custom}{Flourish type hint: color, defaults to \code{NULL}}

\item{content_text_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_text_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{link_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{link_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{link_underline}{Underline. Flourish type hint: boolean, defaults to \code{NULL}}

\item{image_location}{Location. Choices: none, marker, content. Flourish type hint: string, defaults to \code{NULL}}

\item{image_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{image_display}{Content display. Choices: float, inline, banner, background. Flourish type hint: string, defaults to \code{NULL}}

\item{image_sizing}{Sizing. Choices: contain, cover, zoom. Flourish type hint: string, defaults to \code{NULL}}

\item{image_zoom}{Zoom. Sets what size the image should fill the container as a percentage of the container size. Flourish type hint: number, defaults to \code{NULL}}

\item{image_content_sizing}{Sizing. Choices: contain, cover, zoom. Flourish type hint: string, defaults to \code{NULL}}

\item{image_content_zoom}{Zoom. Sets what size the image should fill the container as a percentage of the container size. Flourish type hint: number, defaults to \code{NULL}}

\item{image_float_position}{Position. Choices: 1, 2. Flourish type hint: string, defaults to \code{NULL}}

\item{image_inline_position}{Position. Choices: 1, 2, 3. Flourish type hint: string, defaults to \code{NULL}}

\item{image_alignment}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{content_banner_depth}{Banner depth. Choices: 2, 4, 6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{content_banner_depth_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{image_width}{Width. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{image_margin}{Margin. Flourish type hint: number, defaults to \code{NULL}}

\item{image_aspect}{Aspect ratio. As a multiplier of the images width. For example: a value of 2, means the image height is 2 times the image width. Flourish type hint: number, defaults to \code{NULL}}

\item{navigation_step_type}{Step by. Choices: screen, event. Determines by which measure the timeline is approximately moved each time the navigation buttons are clicked. Flourish type hint: string, defaults to \code{NULL}}

\item{navigation_screen_proportion}{Proportion. Flourish type hint: number, defaults to \code{NULL}}

\item{navigation_event_number}{Number. Flourish type hint: number, defaults to \code{NULL}}

\item{navigation_icon_type}{Icon. Choices: default, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{navigation_icon_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{navigation_icon_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{navigation_icon_custom}{Custom icon. Flourish type hint: url, defaults to \code{NULL}}

\item{navigation_icon_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{navigation_gradient_color}{Color. If undefined the color will default to the background of the visualization. Flourish type hint: color, defaults to \code{NULL}}

\item{navigation_gradient_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{navigation_gradient_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_axis_position}{Position. Choices: bottom, top, hidden. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_y_axis_position}{Y value. Vertical position of the X axis on the Y scale. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_numeric_scale_type}{Type. Choices: linear, log. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_linear_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_linear_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_log_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_log_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_datetime_min}{Min (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_datetime_max}{Max (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_flip}{Flip axis. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_show_scale_settings}{Configure default min/max. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_nice}{Round min/max. Choices: \code{TRUE}, \code{FALSE}. Enabling this option rounds the X axis start and end values to tidy numbers, such as 0, 50, 100 or whole years. Specified min and max values will override this feature. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_zero_axis}{Include zero. Choices: auto, on, off. Enable to extend the axis to include zero, even when data is all positive or all negative. This is ignored if you set specific min and max scale values. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_mode}{Type. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title}{Text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_title_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_title_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_position}{Position. Choices: default, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_tick_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_angle}{Angle. Choices: 0, 30, 45, 60, 90. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_max_lines}{Max lines. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_space_mode}{Space mode. Choices: auto, max, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_space}{Space. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_mode}{Mode. Choices: auto, number, custom. Sets the method for selecting tick values on the axis:  Auto automatically selects the ticks  Number allows you to choose the number of ticks to display.  Custom lets you select specific ticks to show.  Note: If the axis is categorical, Number mode is ignored. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_number}{Number. Approximate number of ticks or gridlines. The actual number will depend on the range of values, chart size, etc. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_custom}{One tick label per line. Specify the ticks to show, one per line. For dates, use the input format as specified in the datasheet.  You can also use \code{{{FIRST}}} and \code{{{LAST}}} to add ticks at the minimum and maximum of the axis. For example:  \code{{{FIRST}}} 25000 50000 75000 \code{{{LAST}}}. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_x_axis_line_and_tick_color}{Line color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_line_and_tick_width}{Line width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_length}{Tick length. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_side}{Tick side. Choices: out, in. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_line_visible}{Axis line. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_edge_padding}{Edge padding (\%). Space between start/end of axis line and first/last category tick as a percentage of the spacing between ticks. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_gridlines_visible}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridlines_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_gridline_style}{Style. Choices: solid, dash, dot, dot_dash. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_gridline_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_gridline_category_dividers}{Put lines between categories. On a categorical axis, this option disables the gridline on each tick and instead adds rules between each category to give a table-like design. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridline_category_dividers_extend}{Extend. Extends the dividers into the axis margins. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_axis_visible}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_numeric_scale_type}{Type. Choices: linear, log. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_linear_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_linear_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_log_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_log_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_datetime_min}{Min (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_datetime_max}{Max (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_flip}{Flip axis. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_show_scale_settings}{Configure default min/max. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_nice}{Round min/max. Choices: \code{TRUE}, \code{FALSE}. When enabled, scales automatically extend to "nice" rounded start/end values, such as hundreds or whole years. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_zero_axis}{Include zero. Choices: auto, on, off. Whether to extend the axis to include zero if the data values are all positive or all negative. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_mode}{Type. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title}{Text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_title_position}{Position. Choices: side, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_title_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_title_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_position}{Position. Choices: default, above, below. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_tick_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_angle}{Angle. Choices: 0, 30, 45, 60, 90. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_max_lines}{Max lines. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_space_mode}{Space mode. Choices: auto, max, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_space}{Space. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_mode}{Mode. Choices: auto, number, custom. Sets the method for selecting tick values on the axis:  Auto automatically selects the ticks  Number allows you to choose the number of ticks to display.  Custom lets you select specific ticks to show.  Note: If the axis is categorical, Number mode is ignored. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_number}{Number. Approximate number of ticks or gridlines. The actual number will depend on the range of values, chart size, etc. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_custom}{One tick label per line. Specify the ticks to show, one per line. For dates, use the input format as specified in the datasheet.  You can also use \code{{{FIRST}}} and \code{{{LAST}}} to add ticks at the minimum and maximum of the axis. For example:  \code{{{FIRST}}} 25000 50000 75000 \code{{{LAST}}}. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_y_axis_line_and_tick_color}{Line color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_line_and_tick_width}{Line width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_length}{Tick length. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_side}{Side. Choices: out, in. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_line_visible}{Axis line. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_edge_padding}{Edge padding (\%). Space between start/end of axis line and first/last category tick as a percentage of the spacing between ticks. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_gridlines_visible}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_gridlines_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_gridline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_gridline_style}{Style. Choices: solid, dash, dot, dot_dash. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_gridline_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_gridline_category_dividers}{Put lines between categories. On a categorical axis, this option disables the gridline on each tick and instead adds rules between each category to give a table-like design. Ignored for date and numeric axes. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_gridline_category_dividers_extend}{Extend. Extends the dividers into the axis margins. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_categorical_legend_show_legend}{Legend mode. Choices: \code{TRUE}, \code{FALSE}. A legend will not show with a single entry. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_categorical_legend_title_mode}{Title mode. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_categorical_legend_title}{Title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_radius}{Roundness. The radius of the corners of the swatch (in pixels). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_legend_items_padding}{Padding. Padding between legend items ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_outline}{Outline. An optional outline for the swatch in the legend. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_outline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_categorical_legend_order_override}{Custom order override. Manually specify the order of legend entries (one entry per line). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_categorical_legend_icon_height}{Height. Height of icon ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_icon_color}{Color. Fallback color (icon color if not determined by template). Flourish type hint: color, defaults to \code{NULL}}

\item{chart_categorical_legend_max_width}{Max width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_body_font}{Main font. This font will apply to the whole graphic by default. You can optionally change the font for the title, subtitle, footer, etc in the Header and Footer settings panels. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_font_color}{Text color. This color will apply to the whole graphic by default, You can optionally change the color for individual text elements, in other settings panels. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_color_enabled}{Color. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_image_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_background_image_size}{Size. Choices: cover, contain, auto, 100\% 100\%. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_background_image_position}{Position. Choices: top left, top center, top right, center left, center center, center right, bottom left, bottom center, bottom right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width_target}{Maximum width. Choices: none, wrapper, primary. Apply a maximum width to just the main graphic or everything (main graphic plus header, footer, etc). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width}{Maximum width. Leave blank to stretch to container width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_max_width_align}{Align. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_layout_order}{Layout order. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections}{Space between sections. Choices: 0.5, 1, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_enabled}{Show borders around visualisation. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_border_top_width}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_top_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_top_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_right_width}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_right_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_right_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_bottom_width}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_bottom_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_bottom_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_left_width}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_left_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_left_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_read_direction}{Read direction. Choices: ltr, rtl. This will change the reading direction of the main text elements on the page. Its not possible to adjust this on all elements, such as axes.  Note that when direction is set to right to left any alignment icons will be reversed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_align}{Alignment. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_title_font}{Title Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_title_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_title_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_subtitle_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_subtitle_font}{Subtitle Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_subtitle_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_subtitle_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_header_text_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_text_size}{Size. Choices: 1.2, 1.4, 1.6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_text_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_space}{Space. Space between border and header text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_enabled}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_header_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_align}{Align. Choices: inside, outside. Align logo inside either the header or the main visualization container. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_inside}{Position. Choices: top, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_outside}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_align}{Alignment. Choices: left, center, right, justify. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_styling}{Advanced footer styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_font}{Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_footer_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_multiple_sources}{Multiple sources. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_source_name_2}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_2}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name_3}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_3}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_label}{Source label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_note}{Note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_footer_note_secondary}{Note (secondary). The secondary note is placed below the source and primary note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_advanced_note_styling}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_timestamp_format}{Time/date stamp formatting. Choices: \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M, \%H:\%M \%p, \%H:\%M:\%S, \%d \%b \%Y, \%H:\%M, \%d \%b \%Y, \%H:\%M, \%d/\%m/\%Y, \%H:\%M, \%m/\%d/\%Y, \%H:\%M, \%d-\%m-\%Y, \%H:\%M, \%m-\%d-\%Y, \%d \%b \%Y, \%m/\%d/\%Y, \%d/\%m/\%Y, \%m-\%d-\%Y, \%d-\%m-\%Y, \%Y-\%m-\%d, \%Y-\%m-\%d \%H:\%M:\%S. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src_light}{Image (light version). If provided this version will be used whenever the background color is dark. Flourish type hint: hidden, defaults to \code{NULL}}

\item{chart_layout_footer_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_margin}{Margin. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_order}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_align_vertical}{V. align. Choices: flex-start, center, flex-end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_space}{Space. Space between border and footer text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_screenreader_text_primary}{Screenreader description. A text alternative to the visual content that will only be visible to screenreaders, e.g. The line chart shows China consistently higher than the other countries since 1990.  Do not replicate your title, since that will also be read by screenreaders. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_layout_screenreader_label}{Screenreader label. A short text label given to the main Flourish embed wrapper to provide an accessible name that is only visible to screenreaders Added in the form of an "aria-label". Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_screenreader_hide_primary}{Screenreader mode for main visual container. Choices: \code{TRUE}, \code{FALSE}. Whether the main visual container is visible to screenreaders (Text in the header and footer are always available to screenreaders.). Flourish type hint: boolean, defaults to \code{NULL}}
}
\value{
A Flourish chart
}
\description{
Set timeline chart details
}
\examples{
try(
  flourish(chart_type = "timeline", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
    bind_timeline_data(gapminder) |>
    set_timeline_details()
)
}
