% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_chord_diagram_data.R
\name{bind_chord_diagram_data}
\alias{bind_chord_diagram_data}
\title{Bind chord diagram data}
\usage{
bind_chord_diagram_data(
  .,
  connections_data = NULL,
  source = NULL,
  target = NULL,
  value = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{connections_data}{Connections data.}

\item{source}{Source}

\item{target}{Target}

\item{value}{Value. A number specifying the size of the flow. Flourish type hint: column}
}
\value{
A Flourish chart
}
\description{
Bind chord diagram data
}
\examples{
try(
  flourish(chart_type = "chord_diagram", api_key = Sys.getenv("FLOURISH_API_KEY")) |>
  bind_chord_diagram_data(gapminder)
)
}
