% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gantt_details.R
\name{set_gantt_details}
\alias{set_gantt_details}
\title{Set gantt chart chart details}
\usage{
set_gantt_details(
  .,
  gantt_height = NULL,
  gantt_padding = NULL,
  border_radius = NULL,
  border_width = NULL,
  border_color = NULL,
  gantt_progress = NULL,
  gantt_text = NULL,
  gantt_font_size = NULL,
  gantt_text_color = NULL,
  gantt_font_weight = NULL,
  missing_value_label = NULL,
  scale_type = NULL,
  chart_color_scale_categorical_palette = NULL,
  chart_color_scale_categorical_extend = NULL,
  chart_color_scale_categorical_custom_palette = NULL,
  chart_color_scale_numeric_type = NULL,
  chart_color_scale_binning = NULL,
  chart_color_scale_bin_mode = NULL,
  chart_color_scale_bin_count = NULL,
  chart_color_scale_bin_thresholds = NULL,
  chart_color_scale_sequential_palette = NULL,
  chart_color_scale_sequential_reverse = NULL,
  chart_color_scale_sequential_custom_min = NULL,
  chart_color_scale_sequential_custom_max = NULL,
  chart_color_scale_sequential_color_space = NULL,
  chart_color_scale_sequential_custom_domain = NULL,
  chart_color_scale_sequential_domain_min = NULL,
  chart_color_scale_sequential_domain_max = NULL,
  chart_color_scale_diverging_palette = NULL,
  chart_color_scale_diverging_reverse = NULL,
  chart_color_scale_diverging_custom_min = NULL,
  chart_color_scale_diverging_custom_mid = NULL,
  chart_color_scale_diverging_custom_max = NULL,
  chart_color_scale_diverging_color_space = NULL,
  chart_color_scale_diverging_custom_domain = NULL,
  chart_color_scale_diverging_domain_min = NULL,
  chart_color_scale_diverging_domain_mid = NULL,
  chart_color_scale_diverging_domain_max = NULL,
  color_single = NULL,
  chart_background_color = NULL,
  chart_background_opacity = NULL,
  chart_background_panels = NULL,
  chart_x_axis_axis_position = NULL,
  chart_x_axis_y_axis_position = NULL,
  chart_x_axis_numeric_scale_type = NULL,
  chart_x_axis_linear_min = NULL,
  chart_x_axis_linear_max = NULL,
  chart_x_axis_log_min = NULL,
  chart_x_axis_log_max = NULL,
  chart_x_axis_datetime_min = NULL,
  chart_x_axis_datetime_max = NULL,
  chart_x_axis_flip = NULL,
  chart_x_axis_show_scale_settings = NULL,
  chart_x_axis_nice = NULL,
  chart_x_axis_zero_axis = NULL,
  chart_x_axis_title_mode = NULL,
  chart_x_axis_title = NULL,
  chart_x_axis_title_styling = NULL,
  chart_x_axis_title_weight = NULL,
  chart_x_axis_title_color = NULL,
  chart_x_axis_title_size = NULL,
  chart_x_axis_title_padding = NULL,
  chart_x_axis_tick_label_position = NULL,
  chart_x_axis_tick_label_styling = NULL,
  chart_x_axis_tick_label_size = NULL,
  chart_x_axis_tick_label_color = NULL,
  chart_x_axis_tick_padding = NULL,
  chart_x_axis_tick_label_angle = NULL,
  chart_x_axis_tick_label_weight = NULL,
  chart_x_axis_tick_label_max_lines = NULL,
  chart_x_axis_tick_label_line_height = NULL,
  chart_x_axis_tick_label_space_mode = NULL,
  chart_x_axis_tick_label_space = NULL,
  chart_x_axis_tick_mode = NULL,
  chart_x_axis_tick_number = NULL,
  chart_x_axis_tick_custom = NULL,
  chart_x_axis_line_and_tick_color = NULL,
  chart_x_axis_line_and_tick_width = NULL,
  chart_x_axis_tick_length = NULL,
  chart_x_axis_tick_side = NULL,
  chart_x_axis_line_visible = NULL,
  chart_x_axis_edge_padding = NULL,
  chart_x_axis_gridlines_visible = NULL,
  chart_x_axis_gridlines_styling = NULL,
  chart_x_axis_gridline_color = NULL,
  chart_x_axis_gridline_style = NULL,
  chart_x_axis_gridline_width = NULL,
  chart_x_axis_gridline_category_dividers = NULL,
  chart_x_axis_gridline_category_dividers_extend = NULL,
  date_format = NULL,
  format_locale = NULL,
  format_weekday = NULL,
  format_year = NULL,
  format_month = NULL,
  format_day = NULL,
  format_hour = NULL,
  format_minute = NULL,
  y_font_size = NULL,
  y_font_weight = NULL,
  y_font_fill = NULL,
  y_max_width = NULL,
  y_text_dx = NULL,
  y_text_dy = NULL,
  y_gridline_color = NULL,
  y_gridline_width = NULL,
  chart_filter_controls_control_type = NULL,
  chart_filter_controls_control_styles = NULL,
  chart_filter_controls_control_title = NULL,
  chart_filter_controls_button_group = NULL,
  chart_filter_controls_button_group_width_mode = NULL,
  chart_filter_controls_button_group_width_fixed = NULL,
  chart_filter_controls_dropdown_width_mode = NULL,
  chart_filter_controls_dropdown_width_fixed = NULL,
  chart_filter_controls_slider_width = NULL,
  chart_filter_controls_slider_max_label_width = NULL,
  chart_filter_controls_slider_step_time = NULL,
  chart_filter_controls_slider_loop = NULL,
  chart_filter_controls_slider_restart_pause = NULL,
  chart_filter_controls_sort = NULL,
  filter_control_all = NULL,
  chart_controls_style_font_size = NULL,
  chart_controls_style_font_weight = NULL,
  chart_controls_style_height = NULL,
  chart_button_control_style_background = NULL,
  chart_button_control_style_background_selected = NULL,
  chart_button_control_style_background_hover = NULL,
  chart_button_control_style_font_color = NULL,
  chart_button_control_style_font_color_selected = NULL,
  chart_button_control_style_font_color_hover = NULL,
  chart_button_control_style_button_styles_advanced = NULL,
  chart_button_control_style_border_width = NULL,
  chart_button_control_style_border_color = NULL,
  chart_button_control_style_border_transparency = NULL,
  chart_button_control_style_border_radius = NULL,
  chart_dropdown_control_style_background = NULL,
  chart_dropdown_control_style_font_color = NULL,
  chart_dropdown_control_style_border_styles_advanced = NULL,
  chart_dropdown_control_style_border_style = NULL,
  chart_dropdown_control_style_border_width = NULL,
  chart_dropdown_control_style_border_color = NULL,
  chart_dropdown_control_style_border_transparency = NULL,
  chart_dropdown_control_style_border_radius = NULL,
  chart_slider_control_style_track_color = NULL,
  chart_slider_control_style_font_color = NULL,
  chart_slider_control_style_handle_color = NULL,
  chart_slider_control_style_play_color = NULL,
  chart_slider_control_style_track_height = NULL,
  chart_slider_control_style_handle_height = NULL,
  chart_slider_control_style_play_button = NULL,
  chart_filter_controls_container_alignment = NULL,
  chart_filter_controls_container_controls_spacing = NULL,
  chart_categorical_legend_show_legend = NULL,
  chart_categorical_legend_title_mode = NULL,
  chart_categorical_legend_title = NULL,
  chart_categorical_legend_swatch_width = NULL,
  chart_categorical_legend_swatch_height = NULL,
  chart_categorical_legend_swatch_radius = NULL,
  chart_categorical_legend_legend_items_padding = NULL,
  chart_categorical_legend_swatch_outline = NULL,
  chart_categorical_legend_swatch_outline_color = NULL,
  chart_categorical_legend_order_override = NULL,
  chart_categorical_legend_icon_height = NULL,
  chart_categorical_legend_icon_color = NULL,
  chart_categorical_legend_max_width = NULL,
  chart_categorical_legend_orientation = NULL,
  chart_continous_legend_show_legend = NULL,
  chart_continous_legend_title_mode = NULL,
  chart_continous_legend_title = NULL,
  chart_continous_legend_color_band_width = NULL,
  chart_continous_legend_color_band_height = NULL,
  chart_continous_legend_color_band_radius = NULL,
  chart_continous_legend_binned_label_mode = NULL,
  chart_continous_legend_binned_label_custom = NULL,
  chart_continous_legend_legend_min = NULL,
  chart_continous_legend_additional_min_label = NULL,
  chart_continous_legend_legend_max = NULL,
  chart_continous_legend_additional_max_label = NULL,
  chart_legend_container_alignment = NULL,
  chart_legend_container_orientation = NULL,
  chart_legend_container_title_weight = NULL,
  chart_legend_container_text_weight = NULL,
  chart_legend_container_text_color = NULL,
  chart_legend_container_text_size = NULL,
  chart_popup_mode = NULL,
  chart_popup_popup_is_custom = NULL,
  chart_popup_show_popup_styles = NULL,
  chart_popup_popup_custom_header = NULL,
  chart_popup_popup_custom_main = NULL,
  chart_popup_popup_custom_accessible = NULL,
  chart_popup_popup_width = NULL,
  chart_popup_popup_font_size = NULL,
  chart_popup_popup_text_color = NULL,
  chart_popup_popup_align = NULL,
  chart_popup_popup_background = NULL,
  chart_popup_popup_background_opacity = NULL,
  chart_popup_popup_padding = NULL,
  chart_popup_popup_radius = NULL,
  chart_popup_popup_header_type = NULL,
  chart_popup_popup_header_background = NULL,
  chart_popup_popup_header_text_color = NULL,
  chart_popup_popup_list_type = NULL,
  chart_popup_popup_list_label_weight = NULL,
  chart_popup_popup_list_separator = NULL,
  chart_popup_popup_shadow = NULL,
  chart_popup_popup_pointer = NULL,
  chart_popup_panel_position = NULL,
  chart_popup_panel_open_duration = NULL,
  chart_popup_panel_side_fixed = NULL,
  chart_popup_panel_side_default_content = NULL,
  chart_popup_panel_is_custom = NULL,
  chart_popup_show_panel_styles = NULL,
  chart_popup_panel_custom_header = NULL,
  chart_popup_panel_custom_main = NULL,
  chart_popup_panel_width_overlay = NULL,
  chart_popup_panel_height_overlay = NULL,
  chart_popup_panel_width_side = NULL,
  chart_popup_panel_height_side = NULL,
  chart_popup_panel_max_width = NULL,
  chart_popup_panel_max_height = NULL,
  chart_popup_panel_vertical_position = NULL,
  chart_popup_panel_text_color = NULL,
  chart_popup_panel_font_size = NULL,
  chart_popup_panel_align = NULL,
  chart_popup_panel_background = NULL,
  chart_popup_panel_background_opacity = NULL,
  chart_popup_panel_padding = NULL,
  chart_popup_panel_radius = NULL,
  chart_popup_panel_list_type = NULL,
  chart_popup_panel_list_label_weight = NULL,
  chart_popup_panel_list_separator = NULL,
  chart_popup_panel_shadow = NULL,
  chart_axes_highlights_x_enabled = NULL,
  chart_axes_highlights_x_lines = NULL,
  chart_axes_highlights_x_line_color = NULL,
  chart_axes_highlights_x_line_width = NULL,
  chart_axes_highlights_x_line_dash = NULL,
  chart_axes_highlights_x_areas = NULL,
  chart_axes_highlights_x_area_color = NULL,
  chart_axes_highlights_x_area_opacity = NULL,
  chart_axes_highlights_x_area_texture = NULL,
  chart_axes_highlights_x_label_vertical = NULL,
  chart_axes_highlights_x_label_y_position = NULL,
  chart_axes_highlights_x_label_x_position = NULL,
  chart_axes_highlights_x_label_max_width = NULL,
  chart_axes_highlights_x_font_size = NULL,
  chart_axes_highlights_x_font_weight = NULL,
  chart_axes_highlights_x_label_color = NULL,
  chart_axes_highlights_x_shadow_color = NULL,
  chart_axes_highlights_x_shadow_enabled = NULL,
  chart_axes_highlights_y_enabled = NULL,
  chart_axes_highlights_y_lines = NULL,
  chart_axes_highlights_y_line_color = NULL,
  chart_axes_highlights_y_line_width = NULL,
  chart_axes_highlights_y_line_dash = NULL,
  chart_axes_highlights_y_areas = NULL,
  chart_axes_highlights_y_area_color = NULL,
  chart_axes_highlights_y_area_opacity = NULL,
  chart_axes_highlights_y_area_texture = NULL,
  chart_axes_highlights_y_label_x_position = NULL,
  chart_axes_highlights_y_label_y_position = NULL,
  chart_axes_highlights_y_label_max_width = NULL,
  chart_axes_highlights_y_font_size = NULL,
  chart_axes_highlights_y_font_weight = NULL,
  chart_axes_highlights_y_label_color = NULL,
  chart_axes_highlights_y_shadow_color = NULL,
  chart_axes_highlights_y_shadow_enabled = NULL,
  animation_duration = NULL,
  chart_layout_body_font = NULL,
  chart_layout_font_color = NULL,
  chart_layout_background_color_enabled = NULL,
  chart_layout_background_image_enabled = NULL,
  chart_layout_background_color = NULL,
  chart_layout_background_image_src = NULL,
  chart_layout_background_image_size = NULL,
  chart_layout_background_image_position = NULL,
  chart_layout_max_width_target = NULL,
  chart_layout_max_width = NULL,
  chart_layout_max_width_align = NULL,
  chart_layout_layout_order = NULL,
  chart_layout_space_between_sections = NULL,
  chart_layout_space_between_sections_custom = NULL,
  chart_layout_margin_top = NULL,
  chart_layout_margin_right = NULL,
  chart_layout_margin_bottom = NULL,
  chart_layout_margin_left = NULL,
  chart_layout_border_enabled = NULL,
  chart_layout_border_top_width = NULL,
  chart_layout_border_top_style = NULL,
  chart_layout_border_top_color = NULL,
  chart_layout_border_right_width = NULL,
  chart_layout_border_right_style = NULL,
  chart_layout_border_right_color = NULL,
  chart_layout_border_bottom_width = NULL,
  chart_layout_border_bottom_style = NULL,
  chart_layout_border_bottom_color = NULL,
  chart_layout_border_left_width = NULL,
  chart_layout_border_left_style = NULL,
  chart_layout_border_left_color = NULL,
  chart_layout_read_direction = NULL,
  chart_layout_font_size_mobile_small = NULL,
  chart_layout_font_size_mobile_big = NULL,
  chart_layout_font_size_tablet = NULL,
  chart_layout_font_size_desktop = NULL,
  chart_layout_font_size_big_screen = NULL,
  chart_layout_breakpoint_mobile_small = NULL,
  chart_layout_breakpoint_mobile_big = NULL,
  chart_layout_breakpoint_tablet = NULL,
  chart_layout_breakpoint_desktop = NULL,
  chart_layout_breakpoint_big_screen = NULL,
  chart_layout_header_align = NULL,
  chart_layout_title = NULL,
  chart_layout_title_styling = NULL,
  chart_layout_title_font = NULL,
  chart_layout_title_size = NULL,
  chart_layout_title_size_custom = NULL,
  chart_layout_title_weight = NULL,
  chart_layout_title_color = NULL,
  chart_layout_title_line_height = NULL,
  chart_layout_title_space_above = NULL,
  chart_layout_title_space_above_custom = NULL,
  chart_layout_subtitle = NULL,
  chart_layout_subtitle_styling = NULL,
  chart_layout_subtitle_font = NULL,
  chart_layout_subtitle_size = NULL,
  chart_layout_subtitle_size_custom = NULL,
  chart_layout_subtitle_weight = NULL,
  chart_layout_subtitle_color = NULL,
  chart_layout_subtitle_line_height = NULL,
  chart_layout_subtitle_space_above = NULL,
  chart_layout_subtitle_space_above_custom = NULL,
  chart_layout_header_text = NULL,
  chart_layout_header_text_styling = NULL,
  chart_layout_header_text_size = NULL,
  chart_layout_header_text_size_custom = NULL,
  chart_layout_header_text_weight = NULL,
  chart_layout_header_text_color = NULL,
  chart_layout_header_text_line_height = NULL,
  chart_layout_header_text_space_above = NULL,
  chart_layout_header_text_space_above_custom = NULL,
  chart_layout_header_border = NULL,
  chart_layout_header_border_width = NULL,
  chart_layout_header_border_color = NULL,
  chart_layout_header_border_style = NULL,
  chart_layout_header_border_space = NULL,
  chart_layout_header_logo_enabled = NULL,
  chart_layout_header_logo_src = NULL,
  chart_layout_header_logo_alt = NULL,
  chart_layout_header_logo_link_url = NULL,
  chart_layout_header_logo_height = NULL,
  chart_layout_header_logo_align = NULL,
  chart_layout_header_logo_position_inside = NULL,
  chart_layout_header_logo_position_outside = NULL,
  chart_layout_header_logo_margin_top = NULL,
  chart_layout_header_logo_margin_right = NULL,
  chart_layout_header_logo_margin_bottom = NULL,
  chart_layout_header_logo_margin_left = NULL,
  chart_layout_footer_align = NULL,
  chart_layout_footer_text_size = NULL,
  chart_layout_footer_text_color = NULL,
  chart_layout_footer_styling = NULL,
  chart_layout_footer_font = NULL,
  chart_layout_footer_text_weight = NULL,
  chart_layout_source_name = NULL,
  chart_layout_source_url = NULL,
  chart_layout_multiple_sources = NULL,
  chart_layout_source_name_2 = NULL,
  chart_layout_source_url_2 = NULL,
  chart_layout_source_name_3 = NULL,
  chart_layout_source_url_3 = NULL,
  chart_layout_source_label = NULL,
  chart_layout_footer_note = NULL,
  chart_layout_footer_note_secondary = NULL,
  chart_layout_advanced_note_styling = NULL,
  chart_layout_footer_timestamp_format = NULL,
  chart_layout_footer_logo_enabled = NULL,
  chart_layout_footer_logo_src = NULL,
  chart_layout_footer_logo_src_light = NULL,
  chart_layout_footer_logo_alt = NULL,
  chart_layout_footer_logo_link_url = NULL,
  chart_layout_footer_logo_height = NULL,
  chart_layout_footer_logo_margin = NULL,
  chart_layout_footer_logo_order = NULL,
  chart_layout_footer_align_vertical = NULL,
  chart_layout_footer_border = NULL,
  chart_layout_footer_border_width = NULL,
  chart_layout_footer_border_color = NULL,
  chart_layout_footer_border_style = NULL,
  chart_layout_footer_border_space = NULL,
  chart_layout_screenreader_text_primary = NULL,
  chart_layout_screenreader_label = NULL,
  chart_layout_screenreader_hide_primary = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{gantt_height}{Height. Choose the height of each bar in rems (i.e. it will scale up and down in relation to screen size). Flourish type hint: number, defaults to \code{NULL}}

\item{gantt_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{border_radius}{Roundness. Flourish type hint: number, defaults to \code{NULL}}

\item{border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{border_color}{Border color. Flourish type hint: color, defaults to \code{NULL}}

\item{gantt_progress}{Show progress. Choices: \code{TRUE}, \code{FALSE}. If enabled, each bar will be filled with an overlay relating to the value supplied in the Progress column. Flourish type hint: boolean, defaults to \code{NULL}}

\item{gantt_text}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{gantt_font_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{gantt_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{gantt_font_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{missing_value_label}{Label for missing values. If a bar dosent have a category, the legend will show the placeholder value.. Flourish type hint: string, defaults to \code{NULL}}

\item{scale_type}{Scale type. Choices: categorical, numeric. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_categorical_palette}{Palette. Flourish type hint: colors, defaults to \code{NULL}}

\item{chart_color_scale_categorical_extend}{Extend. Automatically generate additional colours when needed to avoid the palette colours being used more than once. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_categorical_custom_palette}{Custom overrides. Enter the label name for which you wish to set the color, followed by a colon and the desired color value.  Colors can be set using Hex, RGB, color names or RGBA, if you want to set the opacity Multiple colors can be set using multiple lines. For example:  Party 1: red Party 2: #4455AA Party 3: rgb(30,168,26). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_color_scale_numeric_type}{Scale type. Choices: sequential, diverging. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_binning}{Choices: \code{FALSE}, \code{TRUE}. In linear mode, the color scale will run as a smooth gradient between 2 colors. In binned mode, the gradient will be divided in smaller blocks. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_bin_mode}{Binning mode. Choices: fixed, quantile, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_bin_count}{Number of bins. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_bin_thresholds}{Custom thresholds. Enter your desired thresholds, separating them with a ";". For instance, "5;10;15".  Bins form as follows:  From the datas minimum value (domain minimum) up to the first threshold.  Between consecutive thresholds.  From the last threshold to the datas maximum value (domain maximum).   For "5;10;15", youll get four bins based on your datas range. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_mid}{Midpoint color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_mid}{Mid. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{color_single}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_background_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_background_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_background_panels}{Area. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_axis_position}{Position. Choices: bottom, float-below, float-above, top, off. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_y_axis_position}{Y value. Vertical position of the X axis on the Y scale. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_numeric_scale_type}{Type. Choices: linear, log. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_linear_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_linear_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_log_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_log_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_datetime_min}{Min (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_datetime_max}{Max (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_flip}{Flip axis. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_show_scale_settings}{Configure default min/max. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_nice}{Round min/max. Choices: \code{TRUE}, \code{FALSE}. Enabling this option rounds the X axis start and end values to tidy numbers, such as 0, 50, 100 or whole years. Specified min and max values will override this feature. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_zero_axis}{Include zero. Choices: auto, on, off. Enable to extend the axis to include zero, even when data is all positive or all negative. This is ignored if you set specific min and max scale values. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_mode}{Type. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title}{Text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_title_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_title_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_position}{Position. Choices: default, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_tick_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_angle}{Angle. Choices: 0, 30, 45, 60, 90. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_max_lines}{Max lines. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_space_mode}{Space mode. Choices: auto, max, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_space}{Space. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_mode}{Mode. Choices: auto, number, custom. Sets the method for selecting tick values on the axis:  Auto automatically selects the ticks  Number allows you to choose the number of ticks to display.  Custom lets you select specific ticks to show.  Note: If the axis is categorical, Number mode is ignored. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_number}{Number. Approximate number of ticks or gridlines. The actual number will depend on the range of values, chart size, etc. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_custom}{One tick label per line. Specify the ticks to show, one per line. For dates, use the input format as specified in the datasheet.  You can also use \code{{{FIRST}}} and \code{{{LAST}}} to add ticks at the minimum and maximum of the axis. For example:  \code{{{FIRST}}} 25000 50000 75000 \code{{{LAST}}}. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_x_axis_line_and_tick_color}{Line color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_line_and_tick_width}{Line width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_length}{Tick length. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_side}{Tick side. Choices: out, in. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_line_visible}{Axis line. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_edge_padding}{Edge padding (\%). Space between start/end of axis line and first/last category tick as a percentage of the spacing between ticks. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_gridlines_visible}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridlines_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_gridline_style}{Style. Choices: solid, dash, dot, dot_dash. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_gridline_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_gridline_category_dividers}{Put lines between categories. On a categorical axis, this option disables the gridline on each tick and instead adds rules between each category to give a table-like design. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridline_category_dividers_extend}{Extend. Extends the dividers into the axis margins. Flourish type hint: boolean, defaults to \code{NULL}}

\item{date_format}{Format in data sheet. Choices: auto, \%Y-\%m-\%dT\%H:\%M:\%S.\%LZ, \%Y-\%m-\%d, \%m/\%d/\%Y, \%d/\%m/\%Y, \%d-\%b-\%y, \%m/\%Y, \%b \%Y, \%b \%y, \%B \%d, \%d \%b, \%Y, \%B, \%b, \%A, \%a, \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M. The date/time format in your datasheet.  If it isnt in the list, you can enter a custom format using d3-time-format syntax. See npm - d3 time format for details. Flourish type hint: string, defaults to \code{NULL}}

\item{format_locale}{Date and time locale. Choices: none, ady, af-_n_a, af-_z_a, af, ak, ar-_m_a, ar-_s_a, hy-_a_m, ay-_b_o, eu-_e_s, be-_b_y, bn-_b_d, bn-_i_n, bn, bs-_b_a, my, ck-_u_s, zh-_c_n, zh-_s_g, zh-_hans, zh-_h_k, zh-_t_w, zh-_hant, zh, gx-_g_r, nl-_b_e, nl-_n_l, nl, en-_a_u, en-_c_a, en-_i_n, en-_i_e, en@pirate, en-_z_a, en-_g_b, en-_u_d, en-_u_s, en, eo-_e_o, et-_e_e, fo-_f_o, fi-_f_i, fr-_b_e, fr-_c_a, fr-_f_r, fr-_c_h, fr, fy-_n_l, ff, gl-_e_s, ka-_g_e, de-_a_t, de-_d_e, de-_c_h, de, el-_g_r, el, gn-_p_y, gu-_i_n, ht, is-_i_s, is, ga-_i_e, ga, jv-_i_d, kab, kn-_i_n, kn, cak, kk-_k_z, tlh, ko-_k_r, ko, ku-_t_r, fb-_l_t, li-_n_l, dsb, lb, mk-_m_k, mk, mai, mg-_m_g, mn-_m_n, se-_n_o, nn-_n_o, no, oc, or-_i_n, ps-_a_f, i_r, pt-_b_r, pt-_p_t, pt, pa-_i_n, pa, qu-_p_e, rm-_c_h, sa-_i_n, sr-_r_s, sr, si-_l_k, sk-_s_k, sk, so-_s_o, es-_a_r, es-_c_l, es-_c_o, es-_e_c, es-_m_x, es-_n_i, es-_e_s, es-_u_s, es-_v_e, es, su, sw-_k_e, sw, tt-_r_u, th-_t_h, th, hsb, xh-_z_a, yi-_d_e, yi, zu-_z_a. The regional date/time format ( date locale ). You can specify each part of the format in the fields below. Flourish type hint: string, defaults to \code{NULL}}

\item{format_weekday}{Weekday. Choices: none, narrow, short, long. Flourish type hint: string, defaults to \code{NULL}}

\item{format_year}{Year. Choices: none, numeric, 2-digit. Flourish type hint: string, defaults to \code{NULL}}

\item{format_month}{Month. Choices: none, numeric, 2-digit, narrow, short, long. Flourish type hint: string, defaults to \code{NULL}}

\item{format_day}{Day. Choices: none, numeric, 2-digit. Flourish type hint: string, defaults to \code{NULL}}

\item{format_hour}{Hour. Choices: none, numeric, 2-digit. Flourish type hint: string, defaults to \code{NULL}}

\item{format_minute}{Minute. Choices: none, numeric, 2-digit. Flourish type hint: string, defaults to \code{NULL}}

\item{y_font_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{y_font_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{y_font_fill}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{y_max_width}{Max width. Flourish type hint: number, defaults to \code{NULL}}

\item{y_text_dx}{Horizontal padding. Flourish type hint: number, defaults to \code{NULL}}

\item{y_text_dy}{Vertical padding. Flourish type hint: number, defaults to \code{NULL}}

\item{y_gridline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{y_gridline_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_control_type}{Control type. Choices: dropdown, buttons, slider. Choose between a dropdown, buttons or a  time slider Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_control_styles}{Adv. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_control_title}{Control title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_button_group}{Grouped. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_button_group_width_mode}{Button group width. Choices: auto, full, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_button_group_width_fixed}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_dropdown_width_mode}{Dropdown width. Choices: auto, full, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_dropdown_width_fixed}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_width}{Slider width. Width of the slider track in rems Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_max_label_width}{Max label width. In rems, multiples of base font size. If larger than the longest label, only the space required to fit labels will be used. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_step_time}{Time between steps. Measured in seconds, positive values move the slider left to right, negative values move the slider right to left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_loop}{Loop. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_slider_restart_pause}{Pause before loop. Measured in seconds and in addition to the regular step time displayed above. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_sort}{Sorting. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{filter_control_all}{Name for "All". Leave blank to not include an "All" option. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_controls_style_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_controls_style_font_weight}{Text weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_controls_style_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_background_selected}{Selected. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_background_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color_selected}{Selected. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_button_styles_advanced}{Button border styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_button_control_style_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_border_transparency}{Transparency. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_font_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_styles_advanced}{Dropdown border styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_style}{Border style. Choices: all, bottom. Show border on all sides, or only at the bottom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_transparency}{Transparency. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_control_style_track_color}{Track. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_font_color}{Text. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_handle_color}{Handle. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_play_color}{Play/pause. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_slider_control_style_track_height}{Track height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_control_style_handle_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_slider_control_style_play_button}{Play button. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_container_alignment}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_container_controls_spacing}{Spacing. The space between controls, like dropdowns and adjacent button sets, in em units. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_show_legend}{Legend mode. Choices: \code{TRUE}, \code{FALSE}. A legend will not show with a single entry. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_categorical_legend_title_mode}{Title mode. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_categorical_legend_title}{Title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_radius}{Roundness. The radius of the corners of the swatch (in pixels). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_legend_items_padding}{Padding. Padding between legend items ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_outline}{Outline. An optional outline for the swatch in the legend. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_categorical_legend_swatch_outline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_categorical_legend_order_override}{Custom order override. Manually specify the order of legend entries (one entry per line). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_categorical_legend_icon_height}{Height. Height of icon ( in rems ). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_icon_color}{Color. Fallback color (icon color if not determined by template). Flourish type hint: color, defaults to \code{NULL}}

\item{chart_categorical_legend_max_width}{Max width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_categorical_legend_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_continous_legend_show_legend}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_continous_legend_title_mode}{Title mode. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_continous_legend_title}{Title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_continous_legend_color_band_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_continous_legend_color_band_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_continous_legend_color_band_radius}{Roundness. The radius of the corners of the color band (in pixels). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_continous_legend_binned_label_mode}{Binned label mode. Choices: thresholds, centers, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_continous_legend_binned_label_custom}{Custom bin labels. List numeric values, separated by semi-colons. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_continous_legend_legend_min}{Legend min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_continous_legend_additional_min_label}{Custom min label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_continous_legend_legend_max}{Legend max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_continous_legend_additional_max_label}{Custom max label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_alignment}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_orientation}{Orientation. Choices: horizontal, vertical. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_title_weight}{Title weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_text_weight}{Text weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_legend_container_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_legend_container_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_mode}{Choices: none, popup, panel, both. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_is_custom}{Popup content. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_show_popup_styles}{Show popup styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_popup_custom_header}{Popup custom header. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for popup use.  Advanced users can add HTML styling or add category colors to popup headers.  To hide parts of the popup when a column is empty using \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_popup_custom_main}{Popup custom main content. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for popup use.  Advanced users can add HTML for designs, styles, and images.  To hide parts of the popup when a column is empty use the syntax \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_popup_custom_accessible}{Popup custom accessible content. A text alternative to visual custom popup content that will be read out by screenreaders. This content is added as an aria-label attribute to data point elements. You can still use \code{{{column_name}}} to add values from your data, but do not use HTML. For example: \verb{\{\{Name\}\}, Population: \{\{Population\}\}, GDP: \{\{GDP\}\}}. If left empty, accessible content will fallback to the default popup content. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_popup_popup_width}{Max width. In rems, multiples of base font size. Leave blank to let the width adjust based on the contents. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_text_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_align}{Align. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_background}{Fill color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_background_opacity}{Fill opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_radius}{Radius. Corner radius of popup. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_popup_header_type}{Header style. Choices: block, inline. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_header_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_header_text_color}{Text. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_popup_list_type}{List type. Choices: table, ul. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_list_label_weight}{Label weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_list_separator}{List separator. Choices: :, , , \verb{<br />}. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_popup_shadow}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_popup_pointer}{Pointer. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_panel_position}{Choices: overlay, left, right, top, bottom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_open_duration}{Open duration. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_side_fixed}{Always show. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_panel_side_default_content}{Placeholder content. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_popup_panel_is_custom}{Panel content. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_show_panel_styles}{Show panel styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_popup_panel_custom_header}{Panel custom header. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for panel use.  Advanced users can add HTML for styling or add category colors to panel headers.  To hide parts of the panel when a column is empty using \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_panel_custom_main}{Panel custom main content. Use \code{{{column_name}}} to add a value from your data Ensure the column is selected, or add it to Info for popups for panel use.  Advanced users can add HTML for designs, styles, images, videos and embedded Flourish charts,  To hide parts of the panel when a column is empty using \verb{\{\{#if column_name\}\}Content for non-empty column_name\{\{/if\}\}}. For example: \verb{\{\{#if Capital name\}\}Capital:\{\{/if\}\} \{\{Capital name\}\}}. Here, the word "Capital:" will only show if the corresponding cell in the "Capital name" column is not empty. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_popup_panel_width_overlay}{Width (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_height_overlay}{Height (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_width_side}{Width (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_height_side}{Height (\%). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_max_width}{Max width (px). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_max_height}{Max height (px). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_vertical_position}{Vertical position. Choices: click, center. Choosing Click position is especially useful when your visualization is really tall, this will make sure the panel content is always on top of the element youve clicked on. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_text_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_panel_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_align}{Alignment. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_background}{Fill color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_popup_panel_background_opacity}{Fill opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_radius}{Radius. Corner radius of panel. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_popup_panel_list_type}{List type. Choices: table, ul. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_list_label_weight}{Label weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_list_separator}{List separator. Choices: :, , , \verb{<br />}. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_popup_panel_shadow}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_axes_highlights_x_enabled}{Show highlights on the x axis. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_axes_highlights_x_lines}{One per line, in format My label :: 2012. One per line, using the  format :  My label :: 2012  For dates, use the  input format  as specified in the datasheet. You can optionally specify a line color and a text color For visualizations using a  grid of charts , you can apply the annotation only on a specific chart. Use the format:  \verb{My label :: 2012 :: line_color / text_color :: facet_name}.  For example,  \verb{My label :: 2012 :: #eb0000 / #0084bd :: Category 1}. If you want to add an annotation to just one specific chart but do not want to color the line/text, you can leave that part blank. For example:  My label :: 2012 :: :: Category 1 Flourish type hint: text, defaults to \code{NULL}}

\item{chart_axes_highlights_x_line_color}{Line color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_axes_highlights_x_line_width}{Line width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_axes_highlights_x_line_dash}{Dash. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_axes_highlights_x_areas}{One per line, in format My label :: 2013 >> 2015. One per line, using the format :  My label :: date >> date  For dates, use the input format as specified in the datasheet. You can optionally specify a line color and a text color For visualizations using a  grid of charts , you can apply the annotation only on a specific chart Use the format:  My label :: 2012 :: area_color / text_color :: facet_name For example,  \verb{My label :: 2012 >> 2015 :: #eb0000 / #0084bd :: Category 1}. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_axes_highlights_x_area_color}{Area. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_axes_highlights_x_area_opacity}{Area opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_axes_highlights_x_area_texture}{Area texture. Choices: none, pattern-diagonal, pattern-grid, pattern-dots. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_axes_highlights_x_label_vertical}{Vertical. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_axes_highlights_x_label_y_position}{Y position. Choices: bottom, middle, top. Vertically position the label at the top, bottom or in the middle of the line. This setting will be ignored if there is not enough space for the selected position. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_axes_highlights_x_label_x_position}{X position. Choices: left, middle, right. Horizontally position the label on the left, the right or in the middle of the line. This setting will be ignored if there is not enough space for the selected position. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_axes_highlights_x_label_max_width}{Max width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_axes_highlights_x_font_size}{Font size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_axes_highlights_x_font_weight}{Font weight. Choices: regular, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_axes_highlights_x_label_color}{Text. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_axes_highlights_x_shadow_color}{Shadow. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_axes_highlights_x_shadow_enabled}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_axes_highlights_y_enabled}{Show highlights on the y axis. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_axes_highlights_y_lines}{One per line, in format My label :: 5000. One per line, using the format :  My label :: 2012  For dates, use the input format as specified in the datasheet. You can optionally specify a line color and a text color For visualizations using a  grid of charts , you can apply the annotation only on a specific chart Use the format:  \verb{My label :: 2012 :: line_color / text_color :: facet_name}.  For example:  \verb{My label :: 2012 :: #eb0000 / #0084bd :: Category 1}.  If you want to add an annotation to just one specific chart but do not want to color the line/text, you can leave that part blank. For example:  My label :: 2012 :: :: Category 1 Flourish type hint: text, defaults to \code{NULL}}

\item{chart_axes_highlights_y_line_color}{Line color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_axes_highlights_y_line_width}{line width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_axes_highlights_y_line_dash}{Dash. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_axes_highlights_y_areas}{One per line, in format My label :: 2000 >> 8000. One per line, using the format :  My label :: date >> date  For dates, use the input format as specified in the datasheet. You can optionally specify a line color and a text color For visualizations using a  grid of charts , you can apply the annotation only on a specific chart Use the format:  My label :: 2012 :: area_color / text_color :: facet_name  For example:  \verb{My label :: 2012 >> 2015 :: #eb0000 / #0084bd :: Category 1}. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_axes_highlights_y_area_color}{Area. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_axes_highlights_y_area_opacity}{Area opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_axes_highlights_y_area_texture}{Area texture. Choices: none, pattern-diagonal, pattern-grid, pattern-dots. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_axes_highlights_y_label_x_position}{X position. Choices: left, middle, right. Horizontally position label on the left, the middle or the right end of the line. This setting will be ignored if there is not enough space for the selected position. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_axes_highlights_y_label_y_position}{Y position. Choices: above, middle, below. Vertically position the label above, below or in the middle of the line. This setting will be ignored if there is not enough space for the selected position. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_axes_highlights_y_label_max_width}{Max width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_axes_highlights_y_font_size}{Font size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_axes_highlights_y_font_weight}{Font weight. Choices: regular, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_axes_highlights_y_label_color}{Text. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_axes_highlights_y_shadow_color}{Shadow. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_axes_highlights_y_shadow_enabled}{Shadow. Flourish type hint: boolean, defaults to \code{NULL}}

\item{animation_duration}{Animation duration. The duration, in seconds, of transitions  for example, on load of visualization or between two slides in a story Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_body_font}{Main font. This font will apply to the whole graphic by default. You can optionally change the font for the title, subtitle, footer, etc in the Header and Footer settings panels. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_font_color}{Text color. This color will apply to the whole graphic by default, You can optionally change the color for individual text elements, in other settings panels. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_color_enabled}{Color. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_image_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_background_image_size}{Size. Choices: cover, contain, auto, 100\% 100\%. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_background_image_position}{Position. Choices: top left, top center, top right, center left, center center, center right, bottom left, bottom center, bottom right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width_target}{Maximum width. Choices: none, wrapper, primary. Apply a maximum width to just the main graphic or everything (main graphic plus header, footer, etc). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width}{Maximum width. Leave blank to stretch to container width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_max_width_align}{Align. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_layout_order}{Layout order. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections}{Space between sections. Choices: 0.5, 1, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_enabled}{Show borders around visualisation. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_border_top_width}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_top_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_top_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_right_width}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_right_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_right_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_bottom_width}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_bottom_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_bottom_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_left_width}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_left_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_left_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_read_direction}{Read direction. Choices: ltr, rtl. This will change the reading direction of the main text elements on the page. Its not possible to adjust this on all elements, such as axes.  Note that when direction is set to right to left any alignment icons will be reversed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_align}{Alignment. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_title_font}{Title Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_title_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_title_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_subtitle_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_subtitle_font}{Subtitle Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_subtitle_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_subtitle_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_header_text_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_text_size}{Size. Choices: 1.2, 1.4, 1.6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_text_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_space}{Space. Space between border and header text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_enabled}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_header_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_align}{Align. Choices: inside, outside. Align logo inside either the header or the main visualization container. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_inside}{Position. Choices: top, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_outside}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_align}{Alignment. Choices: left, center, right, justify. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_styling}{Advanced footer styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_font}{Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_footer_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_multiple_sources}{Multiple sources. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_source_name_2}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_2}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name_3}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_3}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_label}{Source label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_note}{Note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_footer_note_secondary}{Note (secondary). The secondary note is placed below the source and primary note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_advanced_note_styling}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_timestamp_format}{Time/date stamp formatting. Choices: \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M, \%H:\%M \%p, \%H:\%M:\%S, \%d \%b \%Y, \%H:\%M, \%d \%b \%Y, \%H:\%M, \%d/\%m/\%Y, \%H:\%M, \%m/\%d/\%Y, \%H:\%M, \%d-\%m-\%Y, \%H:\%M, \%m-\%d-\%Y, \%d \%b \%Y, \%m/\%d/\%Y, \%d/\%m/\%Y, \%m-\%d-\%Y, \%d-\%m-\%Y, \%Y-\%m-\%d, \%Y-\%m-\%d \%H:\%M:\%S. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src_light}{Image (light version). If provided this version will be used whenever the background color is dark. Flourish type hint: hidden, defaults to \code{NULL}}

\item{chart_layout_footer_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_margin}{Margin. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_order}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_align_vertical}{V. align. Choices: flex-start, center, flex-end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_space}{Space. Space between border and footer text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_screenreader_text_primary}{Screenreader description. A text alternative to the visual content that will only be visible to screenreaders, e.g. The line chart shows China consistently higher than the other countries since 1990.  Do not replicate your title, since that will also be read by screenreaders. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_layout_screenreader_label}{Screenreader label. A short text label given to the main Flourish embed wrapper to provide an accessible name that is only visible to screenreaders Added in the form of an "aria-label". Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_screenreader_hide_primary}{Screenreader mode for main visual container. Choices: \code{TRUE}, \code{FALSE}. Whether the main visual container is visible to screenreaders (Text in the header and footer are always available to screenreaders.). Flourish type hint: boolean, defaults to \code{NULL}}
}
\value{
A Flourish chart
}
\description{
Set gantt chart chart details
}
\examples{
try(
  flourish(chart_type = "gantt", api_key = Sys.getenv("FLOURISH_API_KEY")) |> 
  bind_gantt_data(gapminder) |> 
  set_gantt_details()
)
}
