% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_horserace_details.R
\name{set_horserace_details}
\alias{set_horserace_details}
\title{Set line chart race chart details}
\usage{
set_horserace_details(
  .,
  play_mode = NULL,
  zoom_steps_to_show = NULL,
  value_type = NULL,
  higher_scores_win = NULL,
  ties_mode = NULL,
  height_mode = NULL,
  padding_top = NULL,
  padding_right = NULL,
  padding_bottom = NULL,
  padding_left = NULL,
  show_buttons = NULL,
  show_replay = NULL,
  chart_filter_controls_control_type = NULL,
  chart_filter_controls_control_styles = NULL,
  chart_filter_controls_control_title = NULL,
  chart_filter_controls_button_group = NULL,
  chart_filter_controls_button_group_width_mode = NULL,
  chart_filter_controls_button_group_width_fixed = NULL,
  chart_filter_controls_dropdown_width_mode = NULL,
  chart_filter_controls_dropdown_width_fixed = NULL,
  chart_filter_controls_slider_background_color = NULL,
  chart_filter_controls_slider_font_color = NULL,
  chart_filter_controls_slider_handle_color = NULL,
  chart_filter_controls_slider_width = NULL,
  chart_filter_controls_slider_margin = NULL,
  chart_filter_controls_slider_handle_height = NULL,
  chart_filter_controls_slider_track_height = NULL,
  chart_filter_controls_slider_play_button = NULL,
  chart_filter_controls_slider_loop = NULL,
  chart_filter_controls_slider_step_time = NULL,
  chart_filter_controls_slider_restart_pause = NULL,
  chart_filter_controls_sort = NULL,
  filter_include_all = NULL,
  label_ranks = NULL,
  label_scores = NULL,
  label_replay = NULL,
  filter_all_label = NULL,
  chart_controls_style_font_size = NULL,
  chart_controls_style_font_weight = NULL,
  chart_controls_style_height = NULL,
  chart_button_control_style_background = NULL,
  chart_button_control_style_background_selected = NULL,
  chart_button_control_style_background_hover = NULL,
  chart_button_control_style_font_color = NULL,
  chart_button_control_style_font_color_selected = NULL,
  chart_button_control_style_font_color_hover = NULL,
  chart_button_control_style_button_styles_advanced = NULL,
  chart_button_control_style_border_width = NULL,
  chart_button_control_style_border_color = NULL,
  chart_button_control_style_border_transparency = NULL,
  chart_button_control_style_border_radius = NULL,
  chart_dropdown_control_style_background = NULL,
  chart_dropdown_control_style_font_color = NULL,
  chart_dropdown_control_style_border_styles_advanced = NULL,
  chart_dropdown_control_style_border_style = NULL,
  chart_dropdown_control_style_border_width = NULL,
  chart_dropdown_control_style_border_color = NULL,
  chart_dropdown_control_style_border_transparency = NULL,
  chart_dropdown_control_style_border_radius = NULL,
  color_mode = NULL,
  chart_color_scale_categorical_palette = NULL,
  chart_color_scale_categorical_extend = NULL,
  chart_color_scale_categorical_custom_palette = NULL,
  chart_color_scale_numeric_type = NULL,
  chart_color_scale_binning = NULL,
  chart_color_scale_bin_mode = NULL,
  chart_color_scale_bin_count = NULL,
  chart_color_scale_bin_thresholds = NULL,
  chart_color_scale_sequential_palette = NULL,
  chart_color_scale_sequential_reverse = NULL,
  chart_color_scale_sequential_custom_min = NULL,
  chart_color_scale_sequential_custom_max = NULL,
  chart_color_scale_sequential_color_space = NULL,
  chart_color_scale_sequential_custom_domain = NULL,
  chart_color_scale_sequential_domain_min = NULL,
  chart_color_scale_sequential_domain_max = NULL,
  chart_color_scale_diverging_palette = NULL,
  chart_color_scale_diverging_reverse = NULL,
  chart_color_scale_diverging_custom_min = NULL,
  chart_color_scale_diverging_custom_mid = NULL,
  chart_color_scale_diverging_custom_max = NULL,
  chart_color_scale_diverging_color_space = NULL,
  chart_color_scale_diverging_custom_domain = NULL,
  chart_color_scale_diverging_domain_min = NULL,
  chart_color_scale_diverging_domain_mid = NULL,
  chart_color_scale_diverging_domain_max = NULL,
  color_single = NULL,
  color_single_overrides = NULL,
  line_width = NULL,
  line_opacity = NULL,
  curve = NULL,
  shade = NULL,
  shade_width = NULL,
  shade_opacity = NULL,
  missing = NULL,
  missing_opacity = NULL,
  missing_width = NULL,
  missing_stroke_dash = NULL,
  start_circle_r = NULL,
  end_circle_r = NULL,
  end_circle_stroke = NULL,
  circle_space_between = NULL,
  end_circle_stroke_bg = NULL,
  horse_images = NULL,
  show_labels = NULL,
  animate_scores = NULL,
  rank_outside_picture = NULL,
  label_divider = NULL,
  rank_inside_color = NULL,
  chart_labels_max_width = NULL,
  chart_labels_overlap_mode = NULL,
  chart_labels_spreading_margin = NULL,
  chart_labels_data_point_margin = NULL,
  chart_labels_show_only_labels = NULL,
  chart_labels_text_color = NULL,
  chart_labels_text_size = NULL,
  chart_labels_text_bg_color = NULL,
  chart_labels_text_bg_size = NULL,
  chart_labels_text_line_height = NULL,
  chart_labels_text_max_lines = NULL,
  chart_labels_text_weight = NULL,
  chart_labels_lines = NULL,
  chart_labels_line_style = NULL,
  chart_labels_line_color = NULL,
  chart_labels_line_width = NULL,
  chart_labels_line_range = NULL,
  chart_labels_line_padding = NULL,
  chart_x_axis_axis_position = NULL,
  chart_x_axis_y_axis_position = NULL,
  chart_x_axis_numeric_scale_type = NULL,
  chart_x_axis_linear_min = NULL,
  chart_x_axis_linear_max = NULL,
  chart_x_axis_log_min = NULL,
  chart_x_axis_log_max = NULL,
  chart_x_axis_datetime_min = NULL,
  chart_x_axis_datetime_max = NULL,
  chart_x_axis_flip = NULL,
  chart_x_axis_show_scale_settings = NULL,
  chart_x_axis_nice = NULL,
  chart_x_axis_zero_axis = NULL,
  chart_x_axis_title_mode = NULL,
  chart_x_axis_title = NULL,
  chart_x_axis_title_styling = NULL,
  chart_x_axis_title_weight = NULL,
  chart_x_axis_title_color = NULL,
  chart_x_axis_title_size = NULL,
  chart_x_axis_title_padding = NULL,
  chart_x_axis_tick_label_position = NULL,
  chart_x_axis_tick_label_styling = NULL,
  chart_x_axis_tick_label_size = NULL,
  chart_x_axis_tick_label_color = NULL,
  chart_x_axis_tick_padding = NULL,
  chart_x_axis_tick_label_angle = NULL,
  chart_x_axis_tick_label_weight = NULL,
  chart_x_axis_tick_label_max_lines = NULL,
  chart_x_axis_tick_label_line_height = NULL,
  chart_x_axis_tick_label_space_mode = NULL,
  chart_x_axis_tick_label_space = NULL,
  chart_x_axis_tick_mode = NULL,
  chart_x_axis_tick_number = NULL,
  chart_x_axis_tick_custom = NULL,
  chart_x_axis_line_and_tick_color = NULL,
  chart_x_axis_line_and_tick_width = NULL,
  chart_x_axis_tick_length = NULL,
  chart_x_axis_tick_side = NULL,
  chart_x_axis_line_visible = NULL,
  chart_x_axis_edge_padding = NULL,
  chart_x_axis_gridlines_visible = NULL,
  chart_x_axis_gridlines_styling = NULL,
  chart_x_axis_gridline_color = NULL,
  chart_x_axis_gridline_style = NULL,
  chart_x_axis_gridline_width = NULL,
  chart_x_axis_gridline_category_dividers = NULL,
  chart_x_axis_gridline_category_dividers_extend = NULL,
  chart_y_axis_axis_visible = NULL,
  chart_y_axis_numeric_scale_type = NULL,
  chart_y_axis_linear_min = NULL,
  chart_y_axis_linear_max = NULL,
  chart_y_axis_log_min = NULL,
  chart_y_axis_log_max = NULL,
  chart_y_axis_datetime_min = NULL,
  chart_y_axis_datetime_max = NULL,
  chart_y_axis_flip = NULL,
  chart_y_axis_show_scale_settings = NULL,
  chart_y_axis_nice = NULL,
  chart_y_axis_zero_axis = NULL,
  chart_y_axis_title_mode = NULL,
  chart_y_axis_title = NULL,
  chart_y_axis_title_styling = NULL,
  chart_y_axis_title_position = NULL,
  chart_y_axis_title_weight = NULL,
  chart_y_axis_title_color = NULL,
  chart_y_axis_title_size = NULL,
  chart_y_axis_title_padding = NULL,
  chart_y_axis_tick_label_position = NULL,
  chart_y_axis_tick_label_styling = NULL,
  chart_y_axis_tick_label_size = NULL,
  chart_y_axis_tick_label_color = NULL,
  chart_y_axis_tick_padding = NULL,
  chart_y_axis_tick_label_angle = NULL,
  chart_y_axis_tick_label_weight = NULL,
  chart_y_axis_tick_label_max_lines = NULL,
  chart_y_axis_tick_label_line_height = NULL,
  chart_y_axis_tick_label_space_mode = NULL,
  chart_y_axis_tick_label_space = NULL,
  chart_y_axis_tick_mode = NULL,
  chart_y_axis_tick_number = NULL,
  chart_y_axis_tick_custom = NULL,
  chart_y_axis_line_and_tick_color = NULL,
  chart_y_axis_line_and_tick_width = NULL,
  chart_y_axis_tick_length = NULL,
  chart_y_axis_tick_side = NULL,
  chart_y_axis_line_visible = NULL,
  chart_y_axis_edge_padding = NULL,
  chart_y_axis_gridlines_visible = NULL,
  chart_y_axis_gridlines_styling = NULL,
  chart_y_axis_gridline_color = NULL,
  chart_y_axis_gridline_style = NULL,
  chart_y_axis_gridline_width = NULL,
  chart_y_axis_gridline_category_dividers = NULL,
  chart_y_axis_gridline_category_dividers_extend = NULL,
  zoom_y_axis = NULL,
  caption_background_color = NULL,
  caption_border_color = NULL,
  caption_border_radius = NULL,
  caption_opacity = NULL,
  caption_padding = NULL,
  caption_position = NULL,
  caption_text_width = NULL,
  caption_font_size = NULL,
  caption_text_weight = NULL,
  caption_text_color = NULL,
  caption_mode = NULL,
  caption_text_align_h = NULL,
  caption_text_align_v = NULL,
  caption_image_width = NULL,
  caption_image_position = NULL,
  caption_space_between = NULL,
  stage_duration = NULL,
  update_duration = NULL,
  chart_number_date_format_prefix = NULL,
  chart_number_date_format_suffix = NULL,
  chart_number_date_format_n_dec = NULL,
  chart_number_date_format_advanced = NULL,
  chart_number_date_format_negative_sign = NULL,
  chart_number_date_format_strip_zeros = NULL,
  chart_number_date_format_strip_separator = NULL,
  chart_number_date_format_transform_labels = NULL,
  chart_number_date_format_transform = NULL,
  chart_number_date_format_multiply_divide_constant = NULL,
  chart_number_date_format_exponentiate_constant = NULL,
  chart_layout_body_font = NULL,
  chart_layout_font_color = NULL,
  chart_layout_background_color_enabled = NULL,
  chart_layout_background_image_enabled = NULL,
  chart_layout_background_color = NULL,
  chart_layout_background_image_src = NULL,
  chart_layout_background_image_size = NULL,
  chart_layout_background_image_position = NULL,
  chart_layout_max_width_target = NULL,
  chart_layout_max_width = NULL,
  chart_layout_max_width_align = NULL,
  chart_layout_layout_order = NULL,
  chart_layout_space_between_sections = NULL,
  chart_layout_space_between_sections_custom = NULL,
  chart_layout_margin_top = NULL,
  chart_layout_margin_right = NULL,
  chart_layout_margin_bottom = NULL,
  chart_layout_margin_left = NULL,
  chart_layout_border_enabled = NULL,
  chart_layout_border_top_width = NULL,
  chart_layout_border_top_style = NULL,
  chart_layout_border_top_color = NULL,
  chart_layout_border_right_width = NULL,
  chart_layout_border_right_style = NULL,
  chart_layout_border_right_color = NULL,
  chart_layout_border_bottom_width = NULL,
  chart_layout_border_bottom_style = NULL,
  chart_layout_border_bottom_color = NULL,
  chart_layout_border_left_width = NULL,
  chart_layout_border_left_style = NULL,
  chart_layout_border_left_color = NULL,
  chart_layout_read_direction = NULL,
  chart_layout_font_size_mobile_small = NULL,
  chart_layout_font_size_mobile_big = NULL,
  chart_layout_font_size_tablet = NULL,
  chart_layout_font_size_desktop = NULL,
  chart_layout_font_size_big_screen = NULL,
  chart_layout_breakpoint_mobile_small = NULL,
  chart_layout_breakpoint_mobile_big = NULL,
  chart_layout_breakpoint_tablet = NULL,
  chart_layout_breakpoint_desktop = NULL,
  chart_layout_breakpoint_big_screen = NULL,
  chart_layout_header_align = NULL,
  chart_layout_title = NULL,
  chart_layout_title_styling = NULL,
  chart_layout_title_font = NULL,
  chart_layout_title_size = NULL,
  chart_layout_title_size_custom = NULL,
  chart_layout_title_weight = NULL,
  chart_layout_title_color = NULL,
  chart_layout_title_line_height = NULL,
  chart_layout_title_space_above = NULL,
  chart_layout_title_space_above_custom = NULL,
  chart_layout_subtitle = NULL,
  chart_layout_subtitle_styling = NULL,
  chart_layout_subtitle_font = NULL,
  chart_layout_subtitle_size = NULL,
  chart_layout_subtitle_size_custom = NULL,
  chart_layout_subtitle_weight = NULL,
  chart_layout_subtitle_color = NULL,
  chart_layout_subtitle_line_height = NULL,
  chart_layout_subtitle_space_above = NULL,
  chart_layout_subtitle_space_above_custom = NULL,
  chart_layout_header_text = NULL,
  chart_layout_header_text_styling = NULL,
  chart_layout_header_text_size = NULL,
  chart_layout_header_text_size_custom = NULL,
  chart_layout_header_text_weight = NULL,
  chart_layout_header_text_color = NULL,
  chart_layout_header_text_line_height = NULL,
  chart_layout_header_text_space_above = NULL,
  chart_layout_header_text_space_above_custom = NULL,
  chart_layout_header_border = NULL,
  chart_layout_header_border_width = NULL,
  chart_layout_header_border_color = NULL,
  chart_layout_header_border_style = NULL,
  chart_layout_header_border_space = NULL,
  chart_layout_header_logo_enabled = NULL,
  chart_layout_header_logo_src = NULL,
  chart_layout_header_logo_alt = NULL,
  chart_layout_header_logo_link_url = NULL,
  chart_layout_header_logo_height = NULL,
  chart_layout_header_logo_align = NULL,
  chart_layout_header_logo_position_inside = NULL,
  chart_layout_header_logo_position_outside = NULL,
  chart_layout_header_logo_margin_top = NULL,
  chart_layout_header_logo_margin_right = NULL,
  chart_layout_header_logo_margin_bottom = NULL,
  chart_layout_header_logo_margin_left = NULL,
  chart_layout_footer_align = NULL,
  chart_layout_footer_text_size = NULL,
  chart_layout_footer_text_color = NULL,
  chart_layout_footer_styling = NULL,
  chart_layout_footer_font = NULL,
  chart_layout_footer_text_weight = NULL,
  chart_layout_source_name = NULL,
  chart_layout_source_url = NULL,
  chart_layout_multiple_sources = NULL,
  chart_layout_source_name_2 = NULL,
  chart_layout_source_url_2 = NULL,
  chart_layout_source_name_3 = NULL,
  chart_layout_source_url_3 = NULL,
  chart_layout_source_label = NULL,
  chart_layout_footer_note = NULL,
  chart_layout_footer_note_secondary = NULL,
  chart_layout_advanced_note_styling = NULL,
  chart_layout_footer_timestamp_format = NULL,
  chart_layout_footer_logo_enabled = NULL,
  chart_layout_footer_logo_src = NULL,
  chart_layout_footer_logo_src_light = NULL,
  chart_layout_footer_logo_alt = NULL,
  chart_layout_footer_logo_link_url = NULL,
  chart_layout_footer_logo_height = NULL,
  chart_layout_footer_logo_margin = NULL,
  chart_layout_footer_logo_order = NULL,
  chart_layout_footer_align_vertical = NULL,
  chart_layout_footer_border = NULL,
  chart_layout_footer_border_width = NULL,
  chart_layout_footer_border_color = NULL,
  chart_layout_footer_border_style = NULL,
  chart_layout_footer_border_space = NULL,
  chart_layout_screenreader_text_primary = NULL,
  chart_layout_screenreader_label = NULL,
  chart_layout_screenreader_hide_primary = NULL
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{play_mode}{Play mode. In All mode, the entire X axis is visible In Zoomed mode, the X axis is only visible for a specified amount of stages In Reveal mode, the X axis is visible from the start until the current stage Flourish type hint: string, defaults to \code{NULL}}

\item{zoom_steps_to_show}{Visible stages. The number of time slices (each of which is column of data) to show on the X axis.  When in Zoom mode, these time slices are shown to the left and right of the current time.  In Reveal mode, these time slices are shown at the beginning.  (Value is ignored if the number specified is greater than the number of stages in the data.). Flourish type hint: number, defaults to \code{NULL}}

\item{value_type}{Chart mode to show on load. Choices: ranks, scores. Flourish type hint: string, defaults to \code{NULL}}

\item{higher_scores_win}{Data type. Choices: \code{TRUE}, \code{FALSE}. Useful for data where lower scores beat higher scores e.g. race times. Flourish type hint: boolean, defaults to \code{NULL}}

\item{ties_mode}{Rank ties mode. Choices: competition, dense. What ranks should be given when there are ties? See Wikipedia for an explanation of names. Flourish type hint: string, defaults to \code{NULL}}

\item{height_mode}{Height mode. Choices: auto, fixed, flexible. In Fill space mode, the lines will be spread out over the available chart height, which by default will be the standard Flourish chart sizes for any given screen width. In this mode, the circles/images may overlap if you have lots of lines.  Match data mode will resize the embedded graphic so there is enough vertical space to avoid overlaps. (Note this wont have any effect if you embed the graphic as a fixed height.)  Auto mode switches between the two modes depending on the number of lines. Flourish type hint: string, defaults to \code{NULL}}

\item{padding_top}{Top. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{padding_right}{Right. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{padding_bottom}{Bottom. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{padding_left}{Left (mobile). In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{show_buttons}{Ranks/scores toggle. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{show_replay}{Replay button. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_control_type}{Control type. Choices: dropdown, buttons, slider. Choose between a dropdown, buttons or a time slider. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_control_styles}{Adv. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_control_title}{Control title. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_button_group}{Grouped. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_button_group_width_mode}{Button group width. Choices: auto, full, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_button_group_width_fixed}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_dropdown_width_mode}{Dropdown width. Choices: auto, full, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_filter_controls_dropdown_width_fixed}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_background_color}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_filter_controls_slider_font_color}{Text. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_filter_controls_slider_handle_color}{Handle. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_filter_controls_slider_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_margin}{Label width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_handle_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_track_height}{Track height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_play_button}{Play button. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_slider_loop}{Loop. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_filter_controls_slider_step_time}{Time between steps. Measured in seconds, positive values move the slider left to right, negative values move the slider right to left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_slider_restart_pause}{Pause before loop. Measured in seconds and in addition to the regular step time displayed above. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_filter_controls_sort}{Sorting. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{filter_include_all}{Include “All” option. Flourish type hint: boolean, defaults to \code{NULL}}

\item{label_ranks}{“Ranks”. Flourish type hint: string, defaults to \code{NULL}}

\item{label_scores}{“Scores”. Flourish type hint: string, defaults to \code{NULL}}

\item{label_replay}{“Replay”. Flourish type hint: string, defaults to \code{NULL}}

\item{filter_all_label}{“All”. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_controls_style_font_size}{Text size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_controls_style_font_weight}{Text weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_controls_style_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_background_selected}{Selected. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_background_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color_selected}{Selected. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_font_color_hover}{Mouse over. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_button_styles_advanced}{Button border styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_button_control_style_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_button_control_style_border_transparency}{Transparency. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_button_control_style_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_background}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_font_color}{Text color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_styles_advanced}{Dropdown border styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_style}{Border style. Choices: all, bottom. Show border on all sides, or only at the bottom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_width}{Border width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_transparency}{Transparency. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_dropdown_control_style_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{color_mode}{Color mode. Choices: category, line, single. Whether to color by category (if you have a category column set in the Datasheet), by each individual line, or give every line the same color. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_categorical_palette}{Palette. Flourish type hint: colors, defaults to \code{NULL}}

\item{chart_color_scale_categorical_extend}{Extend. Automatically generate additional colours when needed to avoid the palette colours being used more than once. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_categorical_custom_palette}{Custom overrides. Enter the label name for which you wish to set the color, followed by a colon and the desired color value.  Colors can be set using Hex, RGB, color names or RGBA, if you want to set the opacity Multiple colors can be set using multiple lines. For example:  Party 1: red Party 2: #4455AA Party 3: rgb(30,168,26). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_color_scale_numeric_type}{Scale type. Choices: sequential, diverging. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_binning}{Choices: \code{FALSE}, \code{TRUE}. In linear mode, the color scale will run as a smooth gradient between 2 colors. In binned mode, the gradient will be divided in smaller blocks. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_bin_mode}{Binning mode. Choices: fixed, quantile, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_bin_count}{Number of bins. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_bin_thresholds}{Custom thresholds. Enter your desired thresholds, separating them with a ";". For instance, "5;10;15".  Bins form as follows:  From the datas minimum value (domain minimum) up to the first threshold.  Between consecutive thresholds.  From the last threshold to the datas maximum value (domain maximum).   For "5;10;15", youll get four bins based on your datas range. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_sequential_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_sequential_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_sequential_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_palette}{Palette. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_reverse}{Reverse. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_min}{Minimum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_mid}{Midpoint color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_max}{Maximum color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_color_scale_diverging_color_space}{Color space. Choices: rgb, lab, hcl, hsl. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_color_scale_diverging_custom_domain}{Domain. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_mid}{Mid. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_color_scale_diverging_domain_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{color_single}{Single color. Flourish type hint: color, defaults to \code{NULL}}

\item{color_single_overrides}{Color overrides. Type the label of the line whose color you want to set, a colon and then a color (using a name, hex-code or rgb).  Multiple colors can be set using multiple lines. For example:   China: red India: #4455AA Brazil: rgb(30,168,26). Flourish type hint: text, defaults to \code{NULL}}

\item{line_width}{Line width. Flourish type hint: number, defaults to \code{NULL}}

\item{line_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{curve}{Curve. Choices: curve_linear, curve_monotone_x, curve_natural, curve_step, curve_step_after, curve_step_before. Flourish type hint: string, defaults to \code{NULL}}

\item{shade}{Shading behind line. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{shade_width}{Width. Width in pixels. Flourish type hint: number, defaults to \code{NULL}}

\item{shade_opacity}{Opacity. Set to a value between 0 and 1. Flourish type hint: number, defaults to \code{NULL}}

\item{missing}{Lines between missing data points. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{missing_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{missing_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{missing_stroke_dash}{Stroke dash. Flourish type hint: number, defaults to \code{NULL}}

\item{start_circle_r}{Start radius. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{end_circle_r}{End radius. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{end_circle_stroke}{End stroke. Flourish type hint: number, defaults to \code{NULL}}

\item{circle_space_between}{Space between. Minimum space between the circles. The space is ignored when the chart is set to fill available space. Flourish type hint: number, defaults to \code{NULL}}

\item{end_circle_stroke_bg}{Stroke color. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{horse_images}{Image inside circle. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{show_labels}{Show labels. Choices: always, hover, never. Flourish type hint: string, defaults to \code{NULL}}

\item{animate_scores}{Animate scores. Flourish type hint: boolean, defaults to \code{NULL}}

\item{rank_outside_picture}{Value label position. Choices: \code{FALSE}, \code{TRUE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{label_divider}{Divider. The character between label and value. Flourish type hint: string, defaults to \code{NULL}}

\item{rank_inside_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_labels_max_width}{Max label width. The maximum allowed label width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_labels_overlap_mode}{Overlaps. Choices: spread, hide, none. What to do with overlaps. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_labels_spreading_margin}{Space between labels. The space left between labels when spreading them out to avoid overlaps. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_labels_data_point_margin}{Distance. The space between the data point and the label in rems (a multiple of base font size). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_labels_show_only_labels}{Show only certain labels. Manually specify which labels youd like to display (one entry per line that should match entries in the Label column). Flourish type hint: text, defaults to \code{NULL}}

\item{chart_labels_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_labels_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_labels_text_bg_color}{Outline. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_labels_text_bg_size}{Outline width. As a \% of the font size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_labels_text_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_labels_text_max_lines}{Max lines. The max number of lines for each label. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_labels_text_weight}{Weight. Choices: 400, 700. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_labels_lines}{Connector lines. Choices: auto, on, off. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_labels_line_style}{Line style. Choices: straight, step. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_labels_line_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_labels_line_width}{Width. In rems (a multiple of base font size). Flourish type hint: number, defaults to \code{NULL}}

\item{chart_labels_line_range}{Length. The width of the space given over to lines; in rems, a multiple of base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_labels_line_padding}{Padding. The padding to the left and right of the line; in rems, a multiple of base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_axis_position}{Position. Choices: bottom, float-below, float-above, top, off. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_y_axis_position}{Y value. Vertical position of the X axis on the Y scale. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_numeric_scale_type}{Type. Choices: linear, log. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_linear_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_linear_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_log_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_log_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_datetime_min}{Min (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_datetime_max}{Max (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_flip}{Flip axis. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_show_scale_settings}{Configure default min/max. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_nice}{Round min/max. Choices: \code{TRUE}, \code{FALSE}. Enabling this option rounds the X axis start and end values to tidy numbers, such as 0, 50, 100 or whole years. Specified min and max values will override this feature. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_zero_axis}{Include zero. Choices: auto, on, off. Enable to extend the axis to include zero, even when data is all positive or all negative. This is ignored if you set specific min and max scale values. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_mode}{Type. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title}{Text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_title_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_title_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_position}{Position. Choices: default, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_tick_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_angle}{Angle. Choices: 0, 30, 45, 60, 90. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_max_lines}{Max lines. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_space_mode}{Space mode. Choices: auto, max, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_label_space}{Space. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_mode}{Mode. Choices: auto, number, custom. Sets the method for selecting tick values on the axis:  Auto automatically selects the ticks  Number allows you to choose the number of ticks to display.  Custom lets you select specific ticks to show.  Note: If the axis is categorical, Number mode is ignored. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_tick_number}{Number. Approximate number of ticks or gridlines. The actual number will depend on the range of values, chart size, etc. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_custom}{One tick label per line. Specify the ticks to show, one per line. For dates, use the input format as specified in the datasheet.  You can also use \code{{{FIRST}}} and \code{{{LAST}}} to add ticks at the minimum and maximum of the axis. For example:  \code{{{FIRST}}} 25000 50000 75000 \code{{{LAST}}}. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_x_axis_line_and_tick_color}{Line color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_line_and_tick_width}{Line width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_length}{Tick length. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_tick_side}{Tick side. Choices: out, in. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_line_visible}{Axis line. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_edge_padding}{Edge padding (\%). Space between start/end of axis line and first/last category tick as a percentage of the spacing between ticks. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_gridlines_visible}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridlines_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_x_axis_gridline_style}{Style. Choices: solid, dash, dot, dot_dash. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_x_axis_gridline_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_x_axis_gridline_category_dividers}{Put lines between categories. On a categorical axis, this option disables the gridline on each tick and instead adds rules between each category to give a table-like design. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_x_axis_gridline_category_dividers_extend}{Extend. Extends the dividers into the axis margins. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_axis_visible}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_numeric_scale_type}{Type. Choices: linear, log. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_linear_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_linear_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_log_min}{Min. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_log_max}{Max. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_datetime_min}{Min (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_datetime_max}{Max (date). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_flip}{Flip axis. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_show_scale_settings}{Configure default min/max. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_nice}{Round min/max. Choices: \code{TRUE}, \code{FALSE}. When enabled, scales automatically extend to "nice" rounded start/end values, such as hundreds or whole years. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_zero_axis}{Include zero. Choices: auto, on, off. Whether to extend the axis to include zero if the data values are all positive or all negative. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_mode}{Type. Choices: auto, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title}{Text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_title_position}{Position. Choices: side, end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_title_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_title_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_position}{Position. Choices: default, above, below. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_tick_padding}{Padding. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_angle}{Angle. Choices: 0, 30, 45, 60, 90. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_max_lines}{Max lines. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_space_mode}{Space mode. Choices: auto, max, fixed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_label_space}{Space. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_mode}{Mode. Choices: auto, number, custom. Sets the method for selecting tick values on the axis:  Auto automatically selects the ticks  Number allows you to choose the number of ticks to display.  Custom lets you select specific ticks to show.  Note: If the axis is categorical, Number mode is ignored. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_tick_number}{Number. Approximate number of ticks or gridlines. The actual number will depend on the range of values, chart size, etc. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_custom}{One tick label per line. Specify the ticks to show, one per line. For dates, use the input format as specified in the datasheet.  You can also use \code{{{FIRST}}} and \code{{{LAST}}} to add ticks at the minimum and maximum of the axis. For example:  \code{{{FIRST}}} 25000 50000 75000 \code{{{LAST}}}. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_y_axis_line_and_tick_color}{Line color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_line_and_tick_width}{Line width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_length}{Tick length. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_tick_side}{Side. Choices: out, in. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_line_visible}{Axis line. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_edge_padding}{Edge padding (\%). Space between start/end of axis line and first/last category tick as a percentage of the spacing between ticks. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_gridlines_visible}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_gridlines_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_gridline_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_y_axis_gridline_style}{Style. Choices: solid, dash, dot, dot_dash. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_y_axis_gridline_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_y_axis_gridline_category_dividers}{Put lines between categories. On a categorical axis, this option disables the gridline on each tick and instead adds rules between each category to give a table-like design. Ignored for date and numeric axes. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_y_axis_gridline_category_dividers_extend}{Extend. Extends the dividers into the axis margins. Flourish type hint: boolean, defaults to \code{NULL}}

\item{zoom_y_axis}{Dynamic Y axis. Rescales the y axis when filtering or playing in zoomed mode. Flourish type hint: boolean, defaults to \code{NULL}}

\item{caption_background_color}{Background. Flourish type hint: color, defaults to \code{NULL}}

\item{caption_border_color}{Border. Flourish type hint: color, defaults to \code{NULL}}

\item{caption_border_radius}{Radius. Flourish type hint: number, defaults to \code{NULL}}

\item{caption_opacity}{Opacity. Flourish type hint: number, defaults to \code{NULL}}

\item{caption_padding}{Padding. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{caption_position}{Position. Choices: top-left, center-left, bottom-left, top-right, center-right, bottom-right, top-center, center-center, bottom-center. Flourish type hint: string, defaults to \code{NULL}}

\item{caption_text_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{caption_font_size}{Font size. In rems, a multiple of the pages base font size. Flourish type hint: number, defaults to \code{NULL}}

\item{caption_text_weight}{Weight. Choices: normal, bold. Flourish type hint: string, defaults to \code{NULL}}

\item{caption_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{caption_mode}{Content mode. Choices: text_legend, html. In Text with colors mode, any phrases matching category names will be automatically colored.  In HTML mode, no colors will be applied, but you can include any HTML in the datasheet to add images, etc. Flourish type hint: string, defaults to \code{NULL}}

\item{caption_text_align_h}{Horizontal align. Choices: start, center, end. Flourish type hint: string, defaults to \code{NULL}}

\item{caption_text_align_v}{Vertical align. Choices: flex-start, center, flex-end. Flourish type hint: string, defaults to \code{NULL}}

\item{caption_image_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{caption_image_position}{Position. Choices: column, column-reverse, row, row-reverse. Flourish type hint: string, defaults to \code{NULL}}

\item{caption_space_between}{Space between. Space between image and text. Flourish type hint: number, defaults to \code{NULL}}

\item{stage_duration}{Animation duration. The race animation length, in milliseconds per stage(A stage is equal to each column of your data). Flourish type hint: number, defaults to \code{NULL}}

\item{update_duration}{Mode duration. The animation duration when change between ranks and scores modes. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_prefix}{Prefix. Text to place in front of a number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_suffix}{Suffix. Text to place after a number. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_n_dec}{Decimal places. Decimal places. You can also enter a negative integer to round to a whole number with that many zeros. For example, -2 will round to the nearest hundred. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_advanced}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_negative_sign}{Styling of negative numbers. Choices: -$nk, $-nk, ($nk), $(n)k, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_strip_zeros}{Remove trailing zeros. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_strip_separator}{Hide thousands separator below 10,000. Turn off if you want four-digit numbers to include a separator, e.g. 1,234 rather than 1234. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_transform_labels}{Multiply/divide values. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_number_date_format_transform}{Choices: multiply, divide, exponentiate. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_number_date_format_multiply_divide_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_number_date_format_exponentiate_constant}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_body_font}{Main font. This font will apply to the whole graphic by default. You can optionally change the font for the title, subtitle, footer, etc in the Header and Footer settings panels. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_font_color}{Text color. This color will apply to the whole graphic by default, You can optionally change the color for individual text elements, in other settings panels. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_color_enabled}{Color. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_image_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_background_color}{Background color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_background_image_src}{Image URL. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_background_image_size}{Size. Choices: cover, contain, auto, 100\% 100\%. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_background_image_position}{Position. Choices: top left, top center, top right, center left, center center, center right, bottom left, bottom center, bottom right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width_target}{Maximum width. Choices: none, wrapper, primary. Apply a maximum width to just the main graphic or everything (main graphic plus header, footer, etc). Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_max_width}{Maximum width. Leave blank to stretch to container width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_max_width_align}{Align. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_layout_order}{Layout order. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections}{Space between sections. Choices: 0.5, 1, 1.5, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_space_between_sections_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_enabled}{Show borders around visualisation. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_border_top_width}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_top_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_top_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_right_width}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_right_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_right_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_bottom_width}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_bottom_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_bottom_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_border_left_width}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_border_left_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_border_left_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_read_direction}{Read direction. Choices: ltr, rtl. This will change the reading direction of the main text elements on the page. Its not possible to adjust this on all elements, such as axes.  Note that when direction is set to right to left any alignment icons will be reversed. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_font_size_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_small}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_mobile_big}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_tablet}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_desktop}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_breakpoint_big_screen}{Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_align}{Alignment. Choices: left, center, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_title_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_title_font}{Title Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_title_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_title_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_title_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_title_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_subtitle_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_subtitle_font}{Subtitle Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_subtitle_size}{Size. Choices: 1.4, 1.6, 2, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_subtitle_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_subtitle_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text}{Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_header_text_styling}{Styling. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_text_size}{Size. Choices: 1.2, 1.4, 1.6, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_size_custom}{Custom. Specify a custom responsive font size in rems The best results will be with values between 1.2 and 3. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_text_line_height}{Line height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above}{Space above. Choices: 0, 0.5, 1, custom. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_text_space_above_custom}{Custom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_header_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_border_space}{Space. Space between border and header text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_enabled}{Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_header_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_header_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_align}{Align. Choices: inside, outside. Align logo inside either the header or the main visualization container. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_inside}{Position. Choices: top, left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_position_outside}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_top}{Top. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_right}{Right. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_bottom}{Bottom. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_header_logo_margin_left}{Left. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_align}{Alignment. Choices: left, center, right, justify. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_text_size}{Size. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_text_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_styling}{Advanced footer styles. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_font}{Font. Flourish type hint: font, defaults to \code{NULL}}

\item{chart_layout_footer_text_weight}{Weight. Choices: bold, normal. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_multiple_sources}{Multiple sources. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_source_name_2}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_2}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_name_3}{Source name. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_url_3}{Source url. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_source_label}{Source label. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_note}{Note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_footer_note_secondary}{Note (secondary). The secondary note is placed below the source and primary note. To add the time/date stamp of when the data was last updated add \code{{{data_last_updated}}}. For example, "Last updated at \code{{{data_last_updated}}}". To customize the format, use the advanced options below. Flourish type hint: html, defaults to \code{NULL}}

\item{chart_layout_advanced_note_styling}{Advanced. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_timestamp_format}{Time/date stamp formatting. Choices: \%H:\%M:\%S, \%I:\%M \%p, \%H:\%M, \%H:\%M \%p, \%H:\%M:\%S, \%d \%b \%Y, \%H:\%M, \%d \%b \%Y, \%H:\%M, \%d/\%m/\%Y, \%H:\%M, \%m/\%d/\%Y, \%H:\%M, \%d-\%m-\%Y, \%H:\%M, \%m-\%d-\%Y, \%d \%b \%Y, \%m/\%d/\%Y, \%d/\%m/\%Y, \%m-\%d-\%Y, \%d-\%m-\%Y, \%Y-\%m-\%d, \%Y-\%m-\%d \%H:\%M:\%S. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_enabled}{Image. Choices: \code{TRUE}, \code{FALSE}. Flourish type hint: boolean, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src}{Image. Flourish type hint: url, defaults to \code{NULL}}

\item{chart_layout_footer_logo_src_light}{Image (light version). If provided this version will be used whenever the background color is dark. Flourish type hint: hidden, defaults to \code{NULL}}

\item{chart_layout_footer_logo_alt}{Alt text. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_link_url}{Link. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_logo_height}{Height. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_margin}{Margin. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_logo_order}{Position. Choices: left, right. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_align_vertical}{V. align. Choices: flex-start, center, flex-end. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border}{Choices: top, bottom, top_and_bottom, none. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_width}{Width. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_footer_border_color}{Color. Flourish type hint: color, defaults to \code{NULL}}

\item{chart_layout_footer_border_style}{Style. Choices: solid, dashed, dotted. Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_footer_border_space}{Space. Space between border and footer text. Flourish type hint: number, defaults to \code{NULL}}

\item{chart_layout_screenreader_text_primary}{Screenreader description. A text alternative to the visual content that will only be visible to screenreaders, e.g. The line chart shows China consistently higher than the other countries since 1990.  Do not replicate your title, since that will also be read by screenreaders. Flourish type hint: text, defaults to \code{NULL}}

\item{chart_layout_screenreader_label}{Screenreader label. A short text label given to the main Flourish embed wrapper to provide an accessible name that is only visible to screenreaders Added in the form of an "aria-label". Flourish type hint: string, defaults to \code{NULL}}

\item{chart_layout_screenreader_hide_primary}{Screenreader mode for main visual container. Choices: \code{TRUE}, \code{FALSE}. Whether the main visual container is visible to screenreaders (Text in the header and footer are always available to screenreaders.). Flourish type hint: boolean, defaults to \code{NULL}}
}
\value{
A Flourish chart
}
\description{
Set line chart race chart details
}
\examples{
try(
  flourish(chart_type = "horserace", api_key = Sys.getenv("FLOURISH_API_KEY")) |> 
  bind_horserace_data(gapminder) |> 
  set_horserace_details()
)
}
