% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_charts.R
\name{snapshot}
\alias{snapshot}
\title{Snapshot a Flourish graph.}
\usage{
snapshot(
  .,
  format = "png",
  filename = "Flourish API Image",
  download = TRUE,
  scale = 1
)
}
\arguments{
\item{.}{The prior Flourish object. No need to specify name if piping graph as the graph will take the first argument (i.e. the prior existing graph).}

\item{format}{Mandatory - File format of snapshot: one of "png", "jpeg", "svg". Defaults to "png".}

\item{filename}{Optional - If chart is downloaded to local machine, you may specify the filename. If not provided, will default to "Flourish API Image.png".}

\item{download}{Mandatory - Whether the template is downloaded to local machine or the image data will be passed in the callback as a data URL with base64-encoded data. If \code{FALSE}, copy the base-64 encoded string from the chart console log. This makes it easy to use that data as, say, the \code{src} attribute of an \verb{<img>} tag. Defaults to \code{TRUE}.}

\item{scale}{Optional - You can supply a scale parameter (default: 1) to increase the resolution of the generated image.}
}
\value{
The filepath to the saved graph or a data URL with base64-encoded data.
}
\description{
Save a graph using Flourish's API (https://developers.flourish.studio/api/introduction/).
}
\examples{
try(
  flourish(
    chart_type = "scatter", api_key = Sys.getenv("FLOURISH_API_KEY")
  ) |>
  bind_scatter_data(
    data = subset(gapminder, year == 2007),
    y = "gdpPercap", x = "lifeExp"
  ) |>
  snapshot(
    format = "png",
    scale = 1,
    filename = "flourish_scatterplot",
    download = FALSE
  )
 )
}
